Module arrow2::alloc

Defines memory-related functions, such as allocate/deallocate/reallocate memory regions, cache and allocation alignments.

Statics

ALLOCATIONS

Functions

allocate_aligned
Allocates a cache-aligned memory region of size bytes with uninitialized values. This is more performant than using allocate_aligned_zeroed when all bytes will have an unknown or non-zero value and is semantically similar to malloc.

allocate_aligned_zeroed
Allocates a cache-aligned memory region of size bytes with 0 on all of them. This is more performant than using allocate_aligned and setting all bytes to zero and is semantically similar to calloc.

free_aligned⚠
Frees memory previously allocated by allocate_aligned_zeroed or allocate_aligned.

reallocate⚠
Reallocates memory previously allocated by allocate_aligned_zeroed or allocate_aligned.

total_allocated_bytes
Returns the total number of bytes allocated to buffers by the allocator.