Function arrow2::array::get_display

pub fn get_display<'a, F: Write + 'a>(
    array: &'a dyn Array,
    null: &'static str
) -> Box<dyn Fn(&mut F, usize) -> Result + 'a>

Returns a function that writes the element of array at position index to a Write, writing null to the null slots.