Function arrow2::array::get_value_display

pub fn get_value_display<'a, F: Write + 'a>(
    array: &'a dyn Array, 
    null: &'static str
) -> Box<dyn Fn(&mut F, usize) -> Result + 'a>

Returns a function that writes the value of the element of array at position index to a Write, writing null in the null slots.