Module arrow2::array::growable

Contains components suitable to create an array out of N other arrays, by slicing from each in a random-access fashion.

Structs

GrowableBinary

GrowableBoolean
A growable PrimitiveArray

GrowableDictionary
A growable PrimitiveArray

GrowableFixedSizeBinary

GrowableList

GrowableNull
A growable PrimitiveArray

GrowablePrimitive
A growable PrimitiveArray

GrowableStruct
A growable PrimitiveArray

GrowableUtf8

Traits

Growable

Functions

make_growable
Panics