Trait arrow2::array::growable::Growable

pub trait Growable<'a> {
    fn extend(&mut self, index: usize, start: usize, len: usize);

    fn extend_validity(&mut self, additional: usize);

    fn as_box(&mut self) -> Box<dyn Array>;

    fn as_arc(&mut self) -> Arc<dyn Array> { ... }
}

A trait describing a struct that can be extended from slices of pre-existing Arrays. This is used in operations where a new array is built out of other arrays such, as filtering and concatenation.

Required methods

fn extend(&mut self, index: usize, start: usize, len: usize)

Extends this Growable with elements from the bounded Array at index index from a slice starting at start and length len.

Panic

This function panics if the range is out of bounds, i.e. if start + len >= array.len().

fn extend_validity(&mut self, additional: usize)

Extends this [GrowableArray] with null elements, disregarding the bound arrays

fn as_box(&mut self) -> Box<dyn Array>

Converts itself to an Box<dyn Array>, thereby finishing the mutation. Self will be empty after such operation

Loading content...
Provided methods

fn as_arc(&mut self) -> Arc<dyn Array>

Converts itself to an Arc<dyn Array>, thereby finishing the mutation. Self will be empty after such operation

Loading content...