Struct arrow2::array::BinaryArray

pub struct BinaryArray<O: Offset> { /* private fields */ }

A BinaryArray is a nullable array of bytes - the Arrow equivalent of Vec<Option<Vec<u8>>>.

Safety

The following invariants hold:

Two consecutives offsets casted (as) to usize are valid slices of values.

len is equal to validity.len(), when defined.

Implementations

impl<'a, O: Offset> BinaryArray<O>

pub fn iter(&'a self) -> ZipValidity<'a, &'a [u8], BinaryValueIter<'a, O>>

Returns an iterator of Option<&[u8]>

pub fn values_iter(&'a self) -> BinaryValueIter<'a, O>

Returns an iterator of &[u8]

impl<O: Offset> BinaryArray<O>

pub fn from_slice<T: AsRef<[u8]>, P: AsRef<[T]>>(slice: P) -> Self

Creates a new BinaryArray from slices of &[u8].

pub fn from<T: AsRef<[u8]>, P: AsRef<[Option<T>]>>(slice: P) -> Self

Creates a new BinaryArray from a slice of optional &[u8].

pub fn from_trusted_len_values_iter<T: AsRef<[u8]>, I: TrustedLen<Item = T>>(
iterator: I
) -> Self

Creates a BinaryArray from an iterator of trusted length.

pub fn from_iter_values<T: AsRef<[u8]>, I: Iterator<Item = T>>(
iterator: I
) -> Self

Creates a new BinaryArray from a Iterator of &str.

impl<O: Offset> BinaryArray<O>

pub unsafe fn from_trusted_len_iter_unchecked<I, P>(iterator: I) -> Self where
P: AsRef<[u8]>,
I: Iterator<Item = Option<P>>,

Creates a BinaryArray from an iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_trusted_len_iter<I, P>(iterator: I) -> Self where
P: AsRef<[u8]>,
I: TrustedLen<Item = Option<P>>,

Creates a BinaryArray from an iterator of trusted length.

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I, P>(
iterator: I
) -> Result<Self, E> where
P: AsRef<[u8]>,
I: IntoIterator<Item = Result<Option<P>, E>>,

Creates a BinaryArray from an falible iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn try_from_trusted_len_iter<E, I, P>(iter: I) -> Result<Self, E> where
P: AsRef<[u8]>,
I: TrustedLen<Item = Result<Option<P>, E>>,

Creates a BinaryArray from an fallible iterator of trusted length.

impl<O: Offset> BinaryArray<O>

pub fn try_new(
data_type: DataType,
offsets: Buffer<O>,
values: Buffer<u8>,
validity: Option<Bitmap>
) -> Result<Self>

Creates a new BinaryArray.

Errors

This function returns an error iff:

the offsets are not monotonically increasing

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len() - 1.

The data_type’s crate::datatypes::PhysicalType is not equal to either Binary or LargeBinary.

Implementation

This function is O(N) - checking monotinicity is O(N)

pub fn new(
data_type: DataType,
offsets: Buffer<O>,
values: Buffer<u8>,
validity: Option<Bitmap>
) -> Self

Creates a new BinaryArray.

Panics

the offsets are not monotonically increasing

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len() - 1.

The data_type’s crate::datatypes::PhysicalType is not equal to either Binary or LargeBinary.

Implementation

This function is O(N) - checking monotinicity is O(N)

pub fn from_data(
data_type: DataType,
offsets: Buffer<O>,
values: Buffer<u8>,
validity: Option<Bitmap>
) -> Self

Alias for new

pub fn new_empty(data_type: DataType) -> Self

Creates an empty BinaryArray, i.e. whose .len is zero.

pub fn new_null(data_type: DataType, length: usize) -> Self

Creates an null BinaryArray, i.e. whose .null_count() == .len().

pub fn default_data_type() -> DataType

Returns the default DataType, DataType::Binary or DataType::LargeBinary

impl<O: Offset> BinaryArray<O>

pub unsafe fn try_new_unchecked(
data_type: DataType,
offsets: Buffer<O>,
values: Buffer<u8>,
validity: Option<Bitmap>
) -> Result<Self>

Creates a new BinaryArray without checking for offsets monotinicity.

Errors

This function returns an error iff:

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len() - 1.

The data_type’s crate::datatypes::PhysicalType is not equal to either Binary or LargeBinary.

Safety

This function is unsafe iff:

the offsets are not monotonically increasing

Implementation

This function is O(1)

pub unsafe fn new_unchecked(
data_type: DataType,
offsets: Buffer<O>,
values: Buffer<u8>,
validity: Option<Bitmap>
) -> Self

Creates a new BinaryArray without checking for offsets monotinicity.

Panics

This function returns an error iff:

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len() - 1.

The data_type’s crate::datatypes::PhysicalType is not equal to either Binary or LargeBinary.

Safety

This function is unsafe iff:

the offsets are not monotonically increasing

Implementation

This function is O(1)

pub unsafe fn from_data_unchecked(
data_type: DataType,
offsets: Buffer<O>,
values: Buffer<u8>,
validity: Option<Bitmap>
) -> Self

Alias for [new_unchecked]

Safety

This function is unsafe iff:

the offsets are not monotonically increasing

impl<O: Offset> BinaryArray<O>

pub fn slice(&self, offset: usize, length: usize) -> Self

Creates a new BinaryArray by slicing this BinaryArray.

Implementation

This function is O(1): all data will be shared between both arrays.

Panics

iff offset + length > self.len().

pub unsafe fn slice_unchecked(&self, offset: usize, length: usize) -> Self

Creates a new BinaryArray by slicing this BinaryArray.

Implementation

This function is O(1): all data will be shared between both arrays.

Safety

The caller must ensure that offset + length <= self.len().

pub fn with_validity(&self, validity: Option<Bitmap>) -> Self

Clones this BinaryArray with a different validity.

Panic

Panics iff validity.len() != self.len().

impl<O: Offset> BinaryArray<O>

pub fn len(&self) -> usize

Returns the length of this array

pub fn value(&self, i: usize) -> &[u8]

Returns the element at index i

Panics

iff i >= self.len()

pub unsafe fn value_unchecked(&self, i: usize) -> &[u8]

Returns the element at index i

Safety

Assumes that the i < self.len.

pub fn validity(&self) -> Option<&Bitmap>

The optional validity.

pub fn offsets(&self) -> &Buffer<O>

Returns the offsets that slice .values() to return valid values.

pub fn values(&self) -> &Buffer<u8>

Returns all values in this array. Use .offsets() to slice them.