Struct arrow2::array::BooleanArray

pub struct BooleanArray { /* fields omitted */ }

A BooleanArray is arrow’s equivalent to Vec<Option<bool>>, i.e. an array designed for highly performant operations on optionally nullable booleans. The size of this struct is O(1) as all data is stored behind an Arc.

Implementations

impl BooleanArray

pub fn from_trusted_len_values_iter<I: TrustedLen<Item = bool>>(
iterator: I
) -> Self

Creates a new BooleanArray from an TrustedLen of bool.

pub fn from_slice<P: AsRef<[bool]>>(slice: P) -> Self

Creates a new BooleanArray from a slice of bool.

impl BooleanArray

pub unsafe fn from_trusted_len_iter_unchecked<I, P>(iterator: I) -> Self where
P: Borrow<bool>,
I: Iterator<Item = Option<P>>,

Creates a BooleanArray from an iterator of trusted length. Use this over BooleanArray::from_trusted_len_iter when the iterator is trusted len but this crate does not mark it as such.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_trusted_len_iter<I, P>(iterator: I) -> Self where
P: Borrow<bool>,
I: TrustedLen<Item = Option<P>>,

Creates a BooleanArray from a TrustedLen.

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I, P>(
iterator: I
) -> Result<Self, E> where
P: Borrow<bool>,
I: Iterator<Item = Result<Option<P>, E>>,

Creates a BooleanArray from an falible iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn try_from_trusted_len_iter<E, I, P>(iterator: I) -> Result<Self, E> where
P: Borrow<bool>,
I: TrustedLen<Item = Result<Option<P>, E>>,

Creates a BooleanArray from a TrustedLen.

impl<'a> BooleanArray

pub fn iter(&'a self) -> ZipValidity<'a, bool, BitmapIter<'a>>

constructs a new iterator

pub fn values_iter(&'a self) -> BitmapIter<'a>

Returns an iterator of bool

impl BooleanArray

pub fn new_empty() -> Self

Returns a new empty BooleanArray.

pub fn new_null(length: usize) -> Self

Returns a new BooleanArray whose all slots are null / None.

pub fn from_data(values: Bitmap, validity: Option<Bitmap>) -> Self

The canonical method to create a BooleanArray out of low-end APIs.

Panics

This function panics iff:

The validity is not None and its length is different from values’s length

pub fn slice(&self, offset: usize, length: usize) -> Self

Returns a slice of this BooleanArray.

Implementation

This operation is O(1) as it amounts to essentially increase two ref counts.

Panic

This function panics iff offset + length >= self.len().

pub fn value(&self, i: usize) -> bool

Returns the element at index i as bool

pub unsafe fn value_unchecked(&self, i: usize) -> bool

Returns the element at index i as bool

Safety

Caller must be sure that i < self.len()

pub fn values(&self) -> &Bitmap

Returns the values bitmap of this BooleanArray.