Struct arrow2::array::DictionaryArray

pub struct DictionaryArray<K: DictionaryKey> { /* fields omitted */ }

An Array whose values are encoded by keys. This Array is useful when the cardinality of values is low compared to the length of the Array.

Implementations

impl<'a, K: DictionaryKey> DictionaryArray<K>

pub fn iter(
&'a self
) -> ZipValidity<'a, Box<dyn Scalar>, DictionaryValuesIter<'a, K>>

Returns an iterator of Option<Box<dyn Array>>

pub fn values_iter(&'a self) -> DictionaryValuesIter<'a, K>

Returns an iterator of Box<dyn Array>

impl<K: DictionaryKey> DictionaryArray<K>

pub fn new_empty(data_type: DataType) -> Self

Returns a new empty DictionaryArray.

pub fn new_null(data_type: DataType, length: usize) -> Self

Returns an DictionaryArray whose all elements are null

pub fn from_data(keys: PrimitiveArray<K>, values: Arc<dyn Array>) -> Self

The canonical method to create a new DictionaryArray.

pub fn slice(&self, offset: usize, length: usize) -> Self

Creates a new DictionaryArray by slicing the existing DictionaryArray.

pub fn keys(&self) -> &PrimitiveArray<K>

Returns the keys of the DictionaryArray. These keys can be used to fetch values from values.

pub fn values(&self) -> &Arc<dyn Array>

Returns the values of the DictionaryArray.

pub fn value(&self, index: usize) -> Box<dyn Scalar>

Returns the value of the DictionaryArray at position i.