Struct arrow2::array::FixedSizeListArray

pub struct FixedSizeListArray { /* private fields */ }

The Arrow’s equivalent to an immutable Vec<Option<[T; size]>> where T is an Arrow type. Cloning and slicing this struct is O(1).

Implementations

impl<'a> FixedSizeListArray

pub fn iter(
&'a self
) -> ZipValidity<'a, Box<dyn Array>, ListValuesIter<'a, FixedSizeListArray>>

Returns an iterator of Option<Box<dyn Array>>

pub fn values_iter(&'a self) -> ListValuesIter<'a, FixedSizeListArray>

Returns an iterator of Box<dyn Array>

impl FixedSizeListArray

pub fn try_new(
data_type: DataType,
values: Box<dyn Array>,
validity: Option<Bitmap>
) -> Result<Self, Error>

Creates a new FixedSizeListArray.

Errors

This function returns an error iff:

The data_type’s physical type is not crate::datatypes::PhysicalType::FixedSizeList

The data_type’s inner field’s data type is not equal to values.data_type.

The length of values is not a multiple of size in data_type

the validity’s length is not equal to values.len() / size.

pub fn new(
data_type: DataType,
values: Box<dyn Array>,
validity: Option<Bitmap>
) -> Self

Creates a new FixedSizeListArray.

Panics

This function panics iff:

The data_type’s physical type is not crate::datatypes::PhysicalType::FixedSizeList

The data_type’s inner field’s data type is not equal to values.data_type.

The length of values is not a multiple of size in data_type

the validity’s length is not equal to values.len() / size.

pub const fn size(&self) -> usize

Returns the size (number of elements per slot) of this FixedSizeListArray.

pub fn from_data(
data_type: DataType,
values: Box<dyn Array>,
validity: Option<Bitmap>
) -> Self

Alias for new

pub fn new_empty(data_type: DataType) -> Self

Returns a new empty FixedSizeListArray.

pub fn new_null(data_type: DataType, length: usize) -> Self

Returns a new null FixedSizeListArray.

pub fn boxed(self) -> Box<dyn Array>

Boxes self into a Box<dyn Array>.

pub fn arced(self) -> Arc<dyn Array>

Boxes self into a Arc<dyn Array>.

impl FixedSizeListArray

pub fn slice(&self, offset: usize, length: usize) -> Self

Returns a slice of this FixedSizeListArray.

Implementation

This operation is O(1).

Panics

panics iff offset + length > self.len()

pub unsafe fn slice_unchecked(&self, offset: usize, length: usize) -> Self

Returns a slice of this FixedSizeListArray.

Implementation

This operation is O(1).

Safety

The caller must ensure that offset + length <= self.len().

pub fn with_validity(self, validity: Option<Bitmap>) -> Self

Returns this FixedSizeListArray with a new validity.

Panic

This function panics iff validity.len() != self.len().

pub fn set_validity(&mut self, validity: Option<Bitmap>)

Sets the validity of this FixedSizeListArray.

Panics

This function panics iff validity.len() != self.len().

impl FixedSizeListArray

pub fn len(&self) -> usize

Returns the length of this array

pub fn validity(&self) -> Option<&Bitmap>

The optional validity.

pub fn values(&self) -> &Box<dyn Array>

Returns the inner array.

pub fn value(&self, i: usize) -> Box<dyn Array>

Returns the Vec<T> at position i.

Panic:

panics iff i >= self.len()

pub unsafe fn value_unchecked(&self, i: usize) -> Box<dyn Array>

Returns the Vec<T> at position i.

Safety

Caller must ensure that i < self.len()

impl FixedSizeListArray

pub fn default_datatype(data_type: DataType, size: usize) -> DataType

Returns a DataType consistent with FixedSizeListArray.