Struct arrow2::array::MutableBooleanArray

pub struct MutableBooleanArray { /* fields omitted */ }

The mutable version of BooleanArray. See MutableArray for more details.

Implementations

impl MutableBooleanArray

pub fn new() -> Self

pub fn with_capacity(capacity: usize) -> Self

pub fn reserve(&mut self, additional: usize)

pub fn from_data(values: MutableBitmap, validity: Option<MutableBitmap>) -> Self

pub fn push(&mut self, value: Option<bool>)

pub fn into_arc(self) -> Arc<dyn Array>

Converts itself into an Array.

impl MutableBooleanArray

Getters

pub fn values(&self) -> &MutableBitmap

Returns its values.

impl MutableBooleanArray

Setters

pub fn set(&mut self, index: usize, value: Option<bool>)

Sets position index to value. Note that if it is the first time a null appears in this array, this initializes the validity bitmap (O(N)).

Panic

Panics iff index is larger than self.len().

impl MutableBooleanArray

From implementations

pub fn from_trusted_len_values_iter<I: TrustedLen<Item = bool>>(
iterator: I
) -> Self

Creates a new MutableBooleanArray from an TrustedLen of bool.

pub fn from_slice<P: AsRef<[bool]>>(slice: P) -> Self

Creates a new MutableBooleanArray from a slice of bool.

pub unsafe fn from_trusted_len_iter_unchecked<I, P>(iterator: I) -> Self where
P: Borrow<bool>,
I: Iterator<Item = Option<P>>,

Creates a BooleanArray from an iterator of trusted length. Use this over BooleanArray::from_trusted_len_iter when the iterator is trusted len but this crate does not mark it as such.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_trusted_len_iter<I, P>(iterator: I) -> Self where
P: Borrow<bool>,
I: TrustedLen<Item = Option<P>>,

Creates a BooleanArray from a TrustedLen.

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I, P>(
iterator: I
) -> Result<Self, E> where
P: Borrow<bool>,
I: Iterator<Item = Result<Option<P>, E>>,

Creates a BooleanArray from an falible iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn try_from_trusted_len_iter<E, I, P>(iterator: I) -> Result<Self, E> where
P: Borrow<bool>,
I: TrustedLen<Item = Result<Option<P>, E>>,

Creates a BooleanArray from a TrustedLen.