Struct arrow2::array::MutableFixedSizeBinaryArray

pub struct MutableFixedSizeBinaryArray { /* private fields */ }

The Arrow’s equivalent to a mutable Vec<Option<[u8; size]>>. Converting a MutableFixedSizeBinaryArray into a FixedSizeBinaryArray is O(1).

Implementation

This struct does not allocate a validity until one is required (i.e. push a null to it).

Implementations

impl<'a> MutableFixedSizeBinaryArray

pub fn iter(
&'a self
) -> ZipValidity<&'a [u8], ChunksExact<'a, u8>, BitmapIter<'a>>

constructs a new iterator

pub fn iter_values(&'a self) -> ChunksExact<'a, u8>

Returns iterator over the values of MutableFixedSizeBinaryArray

impl MutableFixedSizeBinaryArray

pub fn from_data(
data_type: DataType,
values: Vec<u8>,
validity: Option<MutableBitmap>
) -> Self

Canonical method to create a new MutableFixedSizeBinaryArray.

pub fn new(size: usize) -> Self

Creates a new empty MutableFixedSizeBinaryArray.

pub fn with_capacity(size: usize, capacity: usize) -> Self

Creates a new MutableFixedSizeBinaryArray with capacity for capacity entries.

pub fn from<const N: usize, P: AsRef<[Option<[u8; N]>]>>(slice: P) -> Self

Creates a new MutableFixedSizeBinaryArray from a slice of optional [u8].

pub fn try_push<P: AsRef<[u8]>>(&mut self, value: Option<P>) -> Result<()>

tries to push a new entry to MutableFixedSizeBinaryArray.

Error

Errors iff the size of value is not equal to its own size.

pub fn push<P: AsRef<[u8]>>(&mut self, value: Option<P>)

pushes a new entry to MutableFixedSizeBinaryArray.

Panics

Panics iff the size of value is not equal to its own size.

pub fn len(&self) -> usize

Returns the length of this array

pub fn pop(&mut self) -> Option<Vec<u8>>

Pop the last entry from MutableFixedSizeBinaryArray. This function returns None iff this array is empty

pub fn try_from_iter<P: AsRef<[u8]>, I: IntoIterator<Item = Option<P>>>(
iter: I,
size: usize
) -> Result<Self>

Creates a new MutableFixedSizeBinaryArray from an iterator of values.

Errors

Errors iff the size of any of the value is not equal to its own size.

pub fn size(&self) -> usize

returns the (fixed) size of the MutableFixedSizeBinaryArray.

pub fn capacity(&self) -> usize

Returns the capacity of this array

pub fn value(&self, i: usize) -> &[u8]

Returns the element at index i as &[u8]

pub unsafe fn value_unchecked(&self, i: usize) -> &[u8]

Returns the element at index i as &[u8]

Safety

Assumes that the i < self.len.

pub fn reserve(&mut self, additional: usize)

Reserves additional slots.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the MutableFixedSizeBinaryArray to fit its current length.

impl MutableFixedSizeBinaryArray

Accessors

pub fn values(&self) -> &Vec<u8>

Returns its values.

pub fn values_mut_slice(&mut self) -> &mut [u8]

Returns a mutable slice of values.