Struct arrow2::array::MutableListArray

pub struct MutableListArray<O: Offset, M: MutableArray> { /* private fields */ }

The mutable version of ListArray.

Implementations

impl<O: Offset, M: MutableArray + Default> MutableListArray<O, M>

pub fn new() -> Self

Creates a new empty MutableListArray.

pub fn with_capacity(capacity: usize) -> Self

Creates a new MutableListArray with a capacity.

impl<O: Offset, M: MutableArray> MutableListArray<O, M>

pub fn new_from(values: M, data_type: DataType, capacity: usize) -> Self

Creates a new MutableListArray from a MutableArray and capacity.

pub fn new_with_field(values: M, name: &str, nullable: bool) -> Self

Creates a new MutableListArray from a MutableArray.

pub fn new_with_capacity(values: M, capacity: usize) -> Self

Creates a new MutableListArray from a MutableArray and capacity.

pub fn try_push_valid(&mut self) -> Result<()>

Needs to be called when a valid value was extended to this array. This is a relatively low level function, prefer try_push when you can.

pub fn mut_values(&mut self) -> &mut M

The values

pub fn offsets(&self) -> &Vec<O>

The offseta

pub fn values(&self) -> &M

The values

pub fn into_arc(self) -> Arc<dyn Array>

Converts itself into an Array.

pub fn into_box(self) -> Box<dyn Array>

converts itself into Box<dyn Array>

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the MutableListArray to fit its current length.