Struct arrow2::array::MutablePrimitiveArray

pub struct MutablePrimitiveArray<T: NativeType> { /* private fields */ }

The Arrow’s equivalent to Vec<Option<T>> where T is byte-size (e.g. i32). Converting a MutablePrimitiveArray into a PrimitiveArray is O(1).

Implementations

impl<'a, T: NativeType> MutablePrimitiveArray<T>

pub fn iter(&'a self) -> ZipValidity<'a, &'a T, Iter<'a, T>>

Returns an iterator over Option<T>

pub fn values_iter(&'a self) -> Iter<'a, T>

Returns an iterator of bool

impl<T: NativeType> MutablePrimitiveArray<T>

pub fn new() -> Self

Creates a new empty MutablePrimitiveArray.

pub fn with_capacity(capacity: usize) -> Self

Creates a new MutablePrimitiveArray with a capacity.

pub fn from_data(
data_type: DataType,
values: Vec<T>,
validity: Option<MutableBitmap>
) -> Self

Create a MutablePrimitiveArray out of low-end APIs.

Panics

This function panics iff:

data_type is not supported by the physical type

The validity is not None and its length is different from the values’s length

pub fn into_data(self) -> (DataType, Vec<T>, Option<MutableBitmap>)

Extract the low-end APIs from the MutablePrimitiveArray.

impl<T: NativeType> MutablePrimitiveArray<T>

pub fn with_capacity_from(capacity: usize, data_type: DataType) -> Self

Creates a new MutablePrimitiveArray from a capacity and DataType.

pub fn reserve(&mut self, additional: usize)

Reserves additional entries.

pub fn push(&mut self, value: Option<T>)

Adds a new value to the array.

pub fn extend_constant(&mut self, additional: usize, value: Option<T>)

Extends the MutablePrimitiveArray with a constant

pub fn extend_trusted_len<P, I>(&mut self, iterator: I) where
P: Borrow<T>,
I: TrustedLen<Item = Option<P>>,

Extends the MutablePrimitiveArray from an iterator of trusted len.

pub unsafe fn extend_trusted_len_unchecked<P, I>(&mut self, iterator: I) where
P: Borrow<T>,
I: Iterator<Item = Option<P>>,

Extends the MutablePrimitiveArray from an iterator of trusted len.

Safety

The iterator must be trusted len.

pub fn extend_trusted_len_values<I>(&mut self, iterator: I) where
I: TrustedLen<Item = T>,

Extends the MutablePrimitiveArray from an iterator of values of trusted len. This differs from extend_trusted_len which accepts in iterator of optional values.

pub unsafe fn extend_trusted_len_values_unchecked<I>(&mut self, iterator: I) where
I: Iterator<Item = T>,

Extends the MutablePrimitiveArray from an iterator of values of trusted len. This differs from extend_trusted_len_unchecked which accepts in iterator of optional values.

Safety

The iterator must be trusted len.

pub fn extend_from_slice(&mut self, items: &[T])

Extends the MutablePrimitiveArray from a slice

pub fn to(self, data_type: DataType) -> Self

Changes the arrays’ DataType, returning a new MutablePrimitiveArray. Use to change the logical type without changing the corresponding physical Type.

Implementation

This operation is O(1).

pub fn into_arc(self) -> Arc<dyn Array>

Converts itself into an Array.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the MutablePrimitiveArray to fit its current length.

impl<T: NativeType> MutablePrimitiveArray<T>

Accessors

pub fn values(&self) -> &Vec<T>

Returns its values.

pub fn values_mut_slice(&mut self) -> &mut [T]

Returns a mutable slice of values.

impl<T: NativeType> MutablePrimitiveArray<T>

Setters

pub fn set(&mut self, index: usize, value: Option<T>)

Sets position index to value. Note that if it is the first time a null appears in this array, this initializes the validity bitmap (O(N)).

Panic

Panics iff index is larger than self.len().

pub fn set_validity(&mut self, validity: Option<MutableBitmap>)

Sets the validity.

Panic

Panics iff the validity’s len is not equal to the existing values’ length.

pub fn set_values(&mut self, values: Vec<T>)

Sets values.

Panic

Panics iff the values’ length is not equal to the existing validity’s len.

impl<T: NativeType> MutablePrimitiveArray<T>

pub fn from_slice<P: AsRef<[T]>>(slice: P) -> Self

Creates a MutablePrimitiveArray from a slice of values.

pub unsafe fn from_trusted_len_iter_unchecked<I, P>(iterator: I) -> Self where
P: Borrow<T>,
I: Iterator<Item = Option<P>>,

Creates a MutablePrimitiveArray from an iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. size_hint().1 correctly reports its length.

pub fn from_trusted_len_iter<I, P>(iterator: I) -> Self where
P: Borrow<T>,
I: TrustedLen<Item = Option<P>>,

Creates a MutablePrimitiveArray from a TrustedLen.

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I, P>(
iter: I
) -> Result<Self, E> where
P: Borrow<T>,
I: IntoIterator<Item = Result<Option<P>, E>>,

Creates a MutablePrimitiveArray from an fallible iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn try_from_trusted_len_iter<E, I, P>(iterator: I) -> Result<Self, E> where
P: Borrow<T>,
I: TrustedLen<Item = Result<Option<P>, E>>,

Creates a MutablePrimitiveArray from an fallible iterator of trusted length.

pub fn from_trusted_len_values_iter<I: TrustedLen<Item = T>>(iter: I) -> Self

Creates a new MutablePrimitiveArray out an iterator over values

pub unsafe fn from_trusted_len_values_iter_unchecked<I: Iterator<Item = T>>(
iter: I
) -> Self

Creates a new MutablePrimitiveArray from an iterator over values

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.