Struct arrow2::array::Primitive

pub struct Primitive<T: NativeType> { /* fields omitted */ }

Auxiliary struct used to create a PrimitiveArray out of iterators. Primitive arrays are often built from this struct, that knows how to cheaply convert itself into a primitive array.

Implementations

impl<T: NativeType> Primitive<T>

pub fn from_slice<P: AsRef<[T]>>(slice: P) -> Self

pub fn from_vec(slice: Vec<T>) -> Self

impl<T: NativeType> Primitive<T>

pub fn with_capacity(capacity: usize) -> Self

Initializes a new Primitive with a pre-allocated number of slots.

pub unsafe fn from_trusted_len_iter_unchecked<I, P>(iterator: I) -> Self where
P: Borrow<T>,
I: Iterator<Item = Option<P>>,

Creates a Primitive from an iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. size_hint().1 correctly reports its length.

pub fn from_trusted_len_iter<I, P>(iterator: I) -> Self where
P: Borrow<T>,
I: TrustedLen<Item = Option<P>>,

Creates a Primitive from a TrustedLen.

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I, P>(
iter: I
) -> Result<Self, E> where
P: Borrow<T>,
I: IntoIterator<Item = Result<Option<P>, E>>,

Creates a PrimitiveArray from an fallible iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn try_from_trusted_len_iter<E, I, P>(iterator: I) -> Result<Self, E> where
P: Borrow<T>,
I: TrustedLen<Item = Result<Option<P>, E>>,

Creates a PrimitiveArray from an fallible iterator of trusted length.

impl<T: NativeType> Primitive<T>

pub fn new() -> Self

Initializes itself with a capacity.

pub fn to(self, data_type: DataType) -> PrimitiveArray<T>

Converts itself to a PrimitiveArray.

Panic

This panics if the DataType is not valid for this physical type.