Struct arrow2::array::PrimitiveArray

pub struct PrimitiveArray<T: NativeType> { /* fields omitted */ }

A PrimitiveArray is arrow’s equivalent to Vec<Option<T: NativeType>>, i.e. an array designed for highly performant operations on optionally nullable slots, backed by a physical type of a physical byte-width, such as i32 or f64. The size of this struct is O(1) as all data is stored behind an std::sync::Arc.

Implementations

impl<T: NativeType + NaturalDataType> PrimitiveArray<T>

pub fn from_values<I: IntoIterator<Item = T>>(iter: I) -> Self

Creates a new array out an iterator over values

pub fn from_slice<P: AsRef<[T]>>(slice: P) -> Self

Creates a new array out an iterator over values

impl<T: NativeType + NaturalDataType> PrimitiveArray<T>

pub fn from_trusted_len_values_iter<I: TrustedLen<Item = T>>(iter: I) -> Self

Creates a new array out an iterator over values

pub unsafe fn from_trusted_len_values_iter_unchecked<I: Iterator<Item = T>>(
iter: I
) -> Self

Creates a new PrimitiveArray from an iterator over values

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_trusted_len_iter<I: TrustedLen<Item = Option<T>>>(iter: I) -> Self

Creates a new PrimitiveArray from an iterator over optional values

pub unsafe fn from_trusted_len_iter_unchecked<I: Iterator<Item = Option<T>>>(
iter: I
) -> Self

Creates a new PrimitiveArray from an iterator over optional values

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

impl<'a, T: NativeType> PrimitiveArray<T>

pub fn iter(&'a self) -> ZipValidity<'a, &'a T, Iter<'a, T>>

constructs a new iterator

impl<T: NativeType> PrimitiveArray<T>

pub fn new_empty(data_type: DataType) -> Self

Returns a new empty PrimitiveArray.

pub fn new_null(data_type: DataType, length: usize) -> Self

Returns a new PrimitiveArray whose all slots are null / None.

pub fn from_data(
data_type: DataType,
values: Buffer<T>,
validity: Option<Bitmap>
) -> Self

The canonical method to create a PrimitiveArray out of low-end APIs.

Panics

This function panics iff:

data_type is not supported by the physical type

The validity is not None and its length is different from the values’s length

pub fn slice(&self, offset: usize, length: usize) -> Self

Returns a slice of this PrimitiveArray.

Implementation

This operation is O(1) as it amounts to increase two ref counts.

Panic

This function panics iff offset + length >= self.len().

pub fn values(&self) -> &Buffer<T>

The values Buffer.

pub fn value(&self, i: usize) -> T

Safe method to retrieve the value at slot i. Equivalent to self.values()[i].

pub unsafe fn value_unchecked(&self, i: usize) -> T

Returns the element at index i as T

Safety

Caller must be sure that i < self.len()

pub fn to(self, data_type: DataType) -> Self

Returns a new PrimitiveArray with a different logical type. This is O(1).

Panics

Panics iff the data_type is not supported for the physical type.