Struct arrow2::array::StructArray

pub struct StructArray { /* fields omitted */ }

A StructArray is a nested Array with an optional validity representing multiple Array with the same number of rows.

Example

use std::sync::Arc;
use arrow2::array::*;
use arrow2::datatypes::*;
let boolean = Arc::new(BooleanArray::from_slice(&[false, false, true, true])) as Arc<dyn Array>;
let int = Arc::new(Int32Array::from_slice(&[42, 28, 19, 31])) as Arc<dyn Array>;

let fields = vec![
    Field::new("b", DataType::Boolean, false),
    Field::new("c", DataType::Int32, false),
];

let array = StructArray::from_data(DataType::Struct(fields), vec![boolean, int], None);

Implementations

impl StructArray

pub fn new_empty(data_type: DataType) -> Self

Creates an empty StructArray.

pub fn new_null(data_type: DataType, length: usize) -> Self

Creates a null StructArray of length length.

pub fn from_data(
data_type: DataType,
values: Vec<Arc<dyn Array>>,
validity: Option<Bitmap>
) -> Self

Canonical method to create a StructArray.

Panics

fields are empty

values’s len is different from Fields’ length.

any element of values has a different length than the first element.

pub fn into_data(self) -> (Vec<Field>, Vec<Arc<dyn Array>>, Option<Bitmap>)

Deconstructs the StructArray into its individual components.

pub fn slice(&self, offset: usize, length: usize) -> Self

Creates a new StructArray that is a slice of self.

Panics

offset + length must be smaller than self.len().

Implementation

This operation is O(F) where F is the number of fields.

pub fn values(&self) -> &[Arc<dyn Array>]

Returns the values of this StructArray.

pub fn fields(&self) -> &[Field]

Returns the fields of this StructArray.

impl StructArray

pub fn get_fields(data_type: &DataType) -> &[Field]

Returns the fields the DataType::Struct.