Struct arrow2::array::Utf8Array

pub struct Utf8Array<O: Offset> { /* private fields */ }

A Utf8Array is arrow’s semantic equivalent of an immutable Vec<Option<String>>. Cloning and slicing this struct is O(1).

Example

use arrow2::bitmap::Bitmap;
use arrow2::buffer::Buffer;
use arrow2::array::Utf8Array;
let array = Utf8Array::<i32>::from([Some("hi"), None, Some("there")]);
assert_eq!(array.value(0), "hi");
assert_eq!(array.iter().collect::<Vec<_>>(), vec![Some("hi"), None, Some("there")]);
assert_eq!(array.values_iter().collect::<Vec<_>>(), vec!["hi", "", "there"]);
// the underlying representation
assert_eq!(array.validity(), Some(&Bitmap::from([true, false, true])));
assert_eq!(array.values(), &Buffer::from(b"hithere".to_vec()));
assert_eq!(array.offsets(), &Buffer::from(vec![0, 2, 2, 2 + 5]));

Generic parameter

The generic parameter Offset can only be i32 or i64 and tradeoffs maximum array length with memory usage:

the sum of lengths of all elements cannot exceed Offset::MAX

the total size of the underlying data is array.len() * size_of::<Offset>() + sum of lengths of all elements

Safety

The following invariants hold:

Two consecutives offsets casted (as) to usize are valid slices of values.

A slice of values taken from two consecutives offsets is valid utf8.

len is equal to validity.len(), when defined.

Implementations

impl<O: Offset> Utf8Array<O>

pub fn try_new(
data_type: DataType,
offsets: Buffer<O>,
values: Buffer<u8>,
validity: Option<Bitmap>
) -> Result<Self>

Returns a Utf8Array created from its internal representation.

Errors

This function returns an error iff:

the offsets are not monotonically increasing

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len() - 1.

The data_type’s crate::datatypes::PhysicalType is not equal to either Utf8 or LargeUtf8.

The values between two consecutive offsets are not valid utf8

Implementation

This function is O(N) - checking monotinicity and utf8 is O(N)

pub fn from_slice<T: AsRef<str>, P: AsRef<[T]>>(slice: P) -> Self

Returns a Utf8Array from a slice of &str.

A convenience method that uses Self::from_trusted_len_values_iter.

pub fn from<T: AsRef<str>, P: AsRef<[Option<T>]>>(slice: P) -> Self

Returns a new Utf8Array from a slice of &str.

A convenience method that uses Self::from_trusted_len_iter.

pub fn iter(&self) -> ZipValidity<'_, &str, Utf8ValuesIter<'_, O>>

Returns an iterator of Option<&str>

pub fn values_iter(&self) -> Utf8ValuesIter<'_, O>

Returns an iterator of &str

pub fn len(&self) -> usize

Returns the length of this array

pub fn value(&self, i: usize) -> &str

Returns the value of the element at index i, ignoring the array’s validity.

Panic

This function panics iff i >= self.len.

pub unsafe fn value_unchecked(&self, i: usize) -> &str

Returns the value of the element at index i, ignoring the array’s validity.

Safety

This function is safe iff i < self.len.

pub fn data_type(&self) -> &DataType

Returns the DataType of this array.

pub fn values(&self) -> &Buffer<u8>

Returns the values of this Utf8Array.

pub fn offsets(&self) -> &Buffer<O>

Returns the offsets of this Utf8Array.

pub fn validity(&self) -> Option<&Bitmap>

The optional validity.

pub fn slice(&self, offset: usize, length: usize) -> Self

Returns a slice of this Utf8Array.

Implementation

This operation is O(1) as it amounts to essentially increase two ref counts.

Panic

This function panics iff offset + length >= self.len().

pub unsafe fn slice_unchecked(&self, offset: usize, length: usize) -> Self

Returns a slice of this Utf8Array.

Implementation

This operation is O(1) as it amounts to essentially increase two ref counts.

Safety

The caller must ensure that offset + length <= self.len().

pub fn boxed(self) -> Box<dyn Array>

Boxes self into a Box<dyn Array>.

pub fn arced(self) -> Arc<dyn Array>

Boxes self into a std::sync::Arc<dyn Array>.

pub fn with_validity(self, validity: Option<Bitmap>) -> Self

Returns this Utf8Array with a new validity.

Panics

This function panics iff validity.len() != self.len().

pub fn set_validity(&mut self, validity: Option<Bitmap>)

Sets the validity of this Utf8Array.

Panics

This function panics iff validity.len() != self.len().

pub fn into_mut(self) -> Either<Self, MutableUtf8Array<O>>

Try to convert this Utf8Array to a MutableUtf8Array

pub fn new_empty(data_type: DataType) -> Self

Returns a new empty Utf8Array.

The array is guaranteed to have no elements nor validity.

pub fn new_null(data_type: DataType, length: usize) -> Self

Returns a new Utf8Array whose all slots are null / None.

pub fn default_data_type() -> DataType

Returns a default DataType of this array, which depends on the generic parameter O: DataType::Utf8 or DataType::LargeUtf8

pub unsafe fn try_new_unchecked(
data_type: DataType,
offsets: Buffer<O>,
values: Buffer<u8>,
validity: Option<Bitmap>
) -> Result<Self>

Creates a new Utf8Array without checking for offsets monotinicity nor utf8-validity

Errors

This function returns an error iff:

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len() - 1.

The data_type’s crate::datatypes::PhysicalType is not equal to either Utf8 or LargeUtf8.

Safety

This function is unsound iff:

the offsets are not monotonically increasing

The values between two consecutive offsets are not valid utf8

Implementation

This function is O(1)

pub fn new(
data_type: DataType,
offsets: Buffer<O>,
values: Buffer<u8>,
validity: Option<Bitmap>
) -> Self

Creates a new Utf8Array.

Panics

This function panics iff:

the offsets are not monotonically increasing

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len() - 1.

The data_type’s crate::datatypes::PhysicalType is not equal to either Utf8 or LargeUtf8.

The values between two consecutive offsets are not valid utf8

Implementation

This function is O(N) - checking monotinicity and utf8 is O(N)

pub unsafe fn new_unchecked(
data_type: DataType,
offsets: Buffer<O>,
values: Buffer<u8>,
validity: Option<Bitmap>
) -> Self

Creates a new Utf8Array without checking for offsets monotinicity.

Errors

This function returns an error iff:

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len() - 1.

The data_type’s crate::datatypes::PhysicalType is not equal to either Utf8 or LargeUtf8.

Safety

This function is unsound iff:

the offsets are not monotonically increasing

The values between two consecutive offsets are not valid utf8

Implementation

This function is O(1)

pub fn from_trusted_len_values_iter<T: AsRef<str>, I: TrustedLen<Item = T>>(
iterator: I
) -> Self

Returns a (non-null) Utf8Array created from a TrustedLen of &str.

Implementation

This function is O(N)

pub fn from_iter_values<T: AsRef<str>, I: Iterator<Item = T>>(
iterator: I
) -> Self

Creates a new Utf8Array from a Iterator of &str.

pub unsafe fn from_trusted_len_iter_unchecked<I, P>(iterator: I) -> Selfwhere
P: AsRef<str>,
I: Iterator<Item = Option<P>>,

Creates a Utf8Array from an iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_trusted_len_iter<I, P>(iterator: I) -> Selfwhere
P: AsRef<str>,
I: TrustedLen<Item = Option<P>>,

Creates a Utf8Array from an iterator of trusted length.

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I, P>(
iterator: I
) -> Result<Self, E>where
P: AsRef<str>,
I: IntoIterator<Item = Result<Option<P>, E>>,

Creates a Utf8Array from an falible iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn try_from_trusted_len_iter<E, I, P>(iter: I) -> Result<Self, E>where
P: AsRef<str>,
I: TrustedLen<Item = Result<Option<P>, E>>,

Creates a Utf8Array from an fallible iterator of trusted length.

pub fn from_data(
data_type: DataType,
offsets: Buffer<O>,
values: Buffer<u8>,
validity: Option<Bitmap>
) -> Self

Alias for new

pub unsafe fn from_data_unchecked(
data_type: DataType,
offsets: Buffer<O>,
values: Buffer<u8>,
validity: Option<Bitmap>
) -> Self

Alias for Self::new_unchecked

Safety

This function is unsafe iff:

the offsets are not monotonically increasing

The values between two consecutive offsets are not valid utf8