Trait arrow2::array::Builder

pub trait Builder<T>: TryFromIterator<Option<T>> + IntoArray {
    fn with_capacity(capacity: usize) -> Self;

    fn push(&mut self, item: Option<&T>);

    fn try_push(&mut self, item: Option<&T>) -> Result<()> { ... }
}

A trait describing the ability of a struct to

Required methods

fn with_capacity(capacity: usize) -> Self

Create the builder

fn push(&mut self, item: Option<&T>)

Push a new item to the builder. This operation may panic if the container cannot hold more items. For example, if all possible keys are exausted when building a dictionary.

Loading content...
Provided methods

fn try_push(&mut self, item: Option<&T>) -> Result<()>

Falible version of push, on which the operation errors instead of panicking.

Loading content...