Trait arrow2::array::FromFfi

pub trait FromFfi<T: ArrowArrayRef>: Sized {
    unsafe fn try_from_ffi(array: T) -> Result<Self>;
}

Trait describing how a struct imports into itself from the C data interface (FFI).

Required methods

unsafe fn try_from_ffi(array: T) -> Result<Self>

Convert itself from FFI.

Safety

This function is intrinsically unsafe as it requires the FFI to be made according to the C data interface