Trait arrow2::array::MutableArray

pub trait MutableArray: Debug {
    fn data_type(&self) -> &DataType;

    fn len(&self) -> usize;

    fn validity(&self) -> &Option<MutableBitmap>;

    fn as_arc(&mut self) -> Arc<dyn Array>;

    fn as_any(&self) -> &dyn Any;

    fn as_mut_any(&mut self) -> &mut dyn Any;

    fn push_null(&mut self);

    fn is_empty(&self) -> bool { ... }

    fn is_valid(&self, index: usize) -> bool { ... }
}

A trait describing a mutable array; i.e. an array whose values can be changed. Mutable arrays are not Send + Sync and cannot be cloned but can be mutated in place, thereby making them useful to perform numeric operations without allocations. As in Array, concrete arrays (such as MutablePrimitiveArray) implement how they are mutated.

Required methods

fn data_type(&self) -> &DataType

The DataType of the array.

fn len(&self) -> usize

The length of the array.

fn validity(&self) -> &Option<MutableBitmap>

The optional validity of the array.

fn as_arc(&mut self) -> Arc<dyn Array>

Convert itself to an (immutable) Array.

fn as_any(&self) -> &dyn Any

Convert to Any, to enable dynamic casting.

fn as_mut_any(&mut self) -> &mut dyn Any

Convert to mutable Any, to enable dynamic casting.

fn push_null(&mut self)

Adds a new null element to the array.

Provided methods

fn is_empty(&self) -> bool

Whether the array is empty.

fn is_valid(&self, index: usize) -> bool

Whether index is valid / set.

Panic

Panics if index >= self.len().