Trait arrow2::array::Preallocate

pub trait Preallocate {
    fn with_capacity(capacity: usize) -> Self
    where
        Self: Sized;
}

A trait describing an array with a backing store that can be preallocated to a given size.

Required Methods

fn with_capacity(capacity: usize) -> Self where
Self: Sized,

Create this array with a given capacity.