Trait arrow2::array::TryExtend

pub trait TryExtend<A> {
    fn try_extend<I: IntoIterator<Item = A>>(&mut self, iter: I) -> Result<()>;
}

A trait describing the ability of a struct to create itself from a iterator. This is similar to Extend, but accepted the creation to error.

Required methods

fn try_extend<I: IntoIterator<Item = A>>(&mut self, iter: I) -> Result<()>

Loading content...