Trait arrow2::array::TryExtendFromSelf

pub trait TryExtendFromSelf {
    fn try_extend_from_self(&mut self, other: &Self) -> Result<()>;
}

A trait describing the ability of a struct to extend from a reference of itself. Specialization of TryExtend.

Required Methods

fn try_extend_from_self(&mut self, other: &Self) -> Result<()>

Tries to extend itself with elements from other, failing only on overflow.