Trait arrow2::array::TryFromIterator

pub trait TryFromIterator<A>: Sized {
    fn try_from_iter<T: IntoIterator<Item = Result<A>>>(iter: T) -> Result<Self>;
}

A trait describing the ability of a struct to create itself from a falible iterator Used in the context of creating arrays from non-sized iterators.

Required methods

fn try_from_iter<T: IntoIterator<Item = Result<A>>>(iter: T) -> Result<Self>

Loading content...