Module arrow2::bitmap

Structs

Bitmap
An immutable container whose API is optimized to handle bitmaps. All quantities on this container’s API are measured in bits.

BitmapIter
An iterator of bits according to the LSB format

MutableBitmap
The mutable counterpart of Bitmap.

Functions

binary
Apply a bitwise operation op to two inputs and return the result as a Bitmap.

quaternary
Apply a bitwise operation op to four inputs and return the result as a Bitmap.

ternary
Apply a bitwise operation op to three inputs and return the result as a Bitmap.

unary
Apply a bitwise operation op to one input and return the result as a Bitmap.