Struct arrow2::bitmap::Bitmap

pub struct Bitmap { /* fields omitted */ }

An immutable container whose API is optimized to handle bitmaps. All quantities on this container’s API are measured in bits.

Implementation

memory on this container is sharable across thread boundaries

Cloning Bitmap is O(1)

Slicing Bitmap is O(1)

Implementations

impl Bitmap

pub fn new() -> Self

Initializes an empty Bitmap.

pub fn new_zeroed(length: usize) -> Self

Initializes an new Bitmap filled with unset values.

pub fn len(&self) -> usize

Returns the length of the Bitmap.

pub fn is_empty(&self) -> bool

Returns whether Bitmap is empty

pub fn from_u8_buffer(buffer: MutableBuffer<u8>, length: usize) -> Self

Creates a new Bitmap from MutableBuffer and a length.

Panic

Panics iff length <= buffer.len() * 8

pub fn from_u8_slice<T: AsRef<[u8]>>(buffer: T, length: usize) -> Self

Creates a new Bitmap from [Bytes] and a length.

Panic

Panics iff length <= bytes.len() * 8

pub fn null_count_range(&self, offset: usize, length: usize) -> usize

Counts the nulls (unset bits) starting from offset bits and for length bits.

pub fn null_count(&self) -> usize

Returns the number of unset bits on this Bitmap.

pub fn slice(self, offset: usize, length: usize) -> Self

Slices self, offseting by offset and truncating up to length bits.

Panic

Panics iff self.offset + offset + length <= self.bytes.len() * 8, i.e. if the offset and length exceeds the allocated capacity of self.

pub fn get_bit(&self, i: usize) -> bool

Returns whether the bit at position i is set.

Panics

Panics iff i >= self.len().

pub fn get(&self, i: usize) -> Option<bool>

Returns whether the bit at position i is set.

pub unsafe fn get_bit_unchecked(&self, i: usize) -> bool

Unsafely returns whether the bit at position i is set.

Safety

Unsound iff i >= self.len().

impl Bitmap

pub fn chunks<T: BitChunk>(&self) -> BitChunks<'_, T>

Returns an iterator over bits in chunks of T, which is useful for bit operations.

impl Bitmap

pub unsafe fn from_trusted_len_iter_unchecked<I: Iterator<Item = bool>>(
iterator: I
) -> Self

Creates a new Bitmap from an iterator of booleans.

Safety

The iterator must report an accurate length.

pub fn from_trusted_len_iter<I: TrustedLen<Item = bool>>(iterator: I) -> Self

Creates a new Bitmap from an iterator of booleans.

pub fn try_from_trusted_len_iter<E, I: TrustedLen<Item = Result<bool, E>>>(
iterator: I
) -> Result<Self, E>

Creates a new Bitmap from a fallible iterator of booleans.

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I: Iterator<Item = Result<bool, E>>>(
iterator: I
) -> Result<Self, E>

Creates a new Bitmap from a fallible iterator of booleans.

Safety

The iterator must report an accurate length.

impl<'a> Bitmap

pub fn iter(&'a self) -> BitmapIter<'a>

constructs a new iterator