Struct arrow2::bitmap::MutableBitmap

pub struct MutableBitmap { /* fields omitted */ }

A container to store booleans. MutableBitmap is semantically equivalent to Vec<bool>, but each value is stored as a single bit, thereby achieving a compression of 8x. This container is the counterpart of MutableBuffer for boolean values. MutableBitmap can be converted to a Bitmap at O(1).

Implementation

This container is backed by MutableBuffer<u8>.

Implementations

impl MutableBitmap

pub fn new() -> Self

Initializes an empty MutableBitmap.

pub fn from_len_zeroed(length: usize) -> Self

Initializes a zeroed MutableBitmap.

pub fn with_capacity(capacity: usize) -> Self

Initializes an a pre-allocated MutableBitmap with capacity for capacity bits.

pub fn reserve(&mut self, additional: usize)

Initializes an a pre-allocated MutableBitmap with capacity for capacity bits.

pub fn push(&mut self, value: bool)

Pushes a new bit to the MutableBitmap, re-sizing it if necessary.

pub fn capacity(&self) -> usize

pub unsafe fn push_unchecked(&mut self, value: bool)

Pushes a new bit to the MutableBitmap

Safety

The caller must ensure that the MutableBitmap has sufficient capacity.

pub fn null_count(&self) -> usize

Returns the number of unset bits on this MutableBitmap.

pub fn len(&self) -> usize

Returns the length of the MutableBitmap.

pub fn is_empty(&self) -> bool

Returns whether MutableBitmap is empty.

pub fn extend_constant(&mut self, additional: usize, value: bool)

Extends MutableBitmap by additional values of constant value.

pub fn get(&self, index: usize) -> bool

Returns whether the position index is set.

Panics

Panics iff index >= self.len().

pub fn set(&mut self, index: usize, value: bool)

Sets the position index to value

Panics

Panics iff index >= self.len().

impl MutableBitmap

pub fn from_buffer(buffer: MutableBuffer<u8>, length: usize) -> Self

impl MutableBitmap

pub fn extend_from_trusted_len_iter<I: TrustedLen<Item = bool>>(
&mut self,
iterator: I
)

Extends self from a TrustedLen iterator.

pub unsafe fn extend_from_trusted_len_iter_unchecked<I: Iterator<Item = bool>>(
&mut self,
iterator: I
)

Extends self from an iterator of trusted len.

Safety

The caller must guarantee that the iterator has a trusted len.

pub unsafe fn from_trusted_len_iter_unchecked<I>(iterator: I) -> Self where
I: Iterator<Item = bool>,

Creates a new Bitmap from an iterator of booleans.

Safety

The iterator must report an accurate length.

pub fn from_trusted_len_iter<I>(iterator: I) -> Self where
I: TrustedLen<Item = bool>,

Creates a new Bitmap from an iterator of booleans.

pub fn try_from_trusted_len_iter<E, I>(iterator: I) -> Result<Self, E> where
I: TrustedLen<Item = Result<bool, E>>,

Creates a new Bitmap from an iterator of booleans.

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I>(
iterator: I
) -> Result<Self, E> where
I: Iterator<Item = Result<bool, E>>,

Creates a new Bitmap from an falible iterator of booleans.

Safety

The caller must guarantee that the iterator is TrustedLen.