Trait arrow2::bitmap::utils::BitChunk

pub trait BitChunk: Sealed + NativeType + Binary + BitAnd<Output = Self> + ShlAssign + Not<Output = Self> + ShrAssign<usize> + ShlAssign<usize> + Shl<usize, Output = Self> + Eq + BitAndAssign + BitOr<Output = Self> {
    fn one() -> Self;

    fn zero() -> Self;

    fn to_ne_bytes(self) -> Self::Bytes;

    fn from_ne_bytes(v: Self::Bytes) -> Self;
}

A chunk of bits. This is used to create masks of a given length whose width is 1 bit. In simd_packed notation, this corresponds to m1xY.

Required methods

fn one() -> Self

A value with a single bit set at the most right position.

fn zero() -> Self

A value with no bits set.

fn to_ne_bytes(self) -> Self::Bytes

convert itself into bytes.

fn from_ne_bytes(v: Self::Bytes) -> Self

convert itself from bytes.