Module arrow2::buffer

Contains containers for all Arrow sized types (e.g. i32), Buffer and MutableBuffer.

Structs

Buffer
Buffer is a contiguous memory region that can be shared across thread boundaries. The easiest way to think about Buffer<T> is being equivalent to an immutable Vec<T>, with the following differences:

MutableBuffer
A MutableBuffer is this crates’ interface to store types that are byte-like such as i32. It behaves like a Vec, with the following differences: