Struct arrow2::buffer::Buffer

pub struct Buffer<T: NativeType> { /* fields omitted */ }

Buffer represents a contiguous memory region that can be shared with other buffers and across thread boundaries.

Implementations

impl<T: NativeType> Buffer<T>

pub fn new() -> Self

pub fn new_zeroed(length: usize) -> Self

pub fn from_bytes(bytes: Bytes<T>) -> Self

Auxiliary method to create a new Buffer

pub fn len(&self) -> usize

Returns the number of bytes in the buffer

pub fn is_empty(&self) -> bool

Returns whether the buffer is empty.

pub fn as_slice(&self) -> &[T]

Returns the byte slice stored in this buffer

pub fn slice(self, offset: usize, length: usize) -> Self

Returns a new Buffer that is a slice of this buffer starting at offset. Doing so allows the same memory region to be shared between buffers.

Panics

Panics iff offset is larger than len.

pub fn as_ptr(&self) -> *const T

Returns a pointer to the start of this buffer.

impl<T: NativeType> Buffer<T>

pub fn from_trusted_len_iter<I: TrustedLen<Item = T>>(iterator: I) -> Self

Creates a Buffer from an Iterator with a trusted length. Prefer this to collect whenever possible, as it often enables auto-vectorization.

Example

let v = vec![1u32];
let iter = v.iter().map(|x| x * 2);
let buffer = unsafe { Buffer::from_trusted_len_iter(iter) };
assert_eq!(buffer.len(), 1)

pub fn try_from_trusted_len_iter<E, I: TrustedLen<Item = Result<T, E>>>(
iterator: I
) -> Result<Self, E>

Safety

This method assumes that the iterator’s size is correct and is undefined behavior to use it on an iterator that reports an incorrect length.

pub unsafe fn from_trusted_len_iter_unchecked<I: Iterator<Item = T>>(
iterator: I
) -> Self

Creates a Buffer from an Iterator with a trusted (upper) length.

Safety

This method assumes that the iterator’s size is correct and is undefined behavior to use it on an iterator that reports an incorrect length.

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I: Iterator<Item = Result<T, E>>>(
iterator: I
) -> Result<Self, E>

Safety

This method assumes that the iterator’s size is correct and is undefined behavior to use it on an iterator that reports an incorrect length.