Module arrow2::compute::aggregate

Available on crate feature compute_aggregate only.

Contains different aggregation functions

Traits

SimdOrd

Trait describing a type describing multiple lanes with an order relationship consistent with the same order of T.

Sum

Object that can reduce itself to a number. This is used in the context of SIMD to reduce a MD (e.g. [f32; 16]) into a single number (f32).

Functions

can_max

Whether max supports data_type

can_min

Whether min supports data_type

can_sum

Whether sum supports data_type

estimated_bytes_size

Returns the total (heap) allocated size of the array in bytes.

max

Returns the maximum of Array. The scalar is null when all elements are null.

max_binary

Returns the maximum value in the binary array, according to the natural order.

max_boolean

Returns the maximum value in the boolean array

max_primitive

Returns the maximum value in the array, according to the natural order. For floating point arrays any NaN values are considered to be greater than any other non-null value

max_string

Returns the maximum value in the string array, according to the natural order.

min

Returns the minimum of Array. The scalar is null when all elements are null.

min_binary

Returns the minimum value in the binary array, according to the natural order.

min_boolean

Returns the minimum value in the boolean array.

min_primitive

Returns the minimum value in the array, according to the natural order. For floating point arrays any NaN values are considered to be greater than any other non-null value

min_string

Returns the minimum value in the string array, according to the natural order.

sum

Returns the sum of all elements in array as a Scalar of the same physical and logical types as array.

sum_primitive

Returns the sum of values in the array.

sum_slice

Compute the sum of a slice