Trait arrow2::compute::aggregate::SimdOrd

pub trait SimdOrd<T> {
    const MIN: T;
    const MAX: T;

    fn max_element(self) -> T;

    fn min_element(self) -> T;

    fn max(self, x: Self) -> Self;

    fn min(self, x: Self) -> Self;

    fn new_min() -> Self;

    fn new_max() -> Self;
}

This is supported on crate feature compute_aggregate only.

Trait describing a type describing multiple lanes with an order relationship consistent with the same order of T.

Associated Constants

const MIN: T

The minimum value

const MAX: T

The maximum value

Required methods

fn max_element(self) -> T

reduce itself to the minimum

fn min_element(self) -> T

reduce itself to the maximum

fn max(self, x: Self) -> Self

lane-wise maximum between two instances

fn min(self, x: Self) -> Self

lane-wise minimum between two instances

fn new_min() -> Self

returns a new instance with all lanes equal to MIN

fn new_max() -> Self

returns a new instance with all lanes equal to MAX