Function arrow2::compute::arithmetics::basic::add::saturating_add

pub fn saturating_add<T>(
    lhs: &PrimitiveArray<T>, 
    rhs: &PrimitiveArray<T>
) -> Result<PrimitiveArray<T>> where
    T: NativeType + SaturatingAdd<Output = T>, 

Saturating addition of two primitive arrays. If the result from the sum is larger than the possible number for this type, the result for the operation will be the saturated value.

Examples

use arrow2::compute::arithmetics::basic::add::saturating_add;
use arrow2::array::PrimitiveArray;
use arrow2::datatypes::DataType;

let a = PrimitiveArray::from(&vec![Some(100i8)]).to(DataType::Int8);
let b = PrimitiveArray::from(&vec![Some(100i8)]).to(DataType::Int8);
let result = saturating_add(&a, &b).unwrap();
let expected = PrimitiveArray::from(&vec![Some(127)]).to(DataType::Int8);
assert_eq!(result, expected);