Module arrow2::compute::arithmetics::basic::mul

This is supported on crate feature compute only.

Definition of basic mul operations with primitive arrays

Functions

checked_mul

Checked multiplication of two primitive arrays. If the result from the multiplications overflows, the validity for that index is changed returned.

checked_mul_scalar

Checked multiplication of a scalar T to a primitive array of type T. If the result from the multiplication overflows, then the validity for that index is changed to None

mul

Multiplies two primitive arrays with the same type. Panics if the multiplication of one pair of values overflows.

mul_scalar

Multiply a scalar T to a primitive array of type T. Panics if the multiplication of the values overflows.

overflowing_mul

Overflowing multiplication of two primitive arrays. If the result from the mul overflows, the result for the operation will be an array with overflowed values and a validity array indicating the overflowing elements from the array.

overflowing_mul_scalar

Overflowing multiplication of a scalar T to a primitive array of type T. If the result from the mul overflows for this type, then the result will be an array with overflowed values and a validity array indicating the overflowing elements from the array

saturating_mul

Saturating multiplication of two primitive arrays. If the result from the multiplication overflows, the result for the operation will be the saturated value.

saturating_mul_scalar

Saturated multiplication of a scalar T to a primitive array of type T. If the result from the mul overflows for this type, then the result will be saturated