Function arrow2::compute::arithmetics::basic::rem::rem

pub fn rem<T>(
    lhs: &PrimitiveArray<T>, 
    rhs: &PrimitiveArray<T>
) -> Result<PrimitiveArray<T>> where
    T: NativeType + Rem<Output = T>, 

Remainder of two primitive arrays with the same type. Panics if the divisor is zero of one pair of values overflows.

Examples

use arrow2::compute::arithmetics::basic::rem::rem;
use arrow2::array::Int32Array;

let a = Int32Array::from(&[Some(10), Some(7)]);
let b = Int32Array::from(&[Some(5), Some(6)]);
let result = rem(&a, &b).unwrap();
let expected = Int32Array::from(&[Some(0), Some(1)]);
assert_eq!(result, expected)