Function arrow2::compute::arithmetics::decimal::mul_scalar

pub fn mul_scalar(
    lhs: &PrimitiveArray<i128>,
    rhs: &PrimitiveScalar<i128>
) -> PrimitiveArray<i128>

Available on crate feature compute_arithmetics only.

Multiply a decimal PrimitiveArray with a PrimitiveScalar with the same precision and scale. If the precision and scale is different, then an InvalidArgumentError is returned. This function panics if the multiplied numbers result in a number larger than the possible number for the selected precision.