Module arrow2::compute::arithmetics::decimal::sub

This is supported on crate feature compute only.

Defines the subtract arithmetic kernels for Decimal PrimitiveArrays.

Functions

adaptive_sub

Adaptive subtract of two decimal primitive arrays with different precision and scale. If the precision and scale is different, then the smallest scale and precision is adjusted to the largest precision and scale. If during the addition one of the results is smaller than the min possible value, the result precision is changed to the precision of the min value

checked_sub

Checked subtract of two decimal primitive arrays with the same precision and scale. If the precision and scale is different, then an InvalidArgumentError is returned. If the result from the sub is larger than the possible number with the selected precision (overflowing), then the validity for that index is changed to None

saturating_sub

Saturated subtraction of two decimal primitive arrays with the same precision and scale. If the precision and scale is different, then an InvalidArgumentError is returned. If the result from the sum is smaller than the possible number with the selected precision then the resulted number in the arrow array is the minimum number for the selected precision.

sub

Subtract two decimal primitive arrays with the same precision and scale. If the precision and scale is different, then an InvalidArgumentError is returned. This function panics if the subtracted numbers result in a number smaller than the possible number for the selected precision.