Function arrow2::compute::arithmetics::add_scalar

pub fn add_scalar(lhs: &dyn Array, rhs: &dyn Scalar) -> Box<dyn Array>

Available on crate feature compute_arithmetics only.

Adds an Array and a Scalar.

Panic

This function panics iff

the opertion is not supported for the logical types (use can_add to check)

the arrays have a different length

one of the arrays is a timestamp with timezone and the timezone is not valid.