Trait arrow2::compute::arithmetics::NativeArithmetics

pub trait NativeArithmetics: NativeType { }

This is supported on crate feature compute_arithmetics only.

Trait describing a NativeType whose semantics of arithmetic in Arrow equals the semantics in Rust. A counter example is i128, that in arrow represents a decimal while in rust represents a signed integer.