Module arrow2::compute::boolean

Functions

and
Performs AND operation on two arrays. If either left or right value is null then the result is also null.

is_not_null
Returns a non-null BooleanArray with whether each value of the array is not null.

is_null
Returns a non-null BooleanArray with whether each value of the array is null.

not
Performs unary NOT operation on an arrays. If value is null then the result is also null.

or
Performs OR operation on two arrays. If either left or right value is null then the result is also null.