Function arrow2::compute::cast::dictionary_to_dictionary_values

pub fn dictionary_to_dictionary_values<K: DictionaryKey>(
    from: &DictionaryArray<K>, 
    values_type: &DataType
) -> Result<DictionaryArray<K>>

Casts a DictionaryArray to a new DictionaryArray by keeping the keys and casting the values to values_type.

Errors

This function errors if the values are not castable to values_type