Function arrow2::compute::cast::utf8_to_dictionary

pub fn utf8_to_dictionary<O: Offset, K: DictionaryKey>(
    from: &Utf8Array<O>
) -> Result<DictionaryArray<K>>

Cast Utf8Array to DictionaryArray, also known as packing.

Errors

This function errors if the maximum key is smaller than the number of distinct elements in the array.