Module arrow2::compute::comparison

This is supported on crate feature compute only.

Basic comparison kernels.

The module contains functions that compare either an array and a scalar or two arrays of the same DataType. The scalar-oriented functions are suffixed with _scalar. In general, these comparison functions receive as inputs the two items for comparison and an Operator which specifies the type of comparison that will be conducted, such as <= (Operator::LtEq).

Much like the parent module crate::compute, the comparison functions have two variants - a statically typed one (primitive_compare) which expects concrete types such as Int8Array and a dynamically typed variant (compare) that compares values of type &dyn Array and errors if the underlying concrete types mismsatch. The statically-typed functions are appropriately prefixed with the concrete types they expect.

Also note that the scalar-based comparison functions for the concrete types, like primitive_compare_scalar, are infallible and always return a BooleanArray while the rest of the functions always wrap the returned array in a Result due to their internal checks of the data types and lengths.

Examples

Compare two PrimitiveArrays:

use arrow2::compute::comparison::{primitive_compare, Operator};

let array1 = PrimitiveArray::<i32>::from([Some(1), None, Some(2)]);
let array2 = PrimitiveArray::<i32>::from([Some(1), None, Some(1)]);
let result = primitive_compare(&array1, &array2, Operator::Gt)?;
assert_eq!(result, BooleanArray::from([Some(false), None, Some(true)]));

Compare two dynamically-typed arrays (trait objects):

use arrow2::compute::comparison::{compare, Operator};

let array1: &dyn Array = &PrimitiveArray::<f64>::from(&[Some(10.0), None, Some(20.0)]);
let array2: &dyn Array = &PrimitiveArray::<f64>::from(&[Some(10.0), None, Some(10.0)]);
let result = compare(array1, array2, Operator::LtEq)?;
assert_eq!(result, BooleanArray::from([Some(true), None, Some(false)]));

Compare an array of strings to a “scalar”, i.e a word (note that we use Operator::Neq):

use arrow2::compute::comparison::{utf8_compare_scalar, Operator};

let array = Utf8Array::<i32>::from([Some("compute"), None, Some("compare")]);
let word = Utf8Scalar::new(Some("compare"));
let result = utf8_compare_scalar(&array, &word, Operator::Neq);
assert_eq!(result, BooleanArray::from([Some(true), None, Some(false)]));

Enums

Operator

Comparison operators, such as > (Operator::Gt)

Traits

Simd8

NativeType that supports a representation of 8 lanes

Simd8Lanes

Functions

binary_compare

Compare two BinaryArrays using the given Operator.

binary_compare_scalar

Compare a BinaryArray and a scalar value using the given Operator.

boolean_compare

Compare two BooleanArrays using the given Operator.

boolean_compare_scalar

Compare a BooleanArray and a scalar value using the given Operator.

can_compare

Checks if an array of type datatype can be compared with another array of the same type.

compare

Compares each slot of lhs against each slot of rhs.

compare_scalar

Compares all slots of lhs against rhs.

primitive_compare

Compare two PrimitiveArrays using the given Operator.

primitive_compare_scalar

Compare a PrimitiveArray and a scalar value using the given Operator.

utf8_compare

Compare two Utf8Arrays using the given Operator.

utf8_compare_scalar

Compare a Utf8Array and a scalar value using the given Operator.