Module arrow2::compute::comparison::boolean

Available on crate feature compute_comparison only.

Comparison functions for BooleanArray

Functions

compare_op_scalar

Evaluate op(left, right) for BooleanArray and scalar using a specified comparison function.

eq

Perform lhs == rhs operation on two BooleanArrays.

eq_and_validity

Perform lhs == rhs operation on two BooleanArrays and include validities in comparison.

eq_scalar

Perform lhs == rhs operation on a BooleanArray and a scalar value.

eq_scalar_and_validity

Perform lhs == rhs operation on a BooleanArray and a scalar value and include validities in comparison.

gt

Perform left > right operation on two arrays. Non-null values are greater than null values.

gt_eq

Perform left >= right operation on two arrays. Non-null values are greater than null values.

gt_eq_scalar

Perform left >= right operation on an array and a scalar value. Non-null values are greater than null values.

gt_scalar

Perform left > right operation on an array and a scalar value. Non-null values are greater than null values.

lt

Perform left < right operation on two arrays.

lt_eq

Perform left <= right operation on two arrays.

lt_eq_scalar

Perform left <= right operation on an array and a scalar value. Null values are less than non-null values.

lt_scalar

Perform left < right operation on an array and a scalar value.

neq

lhs != rhs for BooleanArray

neq_and_validity

lhs != rhs for BooleanArray and include validities in comparison.

neq_scalar

Perform left != right operation on an array and a scalar value.

neq_scalar_and_validity

Perform left != right operation on an array and a scalar value.