Function arrow2::compute::comparison::lt_eq

pub fn lt_eq(lhs: &dyn Array, rhs: &dyn Array) -> BooleanArray

Available on crate feature compute_comparison only.

<= between two Arrays. Use can_lt_eq to check whether the operation is valid

Panic

Panics iff either:

the arrays do not have have the same logical type

the arrays do not have the same length

the operation is not supported for the logical type