Function arrow2::compute::comparison::neq_and_validity

pub fn neq_and_validity(lhs: &dyn Array, rhs: &dyn Array) -> BooleanArray

Available on crate feature compute_comparison only.

!= between two Arrays and includes validities in comparison. Use can_neq to check whether the operation is valid

Panic

Panics iff either:

the arrays do not have have the same logical type

the arrays do not have the same length

the operation is not supported for the logical type