Function arrow2::compute::merge_sort::merge_sort_slices

pub fn merge_sort_slices<'a, L: Iterator<Item = &'a MergeSlice>, R: Iterator<Item = &'a MergeSlice>>(
    lhs: L,
    rhs: R,
    comparator: &'a Box<dyn Fn(usize, usize, usize, usize) -> Ordering + '_>
) -> MergeSortSlices<'a, L, R> 

Available on crate feature compute_merge_sort only.

Given two iterators of slices representing two sets of sorted Arrays, and a comparator bound to those Arrays, returns a new iterator of slices denoting how to take slices from each of the arrays such that the resulting array is sorted according to comparator