Function arrow2::compute::merge_sort::take_arrays

pub fn take_arrays<I: IntoIterator<Item = MergeSlice>>(
    arrays: &[&dyn Array],
    slices: I,
    limit: Option<usize>
) -> Box<dyn Array>

Available on crate feature compute_merge_sort only.

Takes N arrays together through slices under the assumption that the slices have a total coverage of the arrays. I.e. they are such that all elements on all arrays are picked (which is the case in sorting).

Panic

This function panics if:

max(slices[i].0) >= arrays.len(), as it indicates that the slices point to an array out of bounds from arrays.

the arrays do not have the same crate::datatypes::DataType (as it makes no sense to take together from them)