Module arrow2::compute::sort::row

Available on crate feature compute_sort only.

A comparable row-oriented representation of a collection of Array.

This module is an arrow2 version of [arrow::row]:[https://docs.rs/arrow/latest/arrow/row/index.html]

As Row are normalized for sorting, they can be very efficiently compared, using memcmp under the hood, or used in non-comparison sorts such as radix sort. This makes the row format ideal for implementing efficient multi-column sorting, grouping, aggregation, windowing and more.

Comparing Rows generated by different RowConverter is not guaranteed to yield a meaningful ordering

Structs

Row

A comparable representation of a row

RowConverter

Converts Box<dyn Array> columns into a row-oriented format.

Rows

A row-oriented representation of arrow data, that is normalized for comparison

RowsIter

An iterator over Rows

SortField

Configure the data type and sort order for a given column