Struct arrow2::compute::sort::row::RowConverter

pub struct RowConverter { /* private fields */ }

Available on crate feature compute_sort only.

Converts Box<dyn Array> columns into a row-oriented format.

Format

The encoding of the row format should not be considered stable, but is documented here for reference.

Unsigned Integer Encoding

A null integer is encoded as a 0_u8, followed by a zero-ed number of bytes corresponding to the integer’s length

A valid integer is encoded as 1_u8, followed by the big-endian representation of the integer

Signed Integer Encoding

Signed integers have their most significant sign bit flipped, and are then encoded in the same manner as an unsigned integer

Float Encoding

Floats are converted from IEEE 754 representation to a signed integer representation by flipping all bar the sign bit if they are negative.

They are then encoded in the same manner as a signed integer

Variable Length Bytes Encoding

A null is encoded as a 0_u8

An empty byte array is encoded as 1_u8

A non-null, non-empty byte array is encoded as 2_u8 followed by the byte array encoded using a block based scheme described below.

The byte array is broken up into 32-byte blocks, each block is written in turn to the output, followed by 0xFF_u8. The final block is padded to 32-bytes with 0_u8 and written to the output, followed by the un-padded length in bytes of this final block as a u8

This is loosely inspired by COBS encoding, and chosen over more traditional byte stuffing as it is more amenable to vectorisation, in particular AVX-256.

Dictionary Encoding

RowConverter needs to support converting dictionary encoded arrays with unsorted, and potentially distinct dictionaries. One simple mechanism to avoid this would be to reverse the dictionary encoding, and encode the array values directly, however, this would lose the benefits of dictionary encoding to reduce memory and CPU consumption.

As such the RowConverter maintains an order-preserving dictionary encoding for each dictionary encoded column. As this is a variable-length encoding, new dictionary values can be added whilst preserving the sort order.

A null dictionary value is encoded as 0_u8.

A non-null dictionary value is encoded as 1_u8 followed by a null-terminated byte array key determined by the order-preserving dictionary encoding

Ordering

Float Ordering

Floats are totally ordered in accordance to the totalOrder predicate as defined in the IEEE 754 (2008 revision) floating point standard.

The ordering established by this does not always agree with the PartialOrd and PartialEq implementations of f32. For example, they consider negative and positive zero equal, while this does not

Null Ordering

The encoding described above will order nulls first, this can be inverted by representing nulls as 0xFF_u8 instead of 0_u8

Reverse Column Ordering

The order of a given column can be reversed by negating the encoded bytes of non-null values

Implementations

impl RowConverter

pub fn new(fields: Vec<SortField>) -> Self

Create a new RowConverter with the provided schema

pub fn convert_columns(&mut self, columns: &[Box<dyn Array>]) -> Result<Rows>

Convert a slice of Box<dyn Array> columns into Rows

See Row for information on when Row can be compared

Panics

Panics if the schema of columns does not match that provided to RowConverter::new