Function arrow2::compute::temporal::minute

pub fn minute(array: &dyn Array) -> Result<PrimitiveArray<u32>>

Available on crate feature compute_temporal only.

Extracts the minutes of a temporal array as PrimitiveArray<u32>. Value ranges from 0 to 59. Use can_minute to check if this operation is supported for the target DataType.