Enum arrow2::error::ArrowError

pub enum ArrowError {
    NotYetImplemented(String),
    External(String, Box<dyn Error + Send + Sync>),
    Schema(String),
    Io(Error),
    InvalidArgumentError(String),
    Ffi(String),
    Ipc(String),
    ExternalFormat(String),
    KeyOverflowError,
    ArithmeticError(String),
    Other(String),
}

Enum with all errors in this crate.

Variants

NotYetImplemented(String)

Returned when functionality is not yet available.

Tuple Fields of NotYetImplemented

0: String

External(String, Box<dyn Error + Send + Sync>)

Triggered by an external error, such as CSV, serde, chrono.

Tuple Fields of External

0: String1: Box<dyn Error + Send + Sync>

Schema(String)

Error associated with incompatible schemas.

Tuple Fields of Schema

0: String

Io(Error)

Errors associated with IO

Tuple Fields of Io

0: Error

InvalidArgumentError(String)

When an invalid argument is passed to a function.

Tuple Fields of InvalidArgumentError

0: String

Ffi(String)

Error during import or export to/from C Data Interface

Tuple Fields of Ffi

0: String

Ipc(String)

Error during import or export to/from IPC

Tuple Fields of Ipc

0: String

ExternalFormat(String)

Error during import or export to/from a format

Tuple Fields of ExternalFormat

0: String

KeyOverflowError

Whenever pushing to a container fails because it does not support more entries. (e.g. maximum size of the keys of a dictionary overflowed)

ArithmeticError(String)

Error during arithmetic operation. Normally returned during checked operations

Tuple Fields of ArithmeticError

0: String

Other(String)

Any other error.

Tuple Fields of Other

0: String

Implementations

impl ArrowError

pub fn from_external_error(error: impl Error + Send + Sync + 'static) -> Self

Wraps an external error in an ArrowError.