Function arrow2::ffi::import_array_from_c

pub unsafe fn import_array_from_c(
    array: ArrowArray,
    data_type: DataType
) -> Result<Box<dyn Array>>

Imports an Array from the C data interface.

Safety

This function is intrinsically unsafe and relies on a ArrowArray being valid according to the C data interface (FFI).