Trait arrow2::ffi::ArrowArrayRef

pub trait ArrowArrayRef {
    fn parent(&self) -> &Arc<ArrowArray>;

    fn array(&self) -> &Ffi_ArrowArray;

    fn schema(&self) -> &Ffi_ArrowSchema;

    fn data_type(&self) -> Result<DataType>;

    fn deallocation(&self) -> Deallocation { ... }

    unsafe fn validity(&self) -> Result<Option<Bitmap>> { ... }

    unsafe fn buffer<T: NativeType>(&self, index: usize) -> Result<Buffer<T>> { ... }

    unsafe fn bitmap(&self, index: usize) -> Result<Bitmap> { ... }

    fn child(&self, index: usize) -> Result<ArrowArrayChild<'_>> { ... }
}

Required methods

fn parent(&self) -> &Arc<ArrowArray>

fn array(&self) -> &Ffi_ArrowArray

fn schema(&self) -> &Ffi_ArrowSchema

fn data_type(&self) -> Result<DataType>

Loading content...
Provided methods

fn deallocation(&self) -> Deallocation

unsafe fn validity(&self) -> Result<Option<Bitmap>>

returns the null bit buffer. Rust implementation uses a buffer that is not part of the array of buffers. The C Data interface’s null buffer is part of the array of buffers.

Safety

The caller must guarantee that the buffer index corresponds to a bitmap. This function assumes that the bitmap created from FFI is valid; this is impossible to prove.

unsafe fn buffer<T: NativeType>(&self, index: usize) -> Result<Buffer<T>>

Safety

The caller must guarantee that the buffer index corresponds to a bitmap. This function assumes that the bitmap created from FFI is valid; this is impossible to prove.

unsafe fn bitmap(&self, index: usize) -> Result<Bitmap>

Safety

The caller must guarantee that the buffer index corresponds to a bitmap. This function assumes that the bitmap created from FFI is valid; this is impossible to prove.

fn child(&self, index: usize) -> Result<ArrowArrayChild<'_>>

Loading content...