Module arrow2::io::avro::write

Available on crate feature io_avro only.

APIs to write to Avro format.

Functions

can_serialize

Whether new_serializer supports data_type.

new_serializer

Creates a StreamingIterator trait object that presents items from array encoded according to schema.

serialize

consumes a set of BoxSerializer into an [Block].

to_record

Converts a Schema to an Avro [Record].

Type Aliases

BoxSerializer

A type alias for a boxed StreamingIterator, used to write arrays into avro rows (i.e. a column -> row transposition of types known at run-time)