Function arrow2::io::avro::write::serialize

pub fn serialize<'a>(serializers: &mut [BoxSerializer<'a>], block: &mut Block)

Available on crate feature io_avro only.

consumes a set of BoxSerializer into an Block.

Panics

Panics iff the number of items in any of the serializers is not equal to the number of rows declared in the block.