Module arrow2::io::csv::read_async

Available on (crate features io_csv_read or io_csv_read_async or io_csv_write) and crate feature io_csv_read_async only.

Asynchronous reading of CSV

Structs

AsyncReader

A already configured CSV reader.

AsyncReaderBuilder

Builds a CSV reader with various configuration knobs.

ByteRecord

A single CSV record stored as raw bytes.

CSVError

An error that can occur when processing CSV data.

Functions

deserialize_batch

Deserializes rows ByteRecord into Chunk. Note that this is a convenience function: column deserialization is trivially parallelizable (e.g. rayon).

deserialize_column

Deserializes column of rows into an Array of DataType datatype.

infer

Infers DataType from bytes

infer_schema

Infers the Fields of a CSV file by reading through the first n records up to max_rows. Seeks back to the begining of the file after the header

read_rows

Asynchronosly read len rows from reader into row, skiping the first skip. This operation has minimal CPU work and is thus the fastest way to read through a CSV without deserializing the contents to Arrow.