Module arrow2::io::csv::write

Structs

ByteRecord
A single CSV record stored as raw bytes.

SerializeOptions

Writer
A already configured CSV writer.

WriterBuilder
Builds a CSV writer with various configuration knobs.

Functions

new_serializer
Returns an Iterator that returns items of Array as Vec<u8>, according to options. For numeric types, this serializes as usual. For dates, times and timestamps, it uses options to Supported types:

serialize
Serializes a RecordBatch as vector of ByteRecord. The vector is guaranteed to have batch.num_rows() entries. Each ByteRecord is guaranteed to have batch.num_columns() fields.

write_batch
Writes the data in a RecordBatch to writer according to the serialization options options.

write_header
Writes a header to writer according to schema