Module arrow2::io::flight

Available on crate feature io_flight only.

Serialization and deserialization to Arrow’s flight protocol

Re-exports

pub use super::ipc::write::default_ipc_fields;

Structs

WriteOptionsio_ipc

Options declaring the behaviour of writing to IPC

Functions

deserialize_batch

Deserializes [FlightData] representing a record batch message to Chunk.

deserialize_dictionary

Deserializes [FlightData], assuming it to be a dictionary message, into dictionaries.

deserialize_message

Deserializes [FlightData] into either a Chunk (when the message is a record batch) or by upserting into dictionaries (when the message is a dictionary)

deserialize_schemas

Deserialize an IPC message into Schema, IpcSchema. Use to deserialize [FlightData::data_header] and [SchemaResult::schema].

serialize_batch

Serializes Chunk to a vector of [FlightData] representing the serialized dictionaries and a [FlightData] representing the batch.

serialize_schema

Serializes a Schema to [FlightData].

serialize_schema_to_info

Convert a Schema to bytes in the format expected in [arrow_format::flight::data::FlightInfo].

serialize_schema_to_result

Serializes a Schema to [SchemaResult].