Module arrow2::io::ipc::gen::File

Structs

Block

Footer

FooterArgs

FooterBuilder

Enums

FooterOffset

Functions

finish_footer_buffer

finish_size_prefixed_footer_buffer

get_root_as_footer	Deprecated

get_size_prefixed_root_as_footer	Deprecated

root_as_footer
Verifies that a buffer of bytes contains a Footer and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_footer_unchecked.

root_as_footer_unchecked⚠
Assumes, without verification, that a buffer of bytes contains a Footer and returns it.

root_as_footer_with_opts
Verifies, with the given options, that a buffer of bytes contains a Footer and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_footer_unchecked.

size_prefixed_root_as_footer
Verifies that a buffer of bytes contains a size prefixed Footer and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use size_prefixed_root_as_footer_unchecked.

size_prefixed_root_as_footer_unchecked⚠
Assumes, without verification, that a buffer of bytes contains a size prefixed Footer and returns it.

size_prefixed_root_as_footer_with_opts
Verifies, with the given verifier options, that a buffer of bytes contains a size prefixed Footer and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_footer_unchecked.