Module arrow2::io::ipc::gen::Tensor

This is supported on crate feature io_ipc only.

Structs

Tensor

TensorArgs

TensorBuilder

TensorDim

TensorDimArgs

TensorDimBuilder

Enums

TensorDimOffset

TensorOffset

Functions

finish_size_prefixed_tensor_buffer

finish_tensor_buffer

get_root_as_tensorDeprecated

get_size_prefixed_root_as_tensorDeprecated

root_as_tensor

Verifies that a buffer of bytes contains a Tensor and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_tensor_unchecked.

root_as_tensor_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a Tensor and returns it.

root_as_tensor_with_opts

Verifies, with the given options, that a buffer of bytes contains a Tensor and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_tensor_unchecked.

size_prefixed_root_as_tensor

Verifies that a buffer of bytes contains a size prefixed Tensor and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use size_prefixed_root_as_tensor_unchecked.

size_prefixed_root_as_tensor_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a size prefixed Tensor and returns it.

size_prefixed_root_as_tensor_with_opts

Verifies, with the given verifier options, that a buffer of bytes contains a size prefixed Tensor and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_tensor_unchecked.