Enum arrow2::io::ipc::read::OutOfSpecKind

#[non_exhaustive]
pub enum OutOfSpecKind {

    InvalidHeader,
    InvalidFooter,
    NegativeFooterLength,
    InvalidFlatbufferFooter(Error),
    MissingRecordBatches,
    InvalidFlatbufferRecordBatches(Error),
    MissingSchema,
    InvalidFlatbufferSchema(Error),
    MissingFields,
    InvalidFlatbufferDictionaries(Error),
    InvalidFlatbufferBlock(Error),
    InvalidFlatbufferMessage(Error),
    MissingMessageHeader,
    InvalidFlatbufferHeader(Error),
    UnexpectedNegativeInteger,
    UnexpectedMessageType,
    MissingMessageBuffers,
    InvalidFlatbufferBuffers(Error),
    MissingMessageNodes,
    InvalidFlatbufferNodes(Error),
    InvalidFlatbufferBodyLength(Error),
    MissingData,
    InvalidFlatbufferData(Error),
    InvalidFlatbufferVersion(Error),
    InvalidFlatbufferCompression(Error),
    ExpectedBuffer,
    InvalidBuffer {
        length: usize,
        type_name: &'static str,
        required_number_of_bytes: usize,
        buffer_length: usize,
    },
    InvalidBuffersLength {
        buffers_size: u64,
        file_size: u64,
    },
    InvalidBitmap {
        length: usize,
        number_of_bits: usize,
    },
    InvalidFlatbufferIsDelta(Error),
    InvalidFlatbufferId(Error),
    InvalidId {
        requested_id: i64,
    },
    InvalidIdDataType {
        requested_id: i64,
    },

}

Available on crate feature io_ipc only.

The different types of errors that reading from IPC can cause

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

InvalidHeader

The IPC file does not start with [b’A’, b’R’, b’R’, b’O’, b’W’, b’1’]

InvalidFooter

The IPC file does not end with [b’A’, b’R’, b’R’, b’O’, b’W’, b’1’]

NegativeFooterLength

The first 4 bytes of the last 10 bytes is < 0

InvalidFlatbufferFooter(Error)

The footer is an invalid flatbuffer

MissingRecordBatches

The file’s footer does not contain record batches

InvalidFlatbufferRecordBatches(Error)

The footer’s record batches is an invalid flatbuffer

MissingSchema

The file’s footer does not contain a schema

InvalidFlatbufferSchema(Error)

The footer’s schema is an invalid flatbuffer

MissingFields

The file’s schema does not contain fields

InvalidFlatbufferDictionaries(Error)

The footer’s dictionaries is an invalid flatbuffer

InvalidFlatbufferBlock(Error)

The block is an invalid flatbuffer

InvalidFlatbufferMessage(Error)

The dictionary message is an invalid flatbuffer

MissingMessageHeader

The message does not contain a header

InvalidFlatbufferHeader(Error)

The message’s header is an invalid flatbuffer

UnexpectedNegativeInteger

Relative positions in the file is < 0

UnexpectedMessageType

dictionaries can only contain dictionary messages; record batches can only contain records

MissingMessageBuffers

RecordBatch messages do not contain buffers

InvalidFlatbufferBuffers(Error)

The message’s buffers is an invalid flatbuffer

MissingMessageNodes

RecordBatch messages does not contain nodes

InvalidFlatbufferNodes(Error)

The message’s nodes is an invalid flatbuffer

InvalidFlatbufferBodyLength(Error)

The message’s body length is an invalid flatbuffer

MissingData

The message does not contain data

InvalidFlatbufferData(Error)

The message’s data is an invalid flatbuffer

InvalidFlatbufferVersion(Error)

The version is an invalid flatbuffer

InvalidFlatbufferCompression(Error)

The compression is an invalid flatbuffer

ExpectedBuffer

The record contains a number of buffers that does not match the required number by the data type

InvalidBuffer

Fields

length: usize

Declared number of elements in the buffer

type_name: &'static str

The name of the NativeType

required_number_of_bytes: usize

Bytes required for the length and type

buffer_length: usize

The size of the IPC buffer

A buffer’s size is smaller than the required for the number of elements

InvalidBuffersLength

Fields

buffers_size: u64

number of bytes of all buffers in the record

file_size: u64

the size of the file

A buffer’s size is larger than the file size

InvalidBitmap

Fields

length: usize

Declared length of the bitmap

number_of_bits: usize

Number of bits on the IPC buffer

A bitmap’s size is smaller than the required for the number of elements

InvalidFlatbufferIsDelta(Error)

The dictionary is_delta is an invalid flatbuffer

InvalidFlatbufferId(Error)

The dictionary id is an invalid flatbuffer

InvalidId

Fields

requested_id: i64

The requested dictionary id

Invalid dictionary id

InvalidIdDataType

Fields

requested_id: i64

The requested dictionary id

Field id is not a dictionary