Function arrow2::io::ipc::read::read_record_batch

pub fn read_record_batch<R: Read + Seek>(
    batch: RecordBatchRef<'_>, 
    fields: &[Field], 
    ipc_schema: &IpcSchema, 
    projection: Option<&[usize]>, 
    dictionaries: &Dictionaries, 
    version: MetadataVersion, 
    reader: &mut R, 
    block_offset: u64
) -> Result<Chunk<Arc<dyn Array>>>

Available on crate feature io_ipc only.

Creates a record batch from binary data using the ipc::RecordBatch indexes and the Schema

Panic

Panics iff the projection is not in increasing order (e.g. [1, 0] nor [0, 1, 1] are valid)