Struct arrow2::io::ipc::write::file_async::FileSink

pub struct FileSink<'a, W: AsyncWrite + Unpin + Send + 'a> { /* private fields */ }

Available on crate features io_ipc and io_ipc_write_async only.

Sink that writes array chunks as an IPC file.

The file header is automatically written before writing the first chunk, and the file footer is automatically written when the sink is closed.

Examples

use futures::{SinkExt, TryStreamExt, io::Cursor};
use arrow2::array::{Array, Int32Array};
use arrow2::datatypes::{DataType, Field, Schema};
use arrow2::chunk::Chunk;
use arrow2::io::ipc::write::file_async::FileSink;
use arrow2::io::ipc::read::file_async::{read_file_metadata_async, FileStream};
let schema = Schema::from(vec![
    Field::new("values", DataType::Int32, true),
]);

let mut buffer = Cursor::new(vec![]);
let mut sink = FileSink::new(
    &mut buffer,
    schema,
    None,
    Default::default(),
);

// Write chunks to file
for i in 0..3 {
    let values = Int32Array::from(&[Some(i), None]);
    let chunk = Chunk::new(vec![values.boxed()]);
    sink.feed(chunk.into()).await?;
}
sink.close().await?;
drop(sink);

// Read chunks from file
buffer.set_position(0);
let metadata = read_file_metadata_async(&mut buffer).await?;
let mut stream = FileStream::new(buffer, metadata, None, None);
let chunks = stream.try_collect::<Vec<_>>().await?;

Implementations

impl<'a, W> FileSink<'a, W>where W: AsyncWrite + Unpin + Send + 'a,

pub fn new( writer: W, schema: Schema, ipc_fields: Option<Vec<IpcField>>, options: WriteOptions ) -> Self

Create a new file writer.