Module arrow2::io::json

Convert data between the Arrow memory format and JSON line-delimited records.

Structs

JsonArray
Produces JSON output as a single JSON array. For example

LineDelimited
Produces JSON output with one record per line. For example

Reader
JSON Reader

ReaderBuilder
JSON file reader builder

Writer
A JSON writer which serializes RecordBatches to a stream of u8 encoded JSON objects. See the module level documentation for detailed usage and examples. The specific format of the stream is controlled by the JsonFormat type parameter.

Traits

JsonFormat
This trait defines how to format a sequence of JSON objects to a byte stream.

ToJson

Functions

infer_json_schema
Infer the fields of a JSON file by reading the first n records of the buffer, with max_read_records controlling the maximum number of records to read.

infer_json_schema_from_iterator
Infer the fields of a JSON file by reading all items from the JSON Value Iterator.

infer_json_schema_from_seekable
Infer the fields of a JSON file by reading the first n records of the file, with max_read_records controlling the maximum number of records to read.

write_record_batches
Serializes a RecordBatch into Json

Type Definitions

ArrayWriter
A JSON writer which serializes RecordBatches to JSON arrays

LineDelimitedWriter
A JSON writer which serializes RecordBatches to newline delimited JSON objects