Module arrow2::io::json::write

Available on crate feature io_json only.

APIs to write to JSON

Structs

Convert

An iterator which wraps a normal Iterator.

Empty

An iterator over no items.

Filter

An iterator which filters elements with a predicate.

Fuse

An iterator which is well-behaved at the beginning and end of iteration.

Map

An iterator which applies a transform to elements.

MapErr

An iterator which applies a transform to errors.

MapRef

An iterator which applies a transform to elements.

RecordSerializer

FallibleStreamingIterator that serializes a Chunk into bytes of JSON in a (pandas-compatible) record-oriented format.

Serializer

FallibleStreamingIterator that serializes an Array to bytes of valid JSON

Skip

An iterator which skips a number of initial elements.

SkipWhile

An iterator which skips initial elements matching a predicate.

Take

An iterator which only returns a number of initial elements.

TakeWhile

An iterator which only returns initial elements matching a predicate.

Traits

DoubleEndedFallibleStreamingIterator

A fallible, streaming iterator which can be advanced from either end.

FallibleStreamingIterator

A fallible, streaming iterator.

Functions

convert

Converts a normal Iterator over Results of references into a FallibleStreamingIterator.

empty

Returns an iterator over no items.

write

Writes valid JSON from an iterator of (assumed JSON-encoded) bytes to writer