Module arrow2::io::json_integration

Available on crate feature io_json_integration only.

Utils for JSON integration testing

These utilities define structs that read the integration JSON format for integration testing purposes.

Modules

read

API to read from Arrow JSON integration format

write

API to write to Arrow JSON integration format

Structs

ArrowJson

A struct that represents an Arrow file with a schema and record batches

ArrowJsonBatch

A struct that partially reads the Arrow JSON record batch

ArrowJsonColumn

A struct that partially reads the Arrow JSON column/array

ArrowJsonDictionaryBatch

A struct that partially reads the Arrow JSON dictionary batch

ArrowJsonField

Fields are left as JSON Value as they vary by DataType

ArrowJsonFieldDictionary

Dictionary metadata

ArrowJsonSchema

A struct that partially reads the Arrow JSON schema.

IntegerType

the type of the integer in the dictionary