Module arrow2::io::ndjson::read

Available on crate feature io_json only.

APIs to read and deserialize NDJSON.

Structs

FileReader

A FallibleStreamingIterator of NDJSON rows.

Traits

FallibleStreamingIterator

A fallible, streaming iterator.

Functions

deserialize

Deserializes rows into an Array of DataType.

deserialize_iter

Deserializes an iterator of rows into an Array of DataType.

infer

Infers the DataType from an NDJSON file, optionally only using number_of_rows rows.

infer_iter

Infers the DataType from an iterator of JSON strings. A limited number of rows can be used by passing rows.take(number_of_rows) as an input.