Function arrow2::io::ndjson::read::deserialize

pub fn deserialize(
    rows: &[String], 
    data_type: DataType
) -> Result<Arc<dyn Array>, ArrowError>

Available on crate feature io_json only.

Deserializes rows into an Array of DataType.

Implementation

This function is CPU-bounded. This function is guaranteed to return an array of length equal to rows.len().

Errors

This function errors iff any of the rows is not a valid JSON (i.e. the format is not valid NDJSON).