Module arrow2::io::parquet::read

Re-exports

pub use schema::get_schema;

pub use schema::is_type_nullable;

Modules

schema

Structs

ColumnDescriptor
A descriptor for leaf-level primitive columns. This encapsulates information such as definition and repetition levels and is used to re-assemble nested data.

TimestampType
Timestamp logical type annotation

Enums

CompressedPage
A CompressedPage is compressed, encoded representation of a Parquet page. It holds actual data and thus cloning it is expensive. Favor passing this enum by value, as it deallocates it as soon as it is not needed, thereby reducing memory usage.

LogicalType

ParquetError

ParquetTimeUnit

ParquetType
Representation of a Parquet type. Used to describe primitive leaf fields and structs, including top-level schema. Note that the top-level schema type is represented using GroupType whose repetition is None.

PhysicalType

PrimitiveConvertedType

Functions

get_page_iterator

page_iter_to_array

read_metadata
Reads a file’s metadata.