Enum arrow2::io::parquet::read::ParquetError

pub enum ParquetError {
    OutOfSpec(String),
    FeatureNotActive(Feature, String),
    FeatureNotSupported(String),
    InvalidParameter(String),
    WouldOverAllocate,
}

Available on crate feature io_parquet only.

Errors generated by this crate

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

OutOfSpec(String)

When the parquet file is known to be out of spec.

FeatureNotActive(Feature, String)

Error presented when trying to use a code branch that requires activating a feature.

FeatureNotSupported(String)

Error presented when trying to use a feature from parquet that is not yet supported

InvalidParameter(String)

When encoding, the user passed an invalid parameter

WouldOverAllocate

When decoding or decompressing, the page would allocate more memory than allowed