Enum arrow2::io::parquet::read::PrimitiveConvertedType

pub enum PrimitiveConvertedType {

    Utf8,
    Enum,
    Decimal(i32, i32),
    Date,
    TimeMillis,
    TimeMicros,
    TimestampMillis,
    TimestampMicros,
    Uint8,
    Uint16,
    Uint32,
    Uint64,
    Int8,
    Int16,
    Int32,
    Int64,
    Json,
    Bson,
    Interval,

}

This is supported on crate feature io_parquet only.

Variants

Utf8

Enum

an enum is converted into a binary field

Decimal(i32, i32)

A decimal value.

This may be used to annotate binary or fixed primitive types. The underlying byte array stores the unscaled value encoded as two’s complement using big-endian byte order (the most significant byte is the zeroth element). The value of the decimal is the value * 10^{-scale}.

This must be accompanied by a (maximum) precision and a scale in the SchemaElement. The precision specifies the number of digits in the decimal and the scale stores the location of the decimal point. For example 1.23 would have precision 3 (3 total digits) and scale 2 (the decimal point is 2 digits over).

Date

A Date

Stored as days since Unix epoch, encoded as the INT32 physical type.

TimeMillis

A time

The total number of milliseconds since midnight. The value is stored as an INT32 physical type.

TimeMicros

A time.

The total number of microseconds since midnight. The value is stored as an INT64 physical type.

TimestampMillis

A date/time combination

Date and time recorded as milliseconds since the Unix epoch. Recorded as a physical type of INT64.

TimestampMicros

A date/time combination

Date and time recorded as microseconds since the Unix epoch. The value is stored as an INT64 physical type.

Uint8

An unsigned integer value.

The number describes the maximum number of meainful data bits in the stored value. 8, 16 and 32 bit values are stored using the INT32 physical type. 64 bit values are stored using the INT64 physical type.

Uint16

Uint32

Uint64

Int8

A signed integer value.

The number describes the maximum number of meainful data bits in the stored value. 8, 16 and 32 bit values are stored using the INT32 physical type. 64 bit values are stored using the INT64 physical type.

Int16

Int32

Int64

Json

An embedded JSON document

A JSON document embedded within a single UTF8 column.

Bson

An embedded BSON document

A BSON document embedded within a single BINARY column.

Interval

An interval of time

This type annotates data stored as a FIXED_LEN_BYTE_ARRAY of length 12 This data is composed of three separate little endian unsigned integers. Each stores a component of a duration of time. The first integer identifies the number of months associated with the duration, the second identifies the number of days associated with the duration and the third identifies the number of milliseconds associated with the provided duration. This duration of time is independent of any particular timezone or date.