Function arrow2::io::parquet::read::get_column_iterator

pub fn get_column_iterator<R>(
    reader: R, 
    metadata: &FileMetaData, 
    row_group: usize, 
    field: usize, 
    page_filter: Option<Arc<dyn Fn(&Descriptor, &DataPageHeader) + Sync + Send + 'static>>, 
    page_buffer: Vec<u8, Global>
) -> ColumnIterator<R> where
    R: Read + Seek, 

Available on crate feature io_parquet only.

Returns a [ColumnIterator] of column chunks corresponding to field. Contrarily to get_page_iterator that returns a single iterator of pages, this iterator returns multiple iterators, one per physical column of the field. For primitive fields (e.g. i64), [ColumnIterator] yields exactly one column. For complex fields, it yields multiple columns.