Function arrow2::io::parquet::read::read_columns_many

pub fn read_columns_many<'a, R: Read + Seek>(
    reader: &mut R,
    row_group: &RowGroupMetaData,
    fields: Vec<Field>,
    chunk_size: Option<usize>,
    limit: Option<usize>,
    pages: Option<Vec<Vec<Vec<FilteredPage>>>>
) -> Result<Vec<ArrayIter<'a>>>

Available on crate feature io_parquet only.

Returns a vector of iterators of Array (ArrayIter) corresponding to the top level parquet fields whose name matches fields’s names.

Implementation

This operation is IO-bounded O(C) where C is the number of columns in the row group - it reads all the columns to memory from the row group associated to the requested fields.

This operation is single-threaded. For readers with stronger invariants (e.g. implement Clone) you can use read_columns to read multiple columns at once and convert them to ArrayIter via to_deserializer.