Function arrow2::io::parquet::read::read_columns_many_async

pub async fn read_columns_many_async<'a, 'b, R: AsyncRead + AsyncSeek + Send + Unpin, F: Fn() -> BoxFuture<'b, Result<R>> + Clone>(
    factory: F, 
    row_group: &RowGroupMetaData, 
    fields: Vec<Field>, 
    chunk_size: Option<usize>
) -> Result<Vec<ArrayIter<'a>>>

Available on crate feature io_parquet only.

Returns a vector of iterators of Array corresponding to the top level parquet fields whose name matches fields’s names.

This operation is IO-bounded O(C) where C is the number of columns in the row group - it reads all the columns to memory from the row group associated to the requested fields.

Implementation

This operation is IO-bounded O(C) where C is the number of columns in the row group - it reads all the columns to memory from the row group associated to the requested fields. It does so asynchronously via join_all