Module arrow2::io::parquet::read::indexes

Available on crate feature io_parquet only.

API to perform page-level filtering (also known as indexes)

Structs

ColumnPageStatistics

ColumnPageStatistics contains the minimum, maximum, and null_count of each page of a parquet column, as an Array. This struct has the following invariants:

FilteredPage

An enum describing a page that was either selected in a filter pushdown or skipped

Interval

An interval

Enums

FieldPageStatistics

Page statistics of an Arrow field.

Functions

has_indexes

Checks whether the row group have page index information (page statistics)

read_filtered_pages

Reads all page locations and index locations (IO-bounded) and uses predicate to compute the set of FilteredPage that fulfill the predicate.