Module arrow2::io::parquet::read::schema

Available on crate feature io_parquet only.

APIs to handle Parquet <-> Arrow schemas.

Structs

FileMetaData

Metadata for a Parquet file.

KeyValue

Wrapper struct to store key values

SchemaDescriptor

A schema descriptor. This encapsulates the top-level schemas for all the columns, as well as all descriptors for all the primitive columns.

SchemaInferenceOptions

Options when inferring schemas from Parquet

Enums

ParquetType

Representation of a Parquet type describing primitive and nested fields, including the top-level schema of the parquet file.

Functions

infer_schema

Infers a Schema from parquet’s FileMetaData. This first looks for the metadata key "ARROW:schema"; if it does not exist, it converts the parquet types declared in the file’s parquet schema to Arrow’s equivalent.

infer_schema_with_options

Like infer_schema but with configurable options which affects the behavior of inference

parquet_to_arrow_schema

Converts ParquetTypes to a Field, ignoring parquet fields that do not contain any physical column.

parquet_to_arrow_schema_with_options

Like parquet_to_arrow_schema but with configurable options which affect the behavior of schema inference

read_schema_from_metadata

Reads an arrow schema from Parquet’s file metadata. Returns None if no schema was found.