Struct arrow2::io::parquet::read::ColumnChunkMetaData

pub struct ColumnChunkMetaData { /* private fields */ }

Available on crate feature io_parquet only.

Metadata for a column chunk.

Implementations

impl ColumnChunkMetaData

pub fn new(
column_chunk: ColumnChunk,
column_descr: ColumnDescriptor
) -> ColumnChunkMetaData

Returns a new ColumnChunkMetaData

pub fn file_path(&self) -> &Option<String>

File where the column chunk is stored.

If not set, assumed to belong to the same file as the metadata. This path is relative to the current file.

pub fn file_offset(&self) -> i64

Byte offset in file_path().

pub fn column_chunk(&self) -> &ColumnChunk

Returns this column’s [ColumnChunk]

pub fn metadata(&self) -> &ColumnMetaData

The column’s [ColumnMetaData]

pub fn descriptor(&self) -> &ColumnDescriptor

The ColumnDescriptor for this column. This descriptor contains the physical and logical type of the pages.

pub fn physical_type(&self) -> PhysicalType

The PhysicalType of this column.

pub fn statistics(&self) -> Option<Result<Arc<dyn Statistics + 'static>, Error>>

Decodes the raw statistics into [Statistics].

pub fn num_values(&self) -> i64

Total number of values in this column chunk. Note that this is not necessarily the number of rows. E.g. the (nested) array [[1, 2], [3]] has 2 rows and 3 values.

pub fn compression(&self) -> Compression

[Compression] for this column.

pub fn compressed_size(&self) -> i64

Returns the total compressed data size of this column chunk.

pub fn uncompressed_size(&self) -> i64

Returns the total uncompressed data size of this column chunk.

pub fn data_page_offset(&self) -> i64

Returns the offset for the column data.

pub fn has_index_page(&self) -> bool

Returns true if this column chunk contains a index page, false otherwise.

pub fn index_page_offset(&self) -> Option<i64>

Returns the offset for the index page.

pub fn dictionary_page_offset(&self) -> Option<i64>

Returns the offset for the dictionary page, if any.

pub fn column_encoding(&self) -> &Vec<Encoding, Global>

Returns the encoding for this column

pub fn byte_range(&self) -> (u64, u64)

Returns the offset and length in bytes of the column chunk within the file

pub fn into_thrift(self) -> ColumnChunk

Method to convert to Thrift.