Module arrow2::io::parquet::write

Modules

stream

Structs

CompressedPage
A CompressedPage is compressed, encoded representation of a Parquet page. It holds actual data and thus cloning it is expensive. Favor passing this enum by value, as it deallocates it as soon as it is not needed, thereby reducing memory usage.

DynIter

RowGroupIterator
An iterator adapter that converts an iterator over RecordBatch into an iterator of row groups. Use it to create an iterator consumable by the parquet’s API.

WriteOptions

Enums

CompressionCodec
Supported compression algorithms.

ParquetType
Representation of a Parquet type. Used to describe primitive leaf fields and structs, including top-level schema. Note that the top-level schema type is represented using GroupType whose repetition is None.

Version

Functions

array_to_page

to_parquet_schema
Creates a parquet SchemaDescriptor from a Schema.

to_parquet_type

write_file
Writes

Type Definitions

RowGroupIter