Enum arrow2::io::parquet::write::ParquetType

pub enum ParquetType {
    PrimitiveType {
        basic_info: BasicTypeInfo,
        logical_type: Option<LogicalType>,
        converted_type: Option<PrimitiveConvertedType>,
        physical_type: PhysicalType,
    },
    GroupType {
        basic_info: BasicTypeInfo,
        logical_type: Option<LogicalType>,
        converted_type: Option<GroupConvertedType>,
        fields: Vec<ParquetType, Global>,
    },
}

Representation of a Parquet type. Used to describe primitive leaf fields and structs, including top-level schema. Note that the top-level schema type is represented using GroupType whose repetition is None.

Variants

PrimitiveType

Fields of PrimitiveType

basic_info: BasicTypeInfological_type: Option<LogicalType>converted_type: Option<PrimitiveConvertedType>physical_type: PhysicalType

GroupType

Fields of GroupType

basic_info: BasicTypeInfological_type: Option<LogicalType>converted_type: Option<GroupConvertedType>fields: Vec<ParquetType, Global>

Implementations

impl ParquetType

pub fn try_from_thrift(
elements: &[&SchemaElement]
) -> Result<ParquetType, ParquetError>

Method to convert from Thrift.

impl ParquetType

pub fn to_thrift(&self) -> Result<Vec<SchemaElement, Global>, ParquetError>

Method to convert to Thrift.

impl ParquetType

Accessors

pub fn get_basic_info(&self) -> &BasicTypeInfo

Returns [BasicTypeInfo] information about the type.

pub fn name(&self) -> &str

Returns this type’s field name.

pub fn is_root(&self) -> bool

pub fn check_contains(&self, sub_type: &ParquetType) -> bool

Checks if sub_type schema is part of current schema. This method can be used to check if projected columns are part of the root schema.

impl ParquetType

Constructors

pub fn new_root(name: String, fields: Vec<ParquetType, Global>) -> ParquetType

pub fn from_converted(
name: String,
fields: Vec<ParquetType, Global>,
repetition: Option<FieldRepetitionType>,
converted_type: Option<GroupConvertedType>,
id: Option<i32>
) -> ParquetType

pub fn try_from_primitive(
name: String,
physical_type: PhysicalType,
repetition: FieldRepetitionType,
converted_type: Option<PrimitiveConvertedType>,
logical_type: Option<LogicalType>,
id: Option<i32>
) -> Result<ParquetType, ParquetError>

pub fn from_physical(name: String, physical_type: PhysicalType) -> ParquetType

pub fn try_from_group(
name: String,
repetition: FieldRepetitionType,
converted_type: Option<GroupConvertedType>,
logical_type: Option<LogicalType>,
fields: Vec<ParquetType, Global>,
id: Option<i32>
) -> Result<ParquetType, ParquetError>