Struct arrow2::offset::Offsets

pub struct Offsets<O: Offset>(_);

A wrapper type of Vec<O> representing the invariants of Arrow’s offsets. It is guaranteed to (sound to assume that):

every element is >= 0

element at position i is >= than element at position i-1.

Implementations

impl<O: Offset> Offsets<O>

pub fn new() -> Self

Returns an empty Offsets (i.e. with a single element, the zero)

pub fn new_zeroed(length: usize) -> Self

Returns an Offsets whose all lengths are zero.

pub fn try_from_iter<I: IntoIterator<Item = usize>>( iter: I ) -> Result<Self, Error>

Creates a new Offsets from an iterator of lengths

pub fn with_capacity(capacity: usize) -> Self

Returns a new Offsets with a capacity, allocating at least capacity + 1 entries.

pub fn capacity(&self) -> usize

Returns the capacity of Offsets.

pub fn reserve(&mut self, additional: usize)

Reserves additional entries.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of self to fit.

pub fn try_push(&mut self, length: O) -> Result<(), Error>

Pushes a new element with a given length.

Error

This function errors iff the new last item is larger than what O supports.

Panic

This function asserts that length > 0.

pub fn try_push_usize(&mut self, length: usize) -> Result<(), Error>

Pushes a new element with a given length.

Error

This function errors iff the new last item is larger than what O supports.

Implementation

This function:

checks that this length does not overflow

pub unsafe fn new_unchecked(offsets: Vec<O>) -> Self

Returns Offsets assuming that offsets fulfills its invariants

Safety

This is safe iff the invariants of this struct are guaranteed in offsets.

pub fn last(&self) -> &O

Returns the last offset of this container.

pub fn start_end(&self, index: usize) -> (usize, usize)

Returns a range (start, end) corresponding to the position index

Panic

This function panics iff index >= self.len()

pub unsafe fn start_end_unchecked(&self, index: usize) -> (usize, usize)

Returns a range (start, end) corresponding to the position index

Safety

index must be < self.len()

pub fn len_proxy(&self) -> usize

Returns the length an array with these offsets would be.

pub fn len(&self) -> usize

Returns the number of offsets in this container.

pub fn as_slice(&self) -> &[O]

Returns the byte slice stored in this buffer

pub fn pop(&mut self) -> Option<O>

Pops the last element

pub fn extend_constant(&mut self, additional: usize)

Extends itself with additional elements equal to the last offset. This is useful to extend offsets with empty values, e.g. for null slots.

pub fn try_from_lengths<I: Iterator<Item = usize>>( lengths: I ) -> Result<Self, Error>

Try to create a new Offsets from a sequence of lengths

Errors

This function errors iff this operation overflows for the maximum value of O.

pub fn try_extend_from_lengths<I: Iterator<Item = usize>>( &mut self, lengths: I ) -> Result<(), Error>

Try extend from an iterator of lengths

Errors

This function errors iff this operation overflows for the maximum value of O.

pub fn try_extend_from_self(&mut self, other: &Self) -> Result<(), Error>

Extends itself from another Offsets

Errors

This function errors iff this operation overflows for the maximum value of O.

pub fn try_extend_from_slice( &mut self, other: &OffsetsBuffer<O>, start: usize, length: usize ) -> Result<(), Error>

Extends itself from another Offsets sliced by start, length

Errors

This function errors iff this operation overflows for the maximum value of O.

pub fn into_inner(self) -> Vec<O>

Returns the inner Vec.