Struct arrow2::record_batch::RecordBatch

pub struct RecordBatch { /* fields omitted */ }

A two-dimensional batch of column-oriented data with a defined schema.

A RecordBatch is a two-dimensional dataset of a number of contiguous arrays, each the same length. A record batch has a schema which must match its arrays’ datatypes.

Record batches are a convenient unit of work for various serialization and computation functions, possibly incremental.

Implementations

impl RecordBatch

pub fn try_new(
schema: Arc<Schema>,
columns: Vec<Arc<dyn Array>>
) -> Result<Self>

Creates a RecordBatch from a schema and columns.

Expects the following:

the vec of columns to not be empty

the schema and column data types to have equal lengths and match

each array in columns to have the same length

If the conditions are not met, an error is returned.

Example

let id_array = Primitive::from_slice(&[1i32, 2, 3, 4, 5]).to(DataType::Int32);
let schema = Arc::new(Schema::new(vec![
    Field::new("id", DataType::Int32, false)
]));

let batch = RecordBatch::try_new(
    schema,
    vec![Arc::new(id_array)]
)?;

pub fn try_new_with_options(
schema: Arc<Schema>,
columns: Vec<Arc<dyn Array>>,
options: &RecordBatchOptions
) -> Result<Self>

Creates a RecordBatch from a schema and columns, with additional options, such as whether to strictly validate field names.

See RecordBatch::try_new for the expected conditions.

pub fn new_empty(schema: Arc<Schema>) -> Self

Creates a new empty RecordBatch.

pub fn schema(&self) -> &Arc<Schema>

Returns the Schema of the record batch.

pub fn num_columns(&self) -> usize

Returns the number of columns in the record batch.

Example

let id_array = Primitive::from_slice(&[1i32, 2, 3, 4, 5]).to(DataType::Int32);
let schema = Arc::new(Schema::new(vec![
    Field::new("id", DataType::Int32, false)
]));

let batch = RecordBatch::try_new(schema, vec![Arc::new(id_array)])?;

assert_eq!(batch.num_columns(), 1);

pub fn num_rows(&self) -> usize

Returns the number of rows in each column.

Panics

Panics if the RecordBatch contains no columns.

Example

let id_array = Primitive::from_slice(&[1i32, 2, 3, 4, 5]).to(DataType::Int32);
let schema = Arc::new(Schema::new(vec![
    Field::new("id", DataType::Int32, false)
]));

let batch = RecordBatch::try_new(schema, vec![Arc::new(id_array)])?;

assert_eq!(batch.num_rows(), 5);

pub fn column(&self, index: usize) -> &Arc<dyn Array>

Get a reference to a column’s array by index.

Panics

Panics if index is outside of 0..num_columns.

pub fn columns(&self) -> &[Arc<dyn Array>]

Get a reference to all columns in the record batch.

pub fn try_from_iter<I, F>(value: I) -> Result<Self> where
I: IntoIterator<Item = (F, Arc<dyn Array>)>,
F: AsRef<str>,

Create a RecordBatch from an iterable list of pairs of the form (field_name, array), with the same requirements on fields and arrays as RecordBatch::try_new. This method is often used to create a single RecordBatch from arrays, e.g. for testing.

The resulting schema is marked as nullable for each column if the array for that column is has any nulls. To explicitly specify nullibility, use RecordBatch::try_from_iter_with_nullable

Example:

use std::sync::Arc;
use arrow2::array::*;
use arrow2::datatypes::DataType;
use arrow2::record_batch::RecordBatch;

let a: ArrayRef = Arc::new(Int32Array::from_slice(&[1, 2]));
let b: ArrayRef = Arc::new(Utf8Array::<i32>::from_slice(&["a", "b"]));

let record_batch = RecordBatch::try_from_iter(vec![
  ("a", a),
  ("b", b),
]);

pub fn try_from_iter_with_nullable<I, F>(value: I) -> Result<Self> where
I: IntoIterator<Item = (F, Arc<dyn Array>, bool)>,
F: AsRef<str>,

Create a RecordBatch from an iterable list of tuples of the form (field_name, array, nullable), with the same requirements on fields and arrays as RecordBatch::try_new. This method is often used to create a single RecordBatch from arrays, e.g. for testing.

Example:

use std::sync::Arc;
use arrow2::array::*;
use arrow2::datatypes::DataType;
use arrow2::record_batch::RecordBatch;

let a: ArrayRef = Arc::new(Int32Array::from_slice(&[1, 2]));
let b: ArrayRef = Arc::new(Utf8Array::<i32>::from_slice(&["a", "b"]));

// Note neither `a` nor `b` has any actual nulls, but we mark
// b an nullable
let record_batch = RecordBatch::try_from_iter_with_nullable(vec![
  ("a", a, false),
  ("b", b, true),
]);