Module arrow2::types::simd

Contains traits and implementations of multi-data used in SIMD. The actual representation is driven by the feature flag "simd", which, if set, uses std::simd.

Structs

days_msx8

Multi-Data correspondence of the native type

f16x32

Multi-Data correspondence of the native type

f32x16

Multi-Data correspondence of the native type

f64x8

Multi-Data correspondence of the native type

i8x64

Multi-Data correspondence of the native type

i16x32

Multi-Data correspondence of the native type

i32x16

Multi-Data correspondence of the native type

i64x8

Multi-Data correspondence of the native type

i128x8

Multi-Data correspondence of the native type

i256x8

Multi-Data correspondence of the native type

months_days_nsx8

Multi-Data correspondence of the native type

u8x64

Multi-Data correspondence of the native type

u16x32

Multi-Data correspondence of the native type

u32x16

Multi-Data correspondence of the native type

u64x8

Multi-Data correspondence of the native type

Traits

FromMaskChunk

Describes the ability to convert itself from a BitChunk.

NativeSimd

A struct that lends itself well to be compiled leveraging SIMD

Simd

Trait implemented by some NativeType that have a SIMD representation.