Trait arrow2::types::simd::NativeSimd

pub trait NativeSimd: Default {
    type Native: NativeType;
    type Chunk: BitChunk;
    type Mask: FromMaskChunk<Self::Chunk>;

    const LANES: usize;

    fn select(self, mask: Self::Mask, default: Self) -> Self;

    fn from_chunk(v: &[Self::Native]) -> Self;

    fn from_incomplete_chunk(
        v: &[Self::Native], 
        remaining: Self::Native
    ) -> Self;
}

A struct lends itself well to be compiled leveraging SIMD

Associated Types

type Native: NativeType

The NativeType of this struct. E.g. f32 for a NativeSimd = f32x16.

type Chunk: BitChunk

The type holding bits for masks.

type Mask: FromMaskChunk<Self::Chunk>

Loading content...
Associated Constants

const LANES: usize

Loading content...
Required methods

fn select(self, mask: Self::Mask, default: Self) -> Self

fn from_chunk(v: &[Self::Native]) -> Self

Convert itself from a slice.

Panics

iff v.len() != T::LANES

fn from_incomplete_chunk(v: &[Self::Native], remaining: Self::Native) -> Self

creates a new Self from remaining by populating items from v up to its length. Items from v at positions larger than the number of lanes are ignored.

Loading content...