Module arrow2::array::growable

Contains the trait Growable and corresponding concreate implementations, one per concrete array, that offer the ability to create a new Array out of slices of existing Arrays.

Structs

GrowableBinary
Concrete Growable for the BinaryArray.

GrowableBoolean
Concrete Growable for the BooleanArray.

GrowableDictionary
Concrete Growable for the DictionaryArray.

GrowableFixedSizeBinary
Concrete Growable for the FixedSizeBinaryArray.

GrowableList
Concrete Growable for the ListArray.

GrowableNull
Concrete Growable for the NullArray.

GrowablePrimitive
Concrete Growable for the PrimitiveArray.

GrowableStruct
Concrete Growable for the StructArray.

GrowableUtf8
Concrete Growable for the Utf8Array.

Traits

Growable
A trait describing a struct that can be extended from slices of pre-existing Arrays. This is used in operations where a new array is built out of other arrays such, as filtering and concatenation.

Functions

make_growable
Creates a new Growable from an arbitrary number of dynamic Arrays.