Struct arrow2::array::BooleanArray

pub struct BooleanArray { /* private fields */ }

The Arrow’s equivalent to an immutable Vec<Option<bool>>, but with 1/16 of its size. Cloning and slicing this struct is O(1).

Implementations

impl BooleanArray

pub fn from_trusted_len_values_iter<I: TrustedLen<Item = bool>>(
iterator: I
) -> Self

Creates a new BooleanArray from an TrustedLen of bool.

pub unsafe fn from_trusted_len_values_iter_unchecked<I: Iterator<Item = bool>>(
iterator: I
) -> Self

Creates a new BooleanArray from an TrustedLen of bool. Use this over BooleanArray::from_trusted_len_iter when the iterator is trusted len but this crate does not mark it as such.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_slice<P: AsRef<[bool]>>(slice: P) -> Self

Creates a new BooleanArray from a slice of bool.

pub unsafe fn from_trusted_len_iter_unchecked<I, P>(iterator: I) -> Self where
P: Borrow<bool>,
I: Iterator<Item = Option<P>>,

Creates a BooleanArray from an iterator of trusted length. Use this over BooleanArray::from_trusted_len_iter when the iterator is trusted len but this crate does not mark it as such.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_trusted_len_iter<I, P>(iterator: I) -> Self where
P: Borrow<bool>,
I: TrustedLen<Item = Option<P>>,

Creates a BooleanArray from a TrustedLen.

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I, P>(
iterator: I
) -> Result<Self, E> where
P: Borrow<bool>,
I: Iterator<Item = Result<Option<P>, E>>,

Creates a BooleanArray from an falible iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn try_from_trusted_len_iter<E, I, P>(iterator: I) -> Result<Self, E> where
P: Borrow<bool>,
I: TrustedLen<Item = Result<Option<P>, E>>,

Creates a BooleanArray from a TrustedLen.

impl<'a> BooleanArray

pub fn iter(&'a self) -> ZipValidity<'a, bool, BitmapIter<'a>>

Returns an iterator over the optional values of this BooleanArray.

pub fn values_iter(&'a self) -> BitmapIter<'a>

Returns an iterator over the values of this BooleanArray

impl BooleanArray

pub fn try_new(
data_type: DataType,
values: Bitmap,
validity: Option<Bitmap>
) -> Result<Self, ArrowError>

The canonical method to create a BooleanArray out of low-end APIs.

Errors

This function errors iff:

The validity is not None and its length is different from values’s length

The data_type’s PhysicalType is not equal to PhysicalType::Boolean.

pub fn new(
data_type: DataType,
values: Bitmap,
validity: Option<Bitmap>
) -> Self

The canonical method to create a BooleanArray

Panics

This function errors iff:

The validity is not None and its length is different from values’s length

The data_type’s PhysicalType is not equal to PhysicalType::Boolean.

pub fn from_data(
data_type: DataType,
values: Bitmap,
validity: Option<Bitmap>
) -> Self

Alias for new

pub fn new_empty(data_type: DataType) -> Self

Returns a new empty BooleanArray.

pub fn new_null(data_type: DataType, length: usize) -> Self

Returns a new BooleanArray whose all slots are null / None.

impl BooleanArray

pub fn slice(&self, offset: usize, length: usize) -> Self

Returns a slice of this BooleanArray.

Implementation

This operation is O(1) as it amounts to increase two ref counts.

Panic

This function panics iff offset + length >= self.len().

pub unsafe fn slice_unchecked(&self, offset: usize, length: usize) -> Self

Returns a slice of this BooleanArray.

Implementation

This operation is O(1) as it amounts to increase two ref counts.

Safety

The caller must ensure that offset + length <= self.len().

pub fn with_validity(&self, validity: Option<Bitmap>) -> Self

Sets the validity bitmap on this BooleanArray.

Panic

This function panics iff validity.len() != self.len().

pub fn into_mut(self) -> Either<Self, MutableBooleanArray>

Try to convert this BooleanArray to a MutableBooleanArray

impl BooleanArray

pub fn len(&self) -> usize

Returns the length of this array

pub fn value(&self, i: usize) -> bool

Returns the value at index i

Panic

This function panics iff i >= self.len().

pub unsafe fn value_unchecked(&self, i: usize) -> bool

Returns the element at index i as bool

Safety

Caller must be sure that i < self.len()

pub fn validity(&self) -> Option<&Bitmap>

The optional validity.

pub fn values(&self) -> &Bitmap

Returns the values of this BooleanArray.