Struct arrow2::array::MapArray

pub struct MapArray { /* private fields */ }

An array representing a (key, value), both of arbitrary logical types.

Implementations

impl<'a> MapArray

pub fn iter( &'a self ) -> ZipValidity<Box<dyn Array>, MapValuesIter<'a>, BitmapIter<'a>>

Returns an iterator of Option<Box<dyn Array>>

pub fn values_iter(&'a self) -> MapValuesIter<'a>

Returns an iterator of Box<dyn Array>

impl MapArray

pub fn try_new( data_type: DataType, offsets: OffsetsBuffer<i32>, field: Box<dyn Array>, validity: Option<Bitmap> ) -> Result<Self, Error>

Returns a new MapArray.

Errors

This function errors iff:

The last offset is not equal to the field’ length

The data_type’s physical type is not crate::datatypes::PhysicalType::Map

The fields’ data_type is not equal to the inner field of data_type

The validity is not None and its length is different from offsets.len() - 1.

pub fn new( data_type: DataType, offsets: OffsetsBuffer<i32>, field: Box<dyn Array>, validity: Option<Bitmap> ) -> Self

Creates a new MapArray.

Panics

The last offset is not equal to the field’ length.

The data_type’s physical type is not crate::datatypes::PhysicalType::Map,

The validity is not None and its length is different from offsets.len() - 1.

pub fn new_null(data_type: DataType, length: usize) -> Self

Returns a new null MapArray of length.

pub fn new_empty(data_type: DataType) -> Self

Returns a new empty MapArray.

impl MapArray

pub fn slice(&mut self, offset: usize, length: usize)

Returns a slice of this MapArray.

Panics

panics iff offset + length >= self.len()

pub unsafe fn slice_unchecked(&mut self, offset: usize, length: usize)

Returns a slice of this MapArray.

Safety

The caller must ensure that offset + length < self.len().

pub fn sliced(self, offset: usize, length: usize) -> Self

Returns this array sliced.

Implementation

This function is O(1).

Panics

iff offset + length > self.len().

pub fn sliced_unchecked(self, offset: usize, length: usize) -> Self

Returns this array sliced.

Implementation

This function is O(1).

Safety

The caller must ensure that offset + length <= self.len().

pub fn with_validity(self, validity: Option<Bitmap>) -> Self

Returns this array with a new validity.

Panic

Panics iff validity.len() != self.len().

pub fn set_validity(&mut self, validity: Option<Bitmap>)

Sets the validity of this array.

Panics

This function panics iff values.len() != self.len().

pub fn boxed(self) -> Box<dyn Array>

Boxes this array into a Box<dyn Array>.

pub fn arced(self) -> Arc<dyn Array>

Arcs this array into a std::sync::Arc<dyn Array>.

impl MapArray

pub fn len(&self) -> usize

Returns the length of this array

pub fn offsets(&self) -> &OffsetsBuffer<i32>

returns the offsets

pub fn field(&self) -> &Box<dyn Array>

Returns the field (guaranteed to be a Struct)

pub fn value(&self, i: usize) -> Box<dyn Array>

Returns the element at index i.

pub unsafe fn value_unchecked(&self, i: usize) -> Box<dyn Array>

Returns the element at index i.

Safety

Assumes that the i < self.len.