Struct arrow2::array::MapArray

pub struct MapArray { /* private fields */ }

An array representing a (key, value), both of arbitrary logical types.

Implementations

impl<'a> MapArray

pub fn iter(&'a self) -> ZipValidity<'a, Box<dyn Array>, MapValuesIter<'a>>

Returns an iterator of Option<Box<dyn Array>>

pub fn values_iter(&'a self) -> MapValuesIter<'a>

Returns an iterator of Box<dyn Array>

impl MapArray

pub fn try_new(
data_type: DataType,
offsets: Buffer<i32>,
field: Arc<dyn Array>,
validity: Option<Bitmap>
) -> Result<Self, ArrowError>

Returns a new MapArray.

Errors

This function errors iff:

the offsets are not monotonically increasing

The last offset is not equal to the field’ length

The data_type’s physical type is not crate::datatypes::PhysicalType::Map

The fields’ data_type is not equal to the inner field of data_type

The validity is not None and its length is different from offsets.len() - 1.

pub fn new(
data_type: DataType,
offsets: Buffer<i32>,
field: Arc<dyn Array>,
validity: Option<Bitmap>
) -> Self

Creates a new MapArray.

Panics

the offsets are not monotonically increasing

The last offset is not equal to the field’ length.

The data_type’s physical type is not crate::datatypes::PhysicalType::Map,

The validity is not None and its length is different from offsets.len() - 1.

pub fn from_data(
data_type: DataType,
offsets: Buffer<i32>,
field: Arc<dyn Array>,
validity: Option<Bitmap>
) -> Self

Alias for new

pub fn new_null(data_type: DataType, length: usize) -> Self

Returns a new null MapArray of length.

pub fn new_empty(data_type: DataType) -> Self

Returns a new empty MapArray.

impl MapArray

pub fn slice(&self, offset: usize, length: usize) -> Self

Returns a slice of this MapArray.

Panics

panics iff offset + length >= self.len()

pub unsafe fn slice_unchecked(&self, offset: usize, length: usize) -> Self

Returns a slice of this MapArray.

Safety

The caller must ensure that offset + length < self.len().

impl MapArray

pub fn len(&self) -> usize

Returns the length of this array

pub fn offsets(&self) -> &Buffer<i32>

returns the offsets

pub fn field(&self) -> &Arc<dyn Array>

Returns the field (guaranteed to be a Struct)

pub fn value(&self, i: usize) -> Box<dyn Array>

Returns the element at index i.

pub unsafe fn value_unchecked(&self, i: usize) -> Box<dyn Array>

Returns the element at index i.

Safety

Assumes that the i < self.len.