Struct arrow2::array::MutableBooleanArray

pub struct MutableBooleanArray { /* private fields */ }

The Arrow’s equivalent to Vec<Option<bool>>, but with 1/16 of its size. Converting a MutableBooleanArray into a BooleanArray is O(1).

Implementation

This struct does not allocate a validity until one is required (i.e. push a null to it).

Implementations

impl<'a> MutableBooleanArray

pub fn iter(&'a self) -> ZipValidity<'a, bool, BitmapIter<'a>>

Returns an iterator over the optional values of this MutableBooleanArray.

pub fn values_iter(&'a self) -> BitmapIter<'a>

Returns an iterator over the values of this MutableBooleanArray

impl MutableBooleanArray

pub fn new() -> Self

Creates an new empty MutableBooleanArray.

pub fn with_capacity(capacity: usize) -> Self

Creates an new MutableBooleanArray with a capacity of values.

pub fn reserve(&mut self, additional: usize)

Reserves additional slots.

pub fn from_data(
data_type: DataType,
values: MutableBitmap,
validity: Option<MutableBitmap>
) -> Self

Canonical method to create a new MutableBooleanArray.

pub fn push(&mut self, value: Option<bool>)

Pushes a new entry to MutableBooleanArray.

pub fn pop(&mut self) -> Option<bool>

Pop an entry from MutableBooleanArray. Note If the values is empty, this method will return None.

pub fn extend_trusted_len_values<I>(&mut self, iterator: I) where
I: TrustedLen<Item = bool>,

Extends the MutableBooleanArray from an iterator of values of trusted len. This differs from extend_trusted_len which accepts in iterator of optional values.

pub unsafe fn extend_trusted_len_values_unchecked<I>(&mut self, iterator: I) where
I: Iterator<Item = bool>,

Extends the MutableBooleanArray from an iterator of values of trusted len. This differs from extend_trusted_len_unchecked, which accepts in iterator of optional values.

Safety

The iterator must be trusted len.

pub fn extend_trusted_len<I, P>(&mut self, iterator: I) where
P: Borrow<bool>,
I: TrustedLen<Item = Option<P>>,

Extends the MutableBooleanArray from an iterator of trusted len.

pub unsafe fn extend_trusted_len_unchecked<I, P>(&mut self, iterator: I) where
P: Borrow<bool>,
I: Iterator<Item = Option<P>>,

Extends the MutableBooleanArray from an iterator of trusted len.

Safety

The iterator must be trusted len.

pub fn into_arc(self) -> Arc<dyn Array>

Converts itself into an Array.

impl MutableBooleanArray

Getters

pub fn values(&self) -> &MutableBitmap

Returns its values.

impl MutableBooleanArray

Setters

pub fn set(&mut self, index: usize, value: Option<bool>)

Sets position index to value. Note that if it is the first time a null appears in this array, this initializes the validity bitmap (O(N)).

Panic

Panics iff index is larger than self.len().

impl MutableBooleanArray

From implementations

pub fn from_trusted_len_values_iter<I: TrustedLen<Item = bool>>(
iterator: I
) -> Self

Creates a new MutableBooleanArray from an TrustedLen of bool.

pub unsafe fn from_trusted_len_values_iter_unchecked<I: Iterator<Item = bool>>(
iterator: I
) -> Self

Creates a new MutableBooleanArray from an TrustedLen of bool. Use this over BooleanArray::from_trusted_len_iter when the iterator is trusted len but this crate does not mark it as such.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_slice<P: AsRef<[bool]>>(slice: P) -> Self

Creates a new MutableBooleanArray from a slice of bool.

pub unsafe fn from_trusted_len_iter_unchecked<I, P>(iterator: I) -> Self where
P: Borrow<bool>,
I: Iterator<Item = Option<P>>,

Creates a BooleanArray from an iterator of trusted length. Use this over BooleanArray::from_trusted_len_iter when the iterator is trusted len but this crate does not mark it as such.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_trusted_len_iter<I, P>(iterator: I) -> Self where
P: Borrow<bool>,
I: TrustedLen<Item = Option<P>>,

Creates a BooleanArray from a TrustedLen.

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I, P>(
iterator: I
) -> Result<Self, E> where
P: Borrow<bool>,
I: Iterator<Item = Result<Option<P>, E>>,

Creates a BooleanArray from an falible iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn try_from_trusted_len_iter<E, I, P>(iterator: I) -> Result<Self, E> where
P: Borrow<bool>,
I: TrustedLen<Item = Result<Option<P>, E>>,

Creates a BooleanArray from a TrustedLen.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the MutableBooleanArray to fit its current length.