Struct arrow2::array::MutableUtf8Array

pub struct MutableUtf8Array<O: Offset> { /* private fields */ }

A MutableArray that builds a Utf8Array. It differs from MutableUtf8ValuesArray in that it can build nullable Utf8Arrays.

Implementations

impl<O: Offset> MutableUtf8Array<O>

pub fn new() -> Self

Initializes a new empty MutableUtf8Array.

pub fn try_new(
data_type: DataType,
offsets: Vec<O>,
values: Vec<u8>,
validity: Option<MutableBitmap>
) -> Result<Self>

Returns a MutableUtf8Array created from its internal representation.

Errors

This function returns an error iff:

the offsets are not monotonically increasing

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len() - 1.

The data_type’s crate::datatypes::PhysicalType is not equal to either Utf8 or LargeUtf8.

The values between two consecutive offsets are not valid utf8

Implementation

This function is O(N) - checking monotinicity and utf8 is O(N)

pub unsafe fn new_unchecked(
data_type: DataType,
offsets: Vec<O>,
values: Vec<u8>,
validity: Option<MutableBitmap>
) -> Self

Create a MutableUtf8Array out of low-end APIs.

Safety

The caller must ensure that every value between offsets is a valid utf8.

Panics

This function panics iff:

The offsets and values are inconsistent

The validity is not None and its length is different from offsets’s length minus one.

pub unsafe fn from_data_unchecked(
data_type: DataType,
offsets: Vec<O>,
values: Vec<u8>,
validity: Option<MutableBitmap>
) -> Self

Alias of new_unchecked

Safety

The caller must ensure that every value between offsets is a valid utf8.

pub fn from_data(
data_type: DataType,
offsets: Vec<O>,
values: Vec<u8>,
validity: Option<MutableBitmap>
) -> Self

The canonical method to create a MutableUtf8Array out of low-end APIs.

Panics

This function panics iff:

The offsets and values are inconsistent

The values between offsets are not utf8 encoded

The validity is not None and its length is different from offsets’s length minus one.

pub fn with_capacity(capacity: usize) -> Self

Initializes a new MutableUtf8Array with a pre-allocated capacity of slots.

pub fn with_capacities(capacity: usize, values: usize) -> Self

Initializes a new MutableUtf8Array with a pre-allocated capacity of slots and values.

pub fn reserve(&mut self, additional: usize, additional_values: usize)

Reserves additional elements and additional_values on the values buffer.

pub fn capacity(&self) -> usize

Reserves additional elements and additional_values on the values buffer.

pub fn len(&self) -> usize

Returns the length of this array

pub fn push<T: AsRef<str>>(&mut self, value: Option<T>)

Pushes a new element to the array.

Panic

This operation panics iff the length of all values (in bytes) exceeds O maximum value.

pub fn value(&self, i: usize) -> &str

Returns the value of the element at index i, ignoring the array’s validity.

Safety

This function is safe iff i < self.len.

pub unsafe fn value_unchecked(&self, i: usize) -> &str

Returns the value of the element at index i, ignoring the array’s validity.

Safety

This function is safe iff i < self.len.

pub fn pop(&mut self) -> Option<String>

Pop the last entry from MutableUtf8Array. This function returns None iff this array is empty.

pub fn iter(
&self
) -> ZipValidity<'_, &str, ArrayValuesIter<'_, MutableUtf8ValuesArray<O>>>

Returns an iterator of Option<&str>

pub fn into_arc(self) -> Arc<dyn Array>

Converts itself into an Array.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the MutableUtf8Array to fit its current length.

pub fn into_data(self) -> (DataType, Vec<O>, Vec<u8>, Option<MutableBitmap>)

Extract the low-end APIs from the MutableUtf8Array.

pub fn values_iter(&self) -> ArrayValuesIter<'_, MutableUtf8ValuesArray<O>>

Returns an iterator of &str

impl<O: Offset> MutableUtf8Array<O>

pub fn values(&self) -> &Vec<u8>

returns its values.

pub fn offsets(&self) -> &Vec<O>

returns its offsets.

impl<O: Offset> MutableUtf8Array<O>

pub fn extend_trusted_len_values<I, P>(&mut self, iterator: I)where
P: AsRef<str>,
I: TrustedLen<Item = P>,

Extends the MutableUtf8Array from an iterator of values of trusted len. This differs from extended_trusted_len which accepts iterator of optional values.

pub fn extend_values<I, P>(&mut self, iterator: I)where
P: AsRef<str>,
I: Iterator<Item = P>,

Extends the MutableUtf8Array from an iterator of values. This differs from extended_trusted_len which accepts iterator of optional values.

pub unsafe fn extend_trusted_len_values_unchecked<I, P>(&mut self, iterator: I)where
P: AsRef<str>,
I: Iterator<Item = P>,

Extends the MutableUtf8Array from an iterator of values of trusted len. This differs from extended_trusted_len_unchecked which accepts iterator of optional values.

Safety

The iterator must be trusted len.

pub fn extend_trusted_len<I, P>(&mut self, iterator: I)where
P: AsRef<str>,
I: TrustedLen<Item = Option<P>>,

Extends the MutableUtf8Array from an iterator of trusted len.

pub unsafe fn extend_trusted_len_unchecked<I, P>(&mut self, iterator: I)where
P: AsRef<str>,
I: Iterator<Item = Option<P>>,

Extends MutableUtf8Array from an iterator of trusted len.

Safety

The iterator must be trusted len.

pub unsafe fn from_trusted_len_iter_unchecked<I, P>(iterator: I) -> Selfwhere
P: AsRef<str>,
I: Iterator<Item = Option<P>>,

Creates a MutableUtf8Array from an iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_trusted_len_iter<I, P>(iterator: I) -> Selfwhere
P: AsRef<str>,
I: TrustedLen<Item = Option<P>>,

Creates a MutableUtf8Array from an iterator of trusted length.

pub unsafe fn from_trusted_len_values_iter_unchecked<T: AsRef<str>, I: Iterator<Item = T>>(
iterator: I
) -> Self

Creates a MutableUtf8Array from an iterator of trusted length of &str.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_trusted_len_values_iter<T: AsRef<str>, I: TrustedLen<Item = T>>(
iterator: I
) -> Self

Creates a new MutableUtf8Array from a TrustedLen of &str.

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I, P>(
iterator: I
) -> Result<Self, E>where
P: AsRef<str>,
I: IntoIterator<Item = Result<Option<P>, E>>,

Creates a MutableUtf8Array from an falible iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn try_from_trusted_len_iter<E, I, P>(iterator: I) -> Result<Self, E>where
P: AsRef<str>,
I: TrustedLen<Item = Result<Option<P>, E>>,

Creates a MutableUtf8Array from an falible iterator of trusted length.

pub fn from_iter_values<T: AsRef<str>, I: Iterator<Item = T>>(
iterator: I
) -> Self

Creates a new MutableUtf8Array from a Iterator of &str.