Struct arrow2::array::PrimitiveArray

pub struct PrimitiveArray<T: NativeType> { /* private fields */ }

A PrimitiveArray is arrow’s equivalent to Vec<Option<T: NativeType>>, i.e. an array designed for highly performant operations on optionally nullable slots, backed by a physical type of a fixed byte-width, such as i32 or f64. The size of this struct is O(1) as all data is stored behind an std::sync::Arc.

Example

use arrow2::array::{PrimitiveArray, Array};
use arrow2::bitmap::Bitmap;
let array = PrimitiveArray::from([Some(1), None, Some(10)]);
assert_eq!(array.values().as_slice(), &[1, 0, 10]);
assert_eq!(array.validity(), Some(&Bitmap::from([true, false, true])));

Implementations

impl<T: NativeType> PrimitiveArray<T>

pub fn from_values<I: IntoIterator<Item = T>>(iter: I) -> Self

Creates a (non-null) PrimitiveArray from an iterator of values.

Implementation

This does not assume that the iterator has a known length.

pub fn from_slice<P: AsRef<[T]>>(slice: P) -> Self

Creates a (non-null) PrimitiveArray from a slice of values.

Implementation

This is essentially a memcopy

pub fn from_vec(array: Vec<T>) -> Self

Creates a (non-null) PrimitiveArray from a vector of values. This does not have memcopy and is the fastest way to create a PrimitiveArray.

impl<T: NativeType> PrimitiveArray<T>

pub fn from_trusted_len_values_iter<I: TrustedLen<Item = T>>(iter: I) -> Self

Creates a (non-null) PrimitiveArray from a TrustedLen of values.

Implementation

This does not assume that the iterator has a known length.

pub unsafe fn from_trusted_len_values_iter_unchecked<I: Iterator<Item = T>>(
iter: I
) -> Self

Creates a new PrimitiveArray from an iterator over values

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_trusted_len_iter<I: TrustedLen<Item = Option<T>>>(iter: I) -> Self

Creates a PrimitiveArray from a TrustedLen of optional values.

pub unsafe fn from_trusted_len_iter_unchecked<I: Iterator<Item = Option<T>>>(
iter: I
) -> Self

Creates a PrimitiveArray from an iterator of optional values.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

impl<'a, T: NativeType> PrimitiveArray<T>

pub fn iter(&'a self) -> ZipValidity<'a, &'a T, Iter<'a, T>>

constructs a new iterator

impl<T: NativeType> PrimitiveArray<T>

pub fn try_new(
data_type: DataType,
values: Buffer<T>,
validity: Option<Bitmap>
) -> Result<Self, ArrowError>

The canonical method to create a PrimitiveArray.

Errors

This function errors iff:

The validity is not None and its length is different from values’s length

The data_type’s PhysicalType is not equal to PhysicalType::Primitive.

pub fn new(
data_type: DataType,
values: Buffer<T>,
validity: Option<Bitmap>
) -> Self

The canonical method to create a PrimitiveArray

Panics

This function errors iff:

The validity is not None and its length is different from values’s length

The data_type’s PhysicalType is not equal to PhysicalType::Primitive.

pub fn from_data(
data_type: DataType,
values: Buffer<T>,
validity: Option<Bitmap>
) -> Self

Alias for new

pub fn new_empty(data_type: DataType) -> Self

Returns a new empty PrimitiveArray.

pub fn new_null(data_type: DataType, length: usize) -> Self

Returns a new PrimitiveArray whose all slots are null / None.

impl<T: NativeType> PrimitiveArray<T>

pub fn slice(&self, offset: usize, length: usize) -> Self

Returns a slice of this PrimitiveArray.

Implementation

This operation is O(1) as it amounts to increase two ref counts.

Panic

This function panics iff offset + length >= self.len().

pub unsafe fn slice_unchecked(&self, offset: usize, length: usize) -> Self

Returns a slice of this PrimitiveArray.

Implementation

This operation is O(1) as it amounts to increase two ref counts.

Safety

The caller must ensure that offset + length <= self.len().

pub fn with_validity(&self, validity: Option<Bitmap>) -> Self

Sets the validity bitmap on this PrimitiveArray.

Panics

This function panics iff validity.len() != self.len().

impl<T: NativeType> PrimitiveArray<T>

pub fn len(&self) -> usize

Returns the length of this array

pub fn validity(&self) -> Option<&Bitmap>

The optional validity.

pub fn data_type(&self) -> &DataType

The arrays’ logical type

pub fn values(&self) -> &Buffer<T>

The values Buffer. Values on null slots are undetermined (they can be anything).

pub fn value(&self, i: usize) -> T

Returns the value at slot i. Equivalent to self.values()[i]. The value on null slots is undetermined (it can be anything).

pub unsafe fn value_unchecked(&self, i: usize) -> T

Returns the element at index i as T. The value on null slots is undetermined (it can be anything).

Safety

Caller must be sure that i < self.len()

pub fn to(self, data_type: DataType) -> Self

Returns a new PrimitiveArray with a different logical type. This is O(1).

Panics

Panics iff the data_type is not supported for the physical type.

pub fn into_mut(self) -> Either<Self, MutablePrimitiveArray<T>>

Try to convert this PrimitiveArray to a MutablePrimitiveArray