Struct arrow2::array::UnionArray

pub struct UnionArray { /* private fields */ }

UnionArray represents an array whose each slot can contain different values.

Implementations

impl<'a> UnionArray

pub fn iter(&'a self) -> UnionIter<'a>

constructs a new iterator

impl UnionArray

pub fn try_new(
data_type: DataType,
types: Buffer<i8>,
fields: Vec<Arc<dyn Array>>,
offsets: Option<Buffer<i32>>
) -> Result<Self, ArrowError>

Returns a new UnionArray.

Errors

This function errors iff:

data_type’s physical type is not crate::datatypes::PhysicalType::Union.

the fields’s len is different from the data_type’s children’s length

any of the values’s data type is different from its corresponding children’ data type

pub fn new(
data_type: DataType,
types: Buffer<i8>,
fields: Vec<Arc<dyn Array>>,
offsets: Option<Buffer<i32>>
) -> Self

Returns a new UnionArray.

Panics

This function panics iff:

data_type’s physical type is not crate::datatypes::PhysicalType::Union.

the fields’s len is different from the data_type’s children’s length

any of the values’s data type is different from its corresponding children’ data type

pub fn from_data(
data_type: DataType,
types: Buffer<i8>,
fields: Vec<Arc<dyn Array>>,
offsets: Option<Buffer<i32>>
) -> Self

Alias for new

pub fn new_null(data_type: DataType, length: usize) -> Self

Creates a new null UnionArray.

pub fn new_empty(data_type: DataType) -> Self

Creates a new empty UnionArray.

impl UnionArray

pub fn slice(&self, offset: usize, length: usize) -> Self

Returns a slice of this UnionArray.

Implementation

This operation is O(F) where F is the number of fields.

Panic

This function panics iff offset + length >= self.len().

pub unsafe fn slice_unchecked(&self, offset: usize, length: usize) -> Self

Returns a slice of this UnionArray.

Implementation

This operation is O(F) where F is the number of fields.

Safety

The caller must ensure that offset + length <= self.len().

impl UnionArray

pub fn len(&self) -> usize

Returns the length of this array

pub fn offsets(&self) -> Option<&Buffer<i32>>

The optional offsets.

pub fn fields(&self) -> &Vec<Arc<dyn Array>>

The fields.

pub fn types(&self) -> &Buffer<i8>

The types.

pub fn index(&self, index: usize) -> (usize, usize)

Returns the index and slot of the field to select from self.fields.

pub fn value(&self, index: usize) -> Box<dyn Scalar>

Returns the slot index as a Scalar.

impl UnionArray

pub fn get_fields(data_type: &DataType) -> &[Field]

Returns all fields from DataType::Union.

Panic

Panics iff data_type’s logical type is not DataType::Union.

pub fn is_sparse(data_type: &DataType) -> bool

Returns whether the DataType::Union is sparse or not.

Panic

Panics iff data_type’s logical type is not DataType::Union.