Trait arrow2::array::Array

pub trait Array: Debug + Send + Sync + ToFfi {
    fn as_any(&self) -> &dyn Any;

    fn len(&self) -> usize;

    fn data_type(&self) -> &DataType;

    fn validity(&self) -> &Option<Bitmap>;

    fn slice(&self, offset: usize, length: usize) -> Box<dyn Array>;

    fn is_empty(&self) -> bool { ... }

    fn null_count(&self) -> usize { ... }

    fn is_null(&self, i: usize) -> bool { ... }

    fn is_valid(&self, i: usize) -> bool { ... }
}

A trait representing an Arrow array. Arrow arrays are trait objects that are infalibly downcasted to concrete types according to the Array::data_type.

Required methods

fn as_any(&self) -> &dyn Any

fn len(&self) -> usize

The length of the Array. Every array has a length corresponding to the number of elements (slots).

fn data_type(&self) -> &DataType

The DataType of the Array. In combination with Array::as_any, this can be used to downcast trait objects (dyn Array) to concrete arrays.

fn validity(&self) -> &Option<Bitmap>

The validity of the Array: every array has an optional Bitmap that, when available specifies whether the array slot is valid or not (null). When the validity is None, all slots are valid.

fn slice(&self, offset: usize, length: usize) -> Box<dyn Array>

Slices the Array, returning a new Box<dyn Array>.

Implementation

This operation is O(1) over len, as it amounts to increase two ref counts and moving the struct to the heap.

Panic

This function panics iff offset + length >= self.len().

Loading content...
Provided methods

fn is_empty(&self) -> bool

whether the array is empty

fn null_count(&self) -> usize

The number of null slots on this Array. This is usually used to branch implementations to cases where optimizations can be made.

Implementation

This is O(1).

fn is_null(&self, i: usize) -> bool

Returns whether slot i is null.

Panic

Panics iff i >= self.len().

fn is_valid(&self, i: usize) -> bool

Returns whether slot i is valid.

Panic

Panics iff i >= self.len().

Loading content...