Struct arrow2::bitmap::Bitmap

pub struct Bitmap { /* private fields */ }

An immutable container semantically equivalent to Arc<Vec<bool>> but represented as Arc<Vec<u8>> where each boolean is represented as a single bit.

Examples

use arrow2::bitmap::{Bitmap, MutableBitmap};

let bitmap = Bitmap::from([true, false, true]);
assert_eq!(bitmap.iter().collect::<Vec<_>>(), vec![true, false, true]);

// creation directly from bytes
let bitmap = Bitmap::try_new(vec![0b00001101], 5).unwrap();
// note: the first bit is the left-most of the first byte
assert_eq!(bitmap.iter().collect::<Vec<_>>(), vec![true, false, true, true, false]);
// we can also get the slice:
assert_eq!(bitmap.as_slice(), ([0b00001101u8].as_ref(), 0, 5));
// debug helps :)
assert_eq!(format!("{:?}", bitmap), "[0b___01101]".to_string());

// it supports copy-on-write semantics (to a `MutableBitmap`)
let bitmap: MutableBitmap = bitmap.into_mut().right().unwrap();
assert_eq!(bitmap, MutableBitmap::from([true, false, true, true, false]));

// slicing is 'O(1)' (data is shared)
let bitmap = Bitmap::try_new(vec![0b00001101], 5).unwrap();
let sliced = bitmap.slice(1, 4);
assert_eq!(sliced.as_slice(), ([0b00001101u8].as_ref(), 1, 4)); // 1 here is the offset:
assert_eq!(format!("{:?}", sliced), "[0b___0110_]".to_string());
// when sliced (or cloned), it is no longer possible to `into_mut`.
let same: Bitmap = sliced.into_mut().left().unwrap();

Implementations

impl Bitmap

pub fn new() -> Self

Initializes an empty Bitmap.

pub fn try_new(bytes: Vec<u8>, length: usize) -> Result<Self, Error>

Initializes a new Bitmap from vector of bytes and a length.

Errors

This function errors iff length > bytes.len() * 8

pub fn len(&self) -> usize

Returns the length of the Bitmap.

pub fn is_empty(&self) -> bool

Returns whether Bitmap is empty

pub fn iter(&self) -> BitmapIter<'_>

Returns a new iterator of bool over this bitmap

pub fn chunks<T: BitChunk>(&self) -> BitChunks<'_, T>

Returns an iterator over bits in bit chunks BitChunk.

This iterator is useful to operate over multiple bits via e.g. bitwise.

pub fn as_slice(&self) -> (&[u8], usize, usize)

Returns the byte slice of this Bitmap.

The returned tuple contains:

.1: The byte slice, truncated to the start of the first bit. So the start of the slice is within the first 8 bits.

.2: The start offset in bits on a range 0 <= offsets < 8.

.3: The length in number of bits.

pub const fn unset_bits(&self) -> usize

Returns the number of unset bits on this Bitmap.

Guaranted to be <= self.len().

Implementation

This function is O(1) - the number of unset bits is computed when the bitmap is created

pub fn null_count(&self) -> usize

ðŸ‘Ž Deprecated since 0.13.0:
use unset_bits instead

Returns the number of unset bits on this Bitmap.

pub fn slice(self, offset: usize, length: usize) -> Self

Slices self, offsetting by offset and truncating up to length bits.

Panic

Panics iff offset + length > self.length, i.e. if the offset and length exceeds the allocated capacity of self.

pub unsafe fn slice_unchecked(self, offset: usize, length: usize) -> Self

Slices self, offseting by offset and truncating up to length bits.

Safety

The caller must ensure that self.offset + offset + length <= self.len()

pub fn get_bit(&self, i: usize) -> bool

Returns whether the bit at position i is set.

Panics

Panics iff i >= self.len().

pub unsafe fn get_bit_unchecked(&self, i: usize) -> bool

Unsafely returns whether the bit at position i is set.

Safety

Unsound iff i >= self.len().

pub fn into_mut(self) -> Either<Self, MutableBitmap>

Converts this Bitmap to MutableBitmap, returning itself if the conversion is not possible

This operation returns a MutableBitmap iff:

this Bitmap is not an offsetted slice of another Bitmap

this Bitmap has not been cloned (i.e. Arc::get_mut yields Some)

this Bitmap was not imported from the c data interface (FFI)

pub fn make_mut(self) -> MutableBitmap

Converts this Bitmap into a MutableBitmap, cloning its internal buffer if required (clone-on-write).

pub fn new_zeroed(length: usize) -> Self

Initializes an new Bitmap filled with unset values.

pub fn null_count_range(&self, offset: usize, length: usize) -> usize

Counts the nulls (unset bits) starting from offset bits and for length bits.

pub fn from_u8_slice<T: AsRef<[u8]>>(slice: T, length: usize) -> Self

Creates a new Bitmap from a slice and length.

Panic

Panics iff length <= bytes.len() * 8

pub fn from_u8_vec(vec: Vec<u8>, length: usize) -> Self

Alias for Bitmap::try_new().unwrap() This function is O(1)

Panic

This function panics iff length <= bytes.len() * 8

pub fn get(&self, i: usize) -> Option<bool>

Returns whether the bit at position i is set.

impl Bitmap

pub unsafe fn from_trusted_len_iter_unchecked<I: Iterator<Item = bool>>(
iterator: I
) -> Self

Creates a new Bitmap from an iterator of booleans.

Safety

The iterator must report an accurate length.

pub fn from_trusted_len_iter<I: TrustedLen<Item = bool>>(iterator: I) -> Self

Creates a new Bitmap from an iterator of booleans.

pub fn try_from_trusted_len_iter<E, I: TrustedLen<Item = Result<bool, E>>>(
iterator: I
) -> Result<Self, E>

Creates a new Bitmap from a fallible iterator of booleans.

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I: Iterator<Item = Result<bool, E>>>(
iterator: I
) -> Result<Self, E>

Creates a new Bitmap from a fallible iterator of booleans.

Safety

The iterator must report an accurate length.