Module arrow2::compute

Contains operators over arrays. This module’s general design is that each operator has two interfaces, a statically-typed version and a dynamically-typed version. The statically-typed version expects concrete arrays (like PrimitiveArray); the dynamically-typed version expects &dyn Array and errors if the the type is not supported. Some dynamically-typed operators have an auxiliary function, can_*, that returns true if the operator can be applied to the particular DataType.

Modules

aggregate

arithmetics
Defines basic arithmetic kernels for PrimitiveArrays.

arity
Defines kernels suitable to perform operations to primitive arrays.

boolean

boolean_kleene

cast

comparison
Defines basic comparison kernels for PrimitiveArrays.

concat
Defines concat kernel for ArrayRef

contains

filter

hash

if_then_else

length
Defines kernel for length of composite arrays

like

limit

merge_sort
This module exposes functions to perform merge-sorts.

nullif

regex_match

sort
Defines sort kernel for ArrayRef

substring
Defines kernel to extract a substring of a [Large]StringArray

take
Defines take kernel for Array

temporal
Defines temporal kernels for time and date related functions.

window
Defines windowing functions, like shifting