Module arrow2::compute

contains a wide range of compute operations (e.g. arithmetics, aggregate, filter, comparison, and sort)

This module’s general design is that each operator has two interfaces, a statically-typed version and a dynamically-typed version. The statically-typed version expects concrete arrays (such as PrimitiveArray); the dynamically-typed version expects &dyn Array and errors if the the type is not supported. Some dynamically-typed operators have an auxiliary function, can_*, that returns true if the operator can be applied to the particular DataType.

Modules

aggregatecompute_aggregate

Contains different aggregation functions

arithmeticscompute_arithmetics

Defines basic arithmetic kernels for PrimitiveArrays.

arity

Defines kernels suitable to perform operations to primitive arrays.

arity_assign

Defines generics suitable to perform operations to PrimitiveArray in-place.

bitwisecompute_bitwise

Contains bitwise operators: or, and, xor and not.

booleancompute_boolean

null-preserving operators such as and, or and not.

boolean_kleenecompute_boolean_kleene

Boolean operators of Kleene logic.

castcompute_cast

Defines different casting operators such as cast or primitive_to_binary.

comparisoncompute_comparison

Contains comparison operators

concatenatecompute_concatenate

Contains the concatenate kernel

containscompute_contains

Declares the contains operator

filtercompute_filter

Contains operators to filter arrays such as filter.

hashcompute_hash

Contains the hash and typed (e.g. hash_primitive) operators.

if_then_elsecompute_if_then_else

Contains the operator if_then_else.

lengthcompute_length

Defines kernel for length of composite arrays

likecompute_like

Contains “like” operators such as like_utf8 and like_utf8_scalar.

limitcompute_limit

Contains the operator limit.

merge_sortcompute_merge_sort

Functions to perform merge-sorts.

nullifcompute_nullif

Contains the operator nullif.

partitioncompute_partition

Defines partition kernel for crate::array::Array

regex_matchcompute_regex_match

Contains regex matching operators regex_match and regex_match_scalar.

sortcompute_sort

Contains operators to sort individual and slices of Arrays.

substringcompute_substring

Defines kernel to extract a substring of a [Large]StringArray

takecompute_take

Defines take kernel for Array

temporalcompute_temporal

Defines temporal kernels for time and date related functions.

utf8compute_utf8

Defines common maps to a Utf8Array

windowcompute_window

Defines windowing functions, like shifting