Function arrow2::compute::aggregate::min_primitive

pub fn min_primitive<T>(array: &PrimitiveArray<T>) -> Option<T> where
    T: NativeType + Simd,
    T::Simd: SimdOrd<T>, 

Returns the minimum value in the array, according to the natural order. For floating point arrays any NaN values are considered to be greater than any other non-null value