Module arrow2::compute::arithmetics::basic::sub

This is supported on crate feature compute only.

Definition of basic sub operations with primitive arrays

Functions

checked_sub

Checked subtraction of two primitive arrays. If the result from the subtraction overflow, the validity for that index is changed

checked_sub_scalar

Checked subtraction of a scalar T to a primitive array of type T. If the result from the subtraction overflows, then the validity for that index is changed to None

overflowing_sub

Overflowing subtraction of two primitive arrays. If the result from the sub is smaller than the possible number for this type, the result for the operation will be an array with overflowed values and a validity array indicating the overflowing elements from the array.

overflowing_sub_scalar

Overflowing subtraction of a scalar T to a primitive array of type T. If the result from the sub is smaller than the possible number for this type, then the result will be an array with overflowed values and a validity array indicating the overflowing elements from the array

saturating_sub

Saturating subtraction of two primitive arrays. If the result from the sub is smaller than the possible number for this type, the result for the operation will be the saturated value.

saturating_sub_scalar

Saturated subtraction of a scalar T to a primitive array of type T. If the result from the sub is smaller than the possible number for this type, then the result will be saturated

sub

Subtracts two primitive arrays with the same type. Panics if the subtraction of one pair of values overflows.

sub_scalar

Subtract a scalar T to a primitive array of type T. Panics if the subtraction of the values overflows.