Function arrow2::compute::cast::dictionary_to_dictionary_keys

pub fn dictionary_to_dictionary_keys<K1, K2>(
    from: &DictionaryArray<K1>
) -> Result<DictionaryArray<K2>> where
    K1: DictionaryKey,
    K2: DictionaryKey, 

Available on crate feature compute_cast only.

Casts a DictionaryArray to a new DictionaryArray backed by a different physical type of the keys, while keeping the values equal.

Errors

Errors if any of the old keys’ values is larger than the maximum value supported by the new physical type.