Module arrow2::compute::sort

Available on crate feature compute_sort only.

Contains operators to sort individual and slices of Arrays.

Structs

SortColumn

One column to be used in lexicographical sort

SortOptions

Options that define how sort kernels should behave

Functions

can_sort

Checks if an array of type datatype can be sorted

lexsort

Sort a list of Array using SortOptions provided for each array.

lexsort_to_indices

Sorts a list of SortColumn into a non-nullable PrimitiveArray representing the indices that would sort the columns.

sort

Sort the Array using SortOptions.

sort_to_indices

Sort elements from values into a non-nullable PrimitiveArray of indices that sort values.