Module arrow2::datatypes

Metadata declarations such as DataType, Field and Schema.

Structs

Field
A logical DataType and its associated metadata per Arrow specification

Schema
An ordered sequence of Field with optional metadata.

Enums

DataType
The set of datatypes that are supported by this implementation of Apache Arrow.

DictionaryIndexType
the set of valid indices used to index a dictionary-encoded Array.

IntervalUnit
Interval units defined in Arrow

PhysicalType
The set of physical types: unique in-memory representations of an Arrow array. A physical type has a one-to-many relationship with a crate::datatypes::DataType and a one-to-one mapping with each struct in this crate that implements crate::array::Array.

PrimitiveType

TimeUnit
Time units defined in Arrow.

Type Definitions

SchemaRef
typedef for Arc<Schema>.