Module arrow2::datatypes

Metadata declarations such as DataType, Field and Schema.

Structs

Field

A logical DataType and its associated metadata per Arrow specification

Schema

An ordered sequence of Field with optional metadata.

Enums

DataType

The set of supported logical types. Each variant uniquely identifies a logical type, which define specific semantics to the data (e.g. how it should be represented). A DataType has an unique corresponding PhysicalType, obtained via DataType::to_physical_type, which uniquely identifies an in-memory representation of data. The DataType::Extension is special in that it augments a DataType with metadata to support custom types. Use to_logical_type to desugar such type and return its correspoding logical type.

DictionaryIndexType

the set of valid indices used to index a dictionary-encoded Array.

IntervalUnit

Interval units defined in Arrow

PhysicalType

The set of physical types: unique in-memory representations of an Arrow array. A physical type has a one-to-many relationship with a crate::datatypes::DataType and a one-to-one mapping with each struct in this crate that implements crate::array::Array.

PrimitiveType

TimeUnit

Time units defined in Arrow.

Type Definitions

SchemaRef

typedef for Arc<Schema>.