Struct arrow2::datatypes::Schema

pub struct Schema { /* fields omitted */ }

An ordered sequence of Field with optional metadata.

Implementations

impl Schema

pub fn empty() -> Self

Creates an empty Schema.

pub fn new(fields: Vec<Field>) -> Self

Creates a new Schema from a sequence of Field values.

Example

let field_a = Field::new("a", DataType::Int64, false);
let field_b = Field::new("b", DataType::Boolean, false);

let schema = Schema::new(vec![field_a, field_b]);

pub const fn new_from(
fields: Vec<Field>,
metadata: HashMap<String, String>
) -> Self

Creates a new Schema from a sequence of Field values and additional metadata.

Example

let field_a = Field::new("a", DataType::Int64, false);
let field_b = Field::new("b", DataType::Boolean, false);

let mut metadata: HashMap<String, String> = HashMap::new();
metadata.insert("row_count".to_string(), "100".to_string());

let schema = Schema::new_from(vec![field_a, field_b], metadata);

pub fn with_metadata(self, metadata: HashMap<String, String>) -> Self

Creates a new Field with metadata.

pub fn try_merge(schemas: impl IntoIterator<Item = Self>) -> Result<Self>

Merge schema into self if it is compatible. Struct fields will be merged recursively.

Example:

use arrow2::datatypes::*;

let merged = Schema::try_merge(vec![
    Schema::new(vec![
        Field::new("c1", DataType::Int64, false),
        Field::new("c2", DataType::Utf8, false),
    ]),
    Schema::new(vec![
        Field::new("c1", DataType::Int64, true),
        Field::new("c2", DataType::Utf8, false),
        Field::new("c3", DataType::Utf8, false),
    ]),
]).unwrap();

assert_eq!(
    merged,
    Schema::new(vec![
        Field::new("c1", DataType::Int64, true),
        Field::new("c2", DataType::Utf8, false),
        Field::new("c3", DataType::Utf8, false),
    ]),
);

pub const fn fields(&self) -> &Vec<Field>

Returns all Fields in this schema.

pub fn field(&self, i: usize) -> &Field

Returns the Field at position i.

Panics

Panics iff i is larger than the number of fields in this Schema.

pub fn field_with_name(&self, name: &str) -> Result<&Field>

Returns the first Field named name.

pub fn fields_with_dict_id(&self, dict_id: i64) -> Vec<&Field>

Returns all Fields with dictionary id dict_id.

pub fn index_of(&self, name: &str) -> Result<usize>

Find the index of the column with the given name.

pub const fn metadata(&self) -> &HashMap<String, String>

Returns an immutable reference to the Map of custom metadata key-value pairs.

pub fn column_with_name(&self, name: &str) -> Option<(usize, &Field)>

Look up a column by name and return a immutable reference to the column along with its index.