Module arrow2::io

Contains modules to interface with other formats such as csv, parquet, json, ipc, print and avro.

Modules

avroio_avro

Read and write from and to Apache Avro

csvio_csv_read or io_csv_read_async or io_csv_write

Convert data between the Arrow and CSV (comma-separated values).

flightio_flight

Serialization and deserialization to Arrow’s flight protocol

ipcio_ipc

APIs to read from and write to Arrow’s IPC format.

jsonio_json

Convert data between the Arrow memory format and JSON line-delimited records.

json_integrationio_json_integration

Utils for JSON integration testing

ndjsonio_json

APIs to read from and write to NDJSON

odbc

API to serialize and deserialize data from and to ODBC

orcio_orc

APIs to read from ORC format.

parquetio_parquet

APIs to read from and write to Parquet format.

printio_print

APIs to represent Chunk as a formatted table.