Module arrow2::io::avro::write

Available on crate feature io_avro only.

APIs to write to Avro format.

Re-exports

pub use super::Compression;

pub use super::Block;

pub use super::CompressedBlock;

Functions

can_serialize

Whether new_serializer supports data_type.

compress

Compresses a Block to a CompressedBlock.

new_serializer

Creates a StreamingIterator trait object that presents items from array encoded according to schema.

serialize

consumes a set of BoxSerializer into an Block.

to_avro_schema

Converts a Schema to a vector of [AvroField] with it.

write_block

Writes a CompressedBlock to writer

write_metadata

Writes Avro’s metadata to writer.

Type Definitions

BoxSerializer

A type alias for a boxed StreamingIterator, used to write arrays into avro rows (i.e. a column -> row transposition of types known at run-time)