Module arrow2::io::csv::read

Available on (crate features io_csv_read or io_csv_read_async or io_csv_write) and crate feature io_csv_read only.

APIs to read from CSV

Structs

ByteRecord

A single CSV record stored as raw bytes.

Reader

A already configured CSV reader.

ReaderBuilder

Builds a CSV reader with various configuration knobs.

Functions

deserialize_batch

Deserializes rows ByteRecord into a Chunk. Note that this is a convenience function: column deserialization is trivially parallelizable (e.g. rayon).

deserialize_column

Deserializes column of rows into an Array of DataType datatype.

infer

Infers DataType from bytes

infer_schema

Infers the Fields of a CSV file by reading through the first n records up to max_rows. Also returns the number of rows used to infer. Seeks back to the begining of the file after the header

read_rows

Reads len rows from reader into row, skiping the first skip. This operation has minimal CPU work and is thus the fastest way to read through a CSV without deserializing the contents to Arrow.