Function arrow2::io::csv::write::new_serializer

pub fn new_serializer<'a>(
    array: &'a dyn Array,
    options: &'a SerializeOptions
) -> Result<Box<dyn StreamingIterator<Item = [u8]> + 'a>>

Available on (crate features io_csv_read or io_csv_read_async or io_csv_write) and crate feature io_csv_write only.

Returns a StreamingIterator that yields &[u8] serialized from array according to options. For numeric types, this serializes as usual. For dates, times and timestamps, it uses options to Supported types:

boolean

numeric types (i.e. floats, int, uint)

times and dates

naive timestamps (timestamps without timezone information)

Error

This function errors if any of the logical types in batch is not supported.