Struct arrow2::io::ipc::gen::Schema::Field

pub struct Field<'a> {
    pub _tab: Table<'a>,
}

A field represents a named column in a record / row batch or child of a nested type.

Fields
_tab: Table<'a>
Implementations

impl<'a> Field<'a>

pub fn init_from_table(table: Table<'a>) -> Self

pub fn create<'bldr: 'args, 'args: 'mut_bldr, 'mut_bldr>(
_fbb: &'mut_bldr mut FlatBufferBuilder<'bldr>,
args: &'args FieldArgs<'args>
) -> WIPOffset<Field<'bldr>>

pub const VT_NAME: VOffsetT

pub const VT_NULLABLE: VOffsetT

pub const VT_TYPE_TYPE: VOffsetT

pub const VT_TYPE_: VOffsetT

pub const VT_DICTIONARY: VOffsetT

pub const VT_CHILDREN: VOffsetT

pub const VT_CUSTOM_METADATA: VOffsetT

pub fn name(&self) -> Option<&'a str>

Name is not required, in i.e. a List

pub fn nullable(&self) -> bool

Whether or not this field can contain nulls. Should be true in general.

pub fn type_type(&self) -> Type

pub fn type_(&self) -> Option<Table<'a>>

This is the type of the decoded value if the field is dictionary encoded.

pub fn dictionary(&self) -> Option<DictionaryEncoding<'a>>

Present only if the field is dictionary encoded.

pub fn children(&self) -> Option<Vector<'a, ForwardsUOffset<Field<'a>>>>

children apply only to nested data types like Struct, List and Union. For primitive types children will have length 0.

pub fn custom_metadata(
&self
) -> Option<Vector<'a, ForwardsUOffset<KeyValue<'a>>>>

User-defined metadata

pub fn type_as_null(&self) -> Option<Null<'a>>

pub fn type_as_int(&self) -> Option<Int<'a>>

pub fn type_as_floating_point(&self) -> Option<FloatingPoint<'a>>

pub fn type_as_binary(&self) -> Option<Binary<'a>>

pub fn type_as_utf_8(&self) -> Option<Utf8<'a>>

pub fn type_as_bool(&self) -> Option<Bool<'a>>

pub fn type_as_decimal(&self) -> Option<Decimal<'a>>

pub fn type_as_date(&self) -> Option<Date<'a>>

pub fn type_as_time(&self) -> Option<Time<'a>>

pub fn type_as_timestamp(&self) -> Option<Timestamp<'a>>

pub fn type_as_interval(&self) -> Option<Interval<'a>>

pub fn type_as_list(&self) -> Option<List<'a>>

pub fn type_as_struct_(&self) -> Option<Struct_<'a>>

pub fn type_as_union(&self) -> Option<Union<'a>>

pub fn type_as_fixed_size_binary(&self) -> Option<FixedSizeBinary<'a>>

pub fn type_as_fixed_size_list(&self) -> Option<FixedSizeList<'a>>

pub fn type_as_map(&self) -> Option<Map<'a>>

pub fn type_as_duration(&self) -> Option<Duration<'a>>

pub fn type_as_large_binary(&self) -> Option<LargeBinary<'a>>

pub fn type_as_large_utf_8(&self) -> Option<LargeUtf8<'a>>

pub fn type_as_large_list(&self) -> Option<LargeList<'a>>