Struct arrow2::io::ipc::gen::SparseTensor::SparseTensor

pub struct SparseTensor<'a> {
    pub _tab: Table<'a>,
}

Fields
_tab: Table<'a>
Implementations

impl<'a> SparseTensor<'a>

pub fn init_from_table(table: Table<'a>) -> Self

pub fn create<'bldr: 'args, 'args: 'mut_bldr, 'mut_bldr>(
_fbb: &'mut_bldr mut FlatBufferBuilder<'bldr>,
args: &'args SparseTensorArgs<'args>
) -> WIPOffset<SparseTensor<'bldr>>

pub const VT_TYPE_TYPE: VOffsetT

pub const VT_TYPE_: VOffsetT

pub const VT_SHAPE: VOffsetT

pub const VT_NON_ZERO_LENGTH: VOffsetT

pub const VT_SPARSEINDEX_TYPE: VOffsetT

pub const VT_SPARSEINDEX: VOffsetT

pub const VT_DATA: VOffsetT

pub fn type_type(&self) -> Type

pub fn type_(&self) -> Table<'a>

The type of data contained in a value cell. Currently only fixed-width value types are supported, no strings or nested types.

pub fn shape(&self) -> Vector<'a, ForwardsUOffset<TensorDim<'a>>>

The dimensions of the tensor, optionally named.

pub fn non_zero_length(&self) -> i64

The number of non-zero values in a sparse tensor.

pub fn sparseIndex_type(&self) -> SparseTensorIndex

pub fn sparseIndex(&self) -> Table<'a>

Sparse tensor index

pub fn data(&self) -> &'a Buffer

The location and size of the tensor’s data

pub fn type_as_null(&self) -> Option<Null<'a>>

pub fn type_as_int(&self) -> Option<Int<'a>>

pub fn type_as_floating_point(&self) -> Option<FloatingPoint<'a>>

pub fn type_as_binary(&self) -> Option<Binary<'a>>

pub fn type_as_utf_8(&self) -> Option<Utf8<'a>>

pub fn type_as_bool(&self) -> Option<Bool<'a>>

pub fn type_as_decimal(&self) -> Option<Decimal<'a>>

pub fn type_as_date(&self) -> Option<Date<'a>>

pub fn type_as_time(&self) -> Option<Time<'a>>

pub fn type_as_timestamp(&self) -> Option<Timestamp<'a>>

pub fn type_as_interval(&self) -> Option<Interval<'a>>

pub fn type_as_list(&self) -> Option<List<'a>>

pub fn type_as_struct_(&self) -> Option<Struct_<'a>>

pub fn type_as_union(&self) -> Option<Union<'a>>

pub fn type_as_fixed_size_binary(&self) -> Option<FixedSizeBinary<'a>>

pub fn type_as_fixed_size_list(&self) -> Option<FixedSizeList<'a>>

pub fn type_as_map(&self) -> Option<Map<'a>>

pub fn type_as_duration(&self) -> Option<Duration<'a>>

pub fn type_as_large_binary(&self) -> Option<LargeBinary<'a>>

pub fn type_as_large_utf_8(&self) -> Option<LargeUtf8<'a>>

pub fn type_as_large_list(&self) -> Option<LargeList<'a>>

pub fn sparseIndex_as_sparse_tensor_index_coo(
&self
) -> Option<SparseTensorIndexCOO<'a>>

pub fn sparseIndex_as_sparse_matrix_index_csx(
&self
) -> Option<SparseMatrixIndexCSX<'a>>

pub fn sparseIndex_as_sparse_tensor_index_csf(
&self
) -> Option<SparseTensorIndexCSF<'a>>