Module arrow2::io::ipc::read

This is supported on crate feature io_ipc only.

APIs to read Arrow’s IPC format.

The two important structs here are the FileReader, which provides arbitrary access to any of its messages, and the StreamReader, which only supports reading data in the order it was written in.

Structs

FileMetadata

FileReader

Arrow File reader

StreamMetadata

StreamReader

Arrow Stream reader.

Enums

StreamState

Encodes the stream’s status after each read.

Functions

read_dictionary

Read the dictionary from the buffer and provided metadata, updating the dictionaries with the resulting dictionary

read_file_metadata

Read the IPC file’s metadata

read_record_batch

Creates a record batch from binary data using the ipc::RecordBatch indexes and the Schema

read_stream_metadata

Reads the metadata of the stream