Module arrow2::io::ipc::read

Structs

FileMetadata

FileReader
Arrow File reader

StreamMetadata

StreamReader
Arrow Stream reader

Enums

StreamState

Functions

read_dictionary
Read the dictionary from the buffer and provided metadata, updating the dictionaries_by_field with the resulting dictionary

read_file_metadata
Read the IPC file’s metadata

read_record_batch
Creates a record batch from binary data using the ipc::RecordBatch indexes and the Schema

read_stream_metadata
Reads the metadata of the stream