Module arrow2::io::ipc::read

Available on crate feature io_ipc only.

APIs to read Arrow’s IPC format.

The two important structs here are the FileReader, which provides arbitrary access to any of its messages, and the StreamReader, which only supports reading data in the order it was written in.

Modules

file_asyncio_ipc_read_async

Async reader for Arrow IPC files

stream_asyncio_ipc_read_async

APIs to read Arrow streams asynchronously

Structs

FileMetadata

Metadata of an Arrow IPC file, written in the footer of the file.

FileReader

An iterator of Chunks from an Arrow IPC file.

StreamMetadata

Metadata of an Arrow IPC stream, written at the start of the stream

StreamReader

Arrow Stream reader.

Enums

OutOfSpecKind

The different types of errors that reading from IPC can cause

StreamState

Encodes the stream’s status after each read.

Functions

deserialize_schema

Deserialize an flatbuffers-encoded Schema message into Schema and IpcSchema.

read_batch

Reads the record batch at position index from the reader.

read_file_dictionaries

Reads all file’s dictionaries, if any This function is IO-bounded

read_file_metadata

Read the Arrow IPC file’s metadata

read_stream_metadata

Reads the metadata of the stream

Type Definitions

Dictionaries

how dictionaries are tracked in this crate