Module arrow2::io::ipc::read

Available on crate feature io_ipc only.

APIs to read Arrow’s IPC format.

The two important structs here are the FileReader, which provides arbitrary access to any of its messages, and the StreamReader, which only supports reading data in the order it was written in.

Modules

file_asyncio_ipc_read_async

Async reader for Arrow IPC files

stream_asyncio_ipc_read_async

APIs to read Arrow streams asynchronously

Structs

FileMetadata

Metadata of an Arrow IPC file, written in the footer of the file.

FileReader

Arrow File reader

StreamMetadata

Metadata of an Arrow IPC stream, written at the start of the stream

StreamReader

Arrow Stream reader.

Enums

StreamState

Encodes the stream’s status after each read.

Functions

deserialize_schema

Deserialize an flatbuffers-encoded Schema message into Schema and IpcSchema.

read_dictionary

Read the dictionary from the buffer and provided metadata, updating the dictionaries with the resulting dictionary

read_file_metadata

Read the IPC file’s metadata

read_record_batch

Creates a record batch from binary data using the ipc::RecordBatch indexes and the Schema

read_stream_metadata

Reads the metadata of the stream

Type Definitions

Dictionaries

how dictionaries are tracked in this crate