Trait arrow2::io::json::JsonFormat

pub trait JsonFormat: Debug + Default {
    fn start_stream<W: Write>(&self, _writer: &mut W) -> Result<()> { ... }

    fn start_row<W: Write>(
        &self, 
        _writer: &mut W, 
        _is_first_row: bool
    ) -> Result<()> { ... }

    fn end_row<W: Write>(&self, _writer: &mut W) -> Result<()> { ... }

    fn end_stream<W: Write>(&self, _writer: &mut W) -> Result<()> { ... }
}

This trait defines how to format a sequence of JSON objects to a byte stream.

Provided methods

fn start_stream<W: Write>(&self, _writer: &mut W) -> Result<()>

write any bytes needed at the start of the file to the writer

fn start_row<W: Write>(
&self,
_writer: &mut W,
_is_first_row: bool
) -> Result<()>

write any bytes needed for the start of each row

fn end_row<W: Write>(&self, _writer: &mut W) -> Result<()>

write any bytes needed for the end of each row

fn end_stream<W: Write>(&self, _writer: &mut W) -> Result<()>

write any bytes needed for the start of each row