Module arrow2::io::parquet::read

Re-exports

pub use schema::get_schema;

pub use schema::is_type_nullable;

pub use parquet2::read::streaming_iterator;

Modules

schema

statistics

Structs

ColumnChunkMetaData
Metadata for a column chunk.

ColumnDescriptor
A descriptor for leaf-level primitive columns. This encapsulates information such as definition and repetition levels and is used to re-assemble nested data.

CompressedDataPage
A CompressedDataPage is compressed, encoded representation of a Parquet data page. It holds actual data and thus cloning it is expensive.

DataPage
A DataPage is an uncompressed, encoded representation of a Parquet data page. It holds actual data and thus cloning it is expensive.

Decompressor
Decompressor that allows re-using the page buffer of PageIterator

FileMetaData
Metadata for a Parquet file.

PageIterator
A page iterator iterates over row group’s pages. In parquet, pages are guaranteed to be contiguously arranged in memory and therefore must be read in sequence.

RecordReader
Single threaded iterator of RecordBatch from a parquet file.

RowGroupMetaData
Metadata for a row group.

TimestampType
Timestamp logical type annotation

Enums

DataPageHeader

LogicalType

ParquetError

ParquetTimeUnit

ParquetType
Representation of a Parquet type. Used to describe primitive leaf fields and structs, including top-level schema. Note that the top-level schema type is represented using GroupType whose repetition is None.

PhysicalType

PrimitiveConvertedType

Traits

StreamingIterator
An interface for dealing with streaming iterators.

Functions

_get_page_iterator

_read_metadata
Reads a file’s metadata.

decompress
Decompresses the page, using buffer for decompression.

get_page_iterator
Creates a new iterator of compressed pages.

int96_to_i64_ns

page_iter_to_array

read_metadata
Reads parquets’ metadata.