Function arrow2::io::parquet::read::column_iter_to_arrays

pub fn column_iter_to_arrays<'a, I: 'a>(
    columns: Vec<I>, 
    types: Vec<&PrimitiveType>, 
    field: Field, 
    chunk_size: usize
) -> Result<ArrayIter<'a>> where
    I: DataPages, 

Available on crate feature io_parquet only.

An iterator adapter that maps multiple iterators of DataPages into an iterator of Arrays.

The arrays are guaranteed to be at most of size chunk_size and data type field.data_type.