Module arrow2::io::parquet::read::schema

This is supported on crate feature io_parquet only.

APIs to handle Parquet <-> Arrow schemas.

Structs

FileMetaData

Metadata for a Parquet file.

SchemaDescriptor

A schema descriptor. This encapsulates the top-level schemas for all the columns, as well as all descriptors for all the primitive columns.

Enums

ParquetType

Representation of a Parquet type. Used to describe primitive leaf fields and structs, including top-level schema. Note that the top-level schema type is represented using GroupType whose repetition is None.

Functions

get_schema

Parses parquet’s metadata into a Schema. This first looks for the metadata key "ARROW:schema"; if it does not exist, it converts logical and converted parquet types to Arrow equivalents.

parquet_to_arrow_schema

Convert parquet schema to arrow schema

read_schema_from_metadata

Reads an arrow schema from Parquet’s file metadata. Returns None if no schema was found.

Type Definitions

KeyValue