Trait arrow2::io::parquet::read::StreamingIterator

pub trait StreamingIterator {
    type Item: ?Sized;

    pub fn advance(&mut self);

    pub fn get(&self) -> Option<&Self::Item>;

    pub fn next(&mut self) -> Option<&Self::Item> { ... }

    pub fn size_hint(&self) -> (usize, Option<usize>) { ... }

    pub fn all<F>(&mut self, f: F) -> bool
    where
        F: FnMut(&Self::Item) -> bool,
    { ... }

    pub fn any<F>(&mut self, f: F) -> bool
    where
        F: FnMut(&Self::Item) -> bool,
    { ... }

    pub fn by_ref(&mut self) -> &mut Self { ... }

    pub fn chain<I>(self, other: I) -> Chain<Self, I>
    where
        I: StreamingIterator<Item = Self::Item>,
    { ... }

    pub fn cloned(self) -> Cloned<Self>
    where
        Self::Item: Clone,
    { ... }

    pub fn count(self) -> usize { ... }

    pub fn filter<F>(self, f: F) -> Filter<Self, F>
    where
        F: FnMut(&Self::Item) -> bool,
    { ... }

    pub fn filter_map<B, F>(self, f: F) -> FilterMap<Self, B, F>
    where
        F: FnMut(&Self::Item) -> Option<B>,
    { ... }

    pub fn flat_map<J, F>(self, f: F) -> FlatMap<Self, J, F>
    where
        F: FnMut(&Self::Item) -> J,
        J: StreamingIterator,
    { ... }

    pub fn filter_map_deref<B, F>(self, f: F) -> FilterMapDeref<Self, F>
    where
        F: FnMut(&Self::Item) -> Option<B>,
    { ... }

    pub fn find<F>(&mut self, f: F) -> Option<&Self::Item>
    where
        F: FnMut(&Self::Item) -> bool,
    { ... }

    pub fn fuse(self) -> Fuse<Self> { ... }

    pub fn inspect<F>(self, f: F) -> Inspect<Self, F>
    where
        F: FnMut(&Self::Item),
    { ... }

    pub fn map<B, F>(self, f: F) -> Map<Self, B, F>
    where
        F: FnMut(&Self::Item) -> B,
    { ... }

    pub fn map_deref<B, F>(self, f: F) -> MapDeref<Self, F>
    where
        F: FnMut(&Self::Item) -> B,
    { ... }

    pub fn map_ref<B, F>(self, f: F) -> MapRef<Self, F>
    where
        B: ?Sized,
        F: Fn(&Self::Item) -> &B,
    { ... }

    pub fn nth(&mut self, n: usize) -> Option<&Self::Item> { ... }

    pub fn position<F>(&mut self, f: F) -> Option<usize>
    where
        F: FnMut(&Self::Item) -> bool,
    { ... }

    pub fn skip(self, n: usize) -> Skip<Self> { ... }

    pub fn skip_while<F>(self, f: F) -> SkipWhile<Self, F>
    where
        F: FnMut(&Self::Item) -> bool,
    { ... }

    pub fn take(self, n: usize) -> Take<Self> { ... }

    pub fn take_while<F>(self, f: F) -> TakeWhile<Self, F>
    where
        F: FnMut(&Self::Item) -> bool,
    { ... }

    pub fn rev(self) -> Rev<Self>
    where
        Self: DoubleEndedStreamingIterator,
    { ... }

    pub fn fold<B, F>(self, init: B, f: F) -> B
    where
        F: FnMut(B, &Self::Item) -> B,
    { ... }

    pub fn for_each<F>(self, f: F)
    where
        F: FnMut(&Self::Item),
    { ... }

}

An interface for dealing with streaming iterators.

Associated Types

type Item: ?Sized

The type of the elements being iterated over.

Loading content...
Required methods

pub fn advance(&mut self)

Advances the iterator to the next element.

Iterators start just before the first element, so this should be called before get.

The behavior of calling this method after the end of the iterator has been reached is unspecified.

pub fn get(&self) -> Option<&Self::Item>

Returns a reference to the current element of the iterator.

The behavior of calling this method before advance has been called is unspecified.

Loading content...
Provided methods

pub fn next(&mut self) -> Option<&Self::Item>

Advances the iterator and returns the next value.

The behavior of calling this method after the end of the iterator has been reached is unspecified.

The default implementation simply calls advance followed by get.

pub fn size_hint(&self) -> (usize, Option<usize>)

Returns the bounds on the remaining length of the iterator.

pub fn all<F>(&mut self, f: F) -> bool where
F: FnMut(&Self::Item) -> bool,

Determines if all elements of the iterator satisfy a predicate.

pub fn any<F>(&mut self, f: F) -> bool where
F: FnMut(&Self::Item) -> bool,

Determines if any elements of the iterator satisfy a predicate.

pub fn by_ref(&mut self) -> &mut Self

Borrows an iterator, rather than consuming it.

This is useful to allow the application of iterator adaptors while still retaining ownership of the original adaptor.

pub fn chain<I>(self, other: I) -> Chain<Self, I> where
I: StreamingIterator<Item = Self::Item>,

Consumes two iterators and returns a new iterator that iterates over both in sequence.

pub fn cloned(self) -> Cloned<Self> where
Self::Item: Clone,

Produces a normal, non-streaming, iterator by cloning the elements of this iterator.

pub fn count(self) -> usize

Consumes the iterator, counting the number of remaining elements and returning it.

pub fn filter<F>(self, f: F) -> Filter<Self, F> where
F: FnMut(&Self::Item) -> bool,

Creates an iterator which uses a closure to determine if an element should be yielded.

pub fn filter_map<B, F>(self, f: F) -> FilterMap<Self, B, F> where
F: FnMut(&Self::Item) -> Option<B>,

Creates an iterator which both filters and maps by applying a closure to elements.

pub fn flat_map<J, F>(self, f: F) -> FlatMap<Self, J, F> where
F: FnMut(&Self::Item) -> J,
J: StreamingIterator,

Creates an iterator which flattens iterators obtained by applying a closure to elements. Note that the returned iterators must be streaming iterators.

pub fn filter_map_deref<B, F>(self, f: F) -> FilterMapDeref<Self, F> where
F: FnMut(&Self::Item) -> Option<B>,

Creates a regular, non-streaming iterator which both filters and maps by applying a closure to elements.

pub fn find<F>(&mut self, f: F) -> Option<&Self::Item> where
F: FnMut(&Self::Item) -> bool,

Returns the first element of the iterator that satisfies the predicate.

pub fn fuse(self) -> Fuse<Self>

Creates an iterator which is “well behaved” at the beginning and end of iteration.

The behavior of calling get before iteration has been started, and of continuing to call advance after get has returned None is normally unspecified, but this guarantees that get will return None in both cases.

pub fn inspect<F>(self, f: F) -> Inspect<Self, F> where
F: FnMut(&Self::Item),

Call a closure on each element, passing the element on. The closure is called upon calls to advance or advance_back, and exactly once per element regardless of how many times (if any) get is called.

pub fn map<B, F>(self, f: F) -> Map<Self, B, F> where
F: FnMut(&Self::Item) -> B,

Creates an iterator which transforms elements of this iterator by passing them to a closure.

pub fn map_deref<B, F>(self, f: F) -> MapDeref<Self, F> where
F: FnMut(&Self::Item) -> B,

Creates a regular, non-streaming iterator which transforms elements of this iterator by passing them to a closure.

pub fn map_ref<B, F>(self, f: F) -> MapRef<Self, F> where
B: ?Sized,
F: Fn(&Self::Item) -> &B,

Creates an iterator which transforms elements of this iterator by passing them to a closure.

Unlike map, this method takes a closure that returns a reference into the original value.

pub fn nth(&mut self, n: usize) -> Option<&Self::Item>

Consumes the first n elements of the iterator, returning the next one.

pub fn position<F>(&mut self, f: F) -> Option<usize> where
F: FnMut(&Self::Item) -> bool,

Returns the index of the first element of the iterator matching a predicate.

pub fn skip(self, n: usize) -> Skip<Self>

Creates an iterator which skips the first n elements.

pub fn skip_while<F>(self, f: F) -> SkipWhile<Self, F> where
F: FnMut(&Self::Item) -> bool,

Creates an iterator that skips initial elements matching a predicate.

pub fn take(self, n: usize) -> Take<Self>

Creates an iterator which only returns the first n elements.

pub fn take_while<F>(self, f: F) -> TakeWhile<Self, F> where
F: FnMut(&Self::Item) -> bool,

Creates an iterator which only returns initial elements matching a predicate.

pub fn rev(self) -> Rev<Self> where
Self: DoubleEndedStreamingIterator,

Creates an iterator which returns elemens in the opposite order.

pub fn fold<B, F>(self, init: B, f: F) -> B where
F: FnMut(B, &Self::Item) -> B,

Reduces the iterator’s elements to a single, final value.

pub fn for_each<F>(self, f: F) where
F: FnMut(&Self::Item),

Calls a closure on each element of an iterator.

Loading content...