Module arrow2::io::parquet::write

This is supported on crate feature io_parquet only.

APIs to write to Parquet format.

Re-exports

pub use parquet2::fallible_streaming_iterator;

Modules

stream

Contains async APIs to write to parquet.

Structs

ColumnDescriptor

A descriptor for leaf-level primitive columns. This encapsulates information such as definition and repetition levels and is used to re-assemble nested data.

CompressedDataPage

A CompressedDataPage is compressed, encoded representation of a Parquet data page. It holds actual data and thus cloning it is expensive.

Compressor

A FallibleStreamingIterator that consumes EncodedPage and yields CompressedPage holding a reusable buffer (Vec<u8>) for compression.

DynIter

DynIter is an implementation of a single-threaded, dynamically-typed iterator.

DynStreamingIterator

Dynamically-typed FallibleStreamingIterator.

RowGroupIterator

An iterator adapter that converts an iterator over RecordBatch into an iterator of row groups. Use it to create an iterator consumable by the parquet’s API.

SchemaDescriptor

A schema descriptor. This encapsulates the top-level schemas for all the columns, as well as all descriptors for all the primitive columns.

WriteOptions

Enums

CompressedPage

A CompressedPage is a compressed, encoded representation of a Parquet page. It holds actual data and thus cloning it is expensive.

Compression

EncodedPage

A EncodedPage is an uncompressed, encoded representation of a Parquet page. It may hold actual data and thus cloning it may be expensive.

Encoding

ParquetType

Representation of a Parquet type. Used to describe primitive leaf fields and structs, including top-level schema. Note that the top-level schema type is represented using GroupType whose repetition is None.

Version

Traits

FallibleStreamingIterator

A fallible, streaming iterator.

Functions

array_to_page

Converts an Array to a CompressedPage based on options, descriptor and encoding.

array_to_pages

Returns an iterator of EncodedPage.

can_encode

Checks whether the data_type can be encoded as encoding. Note that this is whether this implementation supports it, which is a subset of what the parquet spec allows.

compress

parquet_write_file

to_parquet_schema

Creates a parquet SchemaDescriptor from a Schema.

to_parquet_type

Creates a ParquetType from a Field.

write_file

Writes

Type Definitions

KeyValue

RowGroupIter