Module arrow2::io::parquet::write

Available on crate feature io_parquet only.

APIs to write to Parquet format.

Re-exports

pub use parquet2::fallible_streaming_iterator;

Structs

CompressedDataPage

A CompressedDataPage is compressed, encoded representation of a Parquet data page. It holds actual data and thus cloning it is expensive.

Compressor

A FallibleStreamingIterator that consumes EncodedPage and yields CompressedPage holding a reusable buffer (Vec<u8>) for compression.

Descriptor

A descriptor of a parquet column. It contains the necessary information to deserialize a parquet column.

DynIter

DynIter is an implementation of a single-threaded, dynamically-typed iterator.

DynStreamingIterator

Dynamically-typed FallibleStreamingIterator.

FileSink

Sink that writes array chunks as a Parquet file.

FileWriter

An interface to write a parquet to a Write

KeyValue

Wrapper struct to store key values

RowGroupIterator

An iterator adapter that converts an iterator over Chunk into an iterator of row groups. Use it to create an iterator consumable by the parquet’s API.

SchemaDescriptor

A schema descriptor. This encapsulates the top-level schemas for all the columns, as well as all descriptors for all the primitive columns.

WriteOptions

Currently supported options to write to parquet

Enums

CompressedPage

A CompressedPage is a compressed, encoded representation of a Parquet page. It holds actual data and thus cloning it is expensive.

CompressionOptions

Defines the compression settings for writing a parquet file.

EncodedPage

A EncodedPage is an uncompressed, encoded representation of a Parquet page. It may hold actual data and thus cloning it may be expensive.

Encoding

ParquetType

Representation of a Parquet type describing primitive and nested fields, including the top-level schema of the parquet file.

Version

The parquet version to use

Traits

FallibleStreamingIterator

A fallible, streaming iterator.

Functions

array_to_page

Converts an Array to a CompressedPage based on options, descriptor and encoding.

array_to_pages

Returns an iterator of EncodedPage.

can_encode

Checks whether the data_type can be encoded as encoding. Note that this is whether this implementation supports it, which is a subset of what the parquet spec allows.

compress

Compresses an EncodedPage into a CompressedPage using compressed_buffer as the intermediary buffer.

row_group_iter

Maps a Chunk and parquet-specific options to an RowGroupIter used to write to parquet

to_parquet_schema

Creates a parquet SchemaDescriptor from a Schema.

to_parquet_type

Creates a ParquetType from a Field.

Type Definitions

RowGroupIter