Module arrow2::types

Traits and implementations to handle all types used in this crate.

Most physical types used in this crate are native Rust types, like i32. The most important trait is NativeType, the generic trait of crate::array::PrimitiveArray.

Another important trait is BitChunk, describing types that can be used to represent chunks of bits (e.g. u8, u16), and BitChunkIter, that can be used to iterate over bitmaps in BitChunks. Finally, this module also contains traits used to compile code optimized for SIMD instructions at simd.

Modules

simd

Contains traits and implementations of multi-data used in SIMD. The actual representation is driven by the feature flag "simd", which, if set, uses packed_simd2 to get the intrinsics.

Structs

BitChunkIter

An iterator of bool over a BitChunk. This iterator is often compiled to SIMD instructions. The LSB corresponds to the first slot, as defined by the arrow specification.

IndexRange

iterator of Index equivalent to (a..b).

days_ms

The in-memory representation of the DayMillisecond variant of arrow’s “Interval” logical type.

months_days_ns

The in-memory representation of the MonthDayNano variant of the “Interval” logical type.

Traits

BitChunk

Something that can be use as a chunk of bits. This is used to create masks ofa given number of length, whose width is 1. In simd_packed notation, this corresponds to m1xY.

Index

Types that can be used to index a slot of an array.

NativeType

Sealed trait that implemented by all types that can be allocated, serialized and deserialized by this crate. All O(N) in-memory allocations are implemented for this trait alone.

NaturalDataType

Trait denoting anything that has a natural logical DataType. For example, DataType::Int32 for i32.

Relation

describes whether a DataType is valid.