Trait arrow2::types::BitChunk

pub unsafe trait BitChunk: Sized + Copy + Debug + Binary + BitAnd<Output = Self> + ShlAssign + Not<Output = Self> + ShrAssign<u32> + ShlAssign<u32> + Shl<u32, Output = Self> + Eq + BitAndAssign + BitOr<Output = Self> {
    type Bytes: Index<usize, Output = u8> + IndexMut<usize, Output = u8> + for<'a> TryFrom<&'a [u8]> + Debug;
    fn one() -> Self;

    fn zero() -> Self;

    fn to_ne_bytes(self) -> Self::Bytes;

    fn from_ne_bytes(v: Self::Bytes) -> Self;
}

Something that can be use as a chunk of bits. This is used to create masks of a given number of lengths, whose width is 1. In simd_packed notation, this corresponds to m1xY.

Safety

Do not implement.

Associated Types

type Bytes: Index<usize, Output = u8> + IndexMut<usize, Output = u8> + for<'a> TryFrom<&'a [u8]> + Debug

Loading content...
Required methods

fn one() -> Self

fn zero() -> Self

fn to_ne_bytes(self) -> Self::Bytes

fn from_ne_bytes(v: Self::Bytes) -> Self

Loading content...