Trait arrow2::types::BitChunk

pub unsafe trait BitChunk: Sized + NativeType + Copy + Debug + Binary + BitAnd<Output = Self> + ShlAssign + Not<Output = Self> + ShrAssign<usize> + ShlAssign<usize> + Shl<usize, Output = Self> + Eq + BitAndAssign + BitOr<Output = Self> {
    fn one() -> Self;

    fn zero() -> Self;

    fn to_ne_bytes(self) -> Self::Bytes;

    fn from_ne_bytes(v: Self::Bytes) -> Self;
}

Something that can be use as a chunk of bits. This is used to create masks ofa given number of length, whose width is 1. In simd_packed notation, this corresponds to m1xY.

Safety

Do not implement.

Required methods

fn one() -> Self

A value with a single bit set at the most right position.

fn zero() -> Self

A value with no bits set.

fn to_ne_bytes(self) -> Self::Bytes

convert itself into bytes.

fn from_ne_bytes(v: Self::Bytes) -> Self

convert itself from bytes.