Trait arrow2::types::NativeType

pub unsafe trait NativeType: Relation + NaturalDataType + Send + Sync + Sized + Copy + Debug + Display + PartialEq + Default + Sized + 'static {
    type Bytes: AsRef<[u8]> + for<'a> TryFrom<&'a [u8]>;
    fn to_le_bytes(&self) -> Self::Bytes;

    fn to_be_bytes(&self) -> Self::Bytes;

    fn from_be_bytes(bytes: Self::Bytes) -> Self;
}

Trait declaring any type that can be allocated, serialized and deserialized by this crate. All data-heavy memory operations are implemented for this trait alone.

Safety

Do not implement.

Associated Types

type Bytes: AsRef<[u8]> + for<'a> TryFrom<&'a [u8]>

Required methods

fn to_le_bytes(&self) -> Self::Bytes

fn to_be_bytes(&self) -> Self::Bytes

fn from_be_bytes(bytes: Self::Bytes) -> Self