Trait arrow2::types::NativeType

pub trait NativeType: Sealed + Send + Sync + Sized + Copy + Debug + Display + PartialEq + Default + 'static {
    type Bytes: AsRef<[u8]> + Index<usize, Output = u8> + IndexMut<usize, Output = u8> + for<'a> TryFrom<&'a [u8]> + Debug;

    const PRIMITIVE: PrimitiveType;

    fn to_le_bytes(&self) -> Self::Bytes;

    fn to_ne_bytes(&self) -> Self::Bytes;

    fn to_be_bytes(&self) -> Self::Bytes;

    fn from_be_bytes(bytes: Self::Bytes) -> Self;
}

Sealed trait implemented by all physical types that can be allocated, serialized and deserialized by this crate. All O(N) allocations in this crate are done for this trait alone.

Associated Types

type Bytes: AsRef<[u8]> + Index<usize, Output = u8> + IndexMut<usize, Output = u8> + for<'a> TryFrom<&'a [u8]> + Debug

Type denoting its representation as bytes. This is [u8; N] where N = size_of::<T>.

Associated Constants

const PRIMITIVE: PrimitiveType

The corresponding variant of PrimitiveType.

Required methods

fn to_le_bytes(&self) -> Self::Bytes

To bytes in little endian

fn to_ne_bytes(&self) -> Self::Bytes

To bytes in native endian

fn to_be_bytes(&self) -> Self::Bytes

To bytes in big endian

fn from_be_bytes(bytes: Self::Bytes) -> Self

From bytes in big endian